### 2024-08-08 #7088 @editorialbot
[PRE REVIEW]: VoteKit: A Python package for computational social choice research
**Submitting author:** @cdonnay (<a href="http://orcid.org/0000-0003-4498-4067">Moon Duchin</a>)
**Repository:** https://github.com/mggg/votekit
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v2.0.1
