### 2024-08-08 #7088 @samhforbes
@cdonnay I will begin looking for an editor for this submission, but in the meantime if you don't mind, you can see an error generating the pdf, as well as (less urgently) some missing DOIs in the references.
