### 2024-08-08 #7088 @cdonnay
Will do!

There is a similar paper in submission that we think compliments ours very well: https://github.com/openjournals/joss-reviews/issues/7020

Perhaps some of the editors from that would be appropriate for us as well.
