+++ tests/elections/election_types/ranking/test_irv.py
@@ -1,0 +1,290 @@
+from votekit.elections import IRV, ElectionState
+from votekit import Ballot, PreferenceProfile
+import pandas as pd
+import pytest
+
+# taken from STV wiki
+test_profile = PreferenceProfile(
+    ballots=[
+        Ballot(ranking=({"Orange"}, {"Pear"}), weight=3),
+        Ballot(ranking=({"Pear"}, {"Strawberry"}, {"Cake"}), weight=8),
+        Ballot(ranking=({"Strawberry"}, {"Orange"}, {"Pear"}), weight=1 / 2),
+        Ballot(ranking=({"Cake"}, {"Chocolate"}), weight=3),
+        Ballot(ranking=({"Chocolate"}, {"Cake"}, {"Burger"}), weight=1),
+        Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+        Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+    ]
+)
+
+profile_list = [
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"}, {"Pear"}), weight=3),
+            Ballot(ranking=({"Pear"}, {"Strawberry"}, {"Cake"}), weight=8),
+            Ballot(ranking=({"Strawberry"}, {"Orange"}, {"Pear"}), weight=1 / 2),
+            Ballot(ranking=({"Cake"}, {"Chocolate"}), weight=3),
+            Ballot(ranking=({"Chocolate"}, {"Cake"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+            Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+        ]
+    ),
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"}, {"Pear"}), weight=3.5),
+            Ballot(ranking=({"Pear"}, {"Cake"}), weight=8),
+            Ballot(ranking=({"Cake"}, {"Chocolate"}), weight=3),
+            Ballot(ranking=({"Chocolate"}, {"Cake"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+            Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+        ]
+    ),
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"}, {"Pear"}), weight=3.5),
+            Ballot(ranking=({"Pear"}, {"Cake"}), weight=8),
+            Ballot(ranking=({"Cake"},), weight=3),
+            Ballot(ranking=({"Cake"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+            Ballot(ranking=({"Chicken"}, {"Burger"}), weight=3),
+        ]
+    ),
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"}, {"Pear"}), weight=3.5),
+            Ballot(ranking=({"Pear"}, {"Cake"}), weight=8),
+            Ballot(ranking=({"Cake"},), weight=3),
+            Ballot(ranking=({"Cake"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"},), weight=7),
+        ]
+    ),
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Pear"},), weight=3.5),
+            Ballot(ranking=({"Pear"}, {"Cake"}), weight=8),
[…]
