+++ tests/elections/election_types/ranking/test_sequential_rcv.py
@@ -1,0 +1,264 @@
+from votekit.elections import SequentialRCV, ElectionState
+from votekit import Ballot, PreferenceProfile
+import pytest
+
+# taken from STV wiki
+simult_same_as_one_by_one_profile = PreferenceProfile(
+    ballots=[
+        Ballot(ranking=({"Orange"}, {"Pear"}), weight=3),
+        Ballot(ranking=({"Pear"}, {"Strawberry"}, {"Cake"}), weight=8),
+        Ballot(ranking=({"Strawberry"}, {"Orange"}, {"Pear"}), weight=1),
+        Ballot(ranking=({"Cake"}, {"Chocolate"}), weight=3),
+        Ballot(ranking=({"Chocolate"}, {"Cake"}, {"Burger"}), weight=1),
+        Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+        Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+    ]
+)
+
+profile_list = [
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"}, {"Pear"}), weight=3),
+            Ballot(ranking=({"Pear"}, {"Strawberry"}, {"Cake"}), weight=8),
+            Ballot(ranking=({"Strawberry"}, {"Orange"}, {"Pear"}), weight=1),
+            Ballot(ranking=({"Cake"}, {"Chocolate"}), weight=3),
+            Ballot(ranking=({"Chocolate"}, {"Cake"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+            Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+        ]
+    ),
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"},), weight=3),
+            Ballot(ranking=({"Strawberry"}, {"Cake"}), weight=8),
+            Ballot(ranking=({"Strawberry"}, {"Orange"}), weight=1),
+            Ballot(ranking=({"Cake"}, {"Chocolate"}), weight=3),
+            Ballot(ranking=({"Chocolate"}, {"Cake"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+            Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+        ]
+    ),
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"},), weight=4),
+            Ballot(ranking=({"Cake"},), weight=8),
+            Ballot(ranking=({"Cake"}, {"Chocolate"}), weight=3),
+            Ballot(ranking=({"Chocolate"}, {"Cake"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+            Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+        ]
+    ),
+    PreferenceProfile(
+        ballots=[
+            Ballot(ranking=({"Orange"},), weight=4),
+            Ballot(ranking=({"Chocolate"},), weight=3),
+            Ballot(ranking=({"Chocolate"}, {"Burger"}), weight=1),
+            Ballot(ranking=({"Burger"}, {"Chicken"}), weight=4),
+            Ballot(ranking=({"Chicken"}, {"Chocolate"}, {"Burger"}), weight=3),
+        ]
+    ),
+]
+
+states = [
+    ElectionState(
+        round_number=0,
+        remaining=(
[…]

[…]
