+++ tests/elections/election_types/ranking/test_stv.py
@@ -314,13 +314,22 @@ def test_simul_v_1by1_():
 
 
 def test_errors():
-    with pytest.raises(ValueError):  # m must be non negative
+    with pytest.raises(
+        ValueError,
+        match="m must be non-negative and less than or equal to the number of candidates.",
+    ):
         STV(simult_same_as_one_by_one_profile, m=0)
 
-    with pytest.raises(ValueError):  # m must be less than num cands
+    with pytest.raises(
+        ValueError,
+        match="m must be non-negative and less than or equal to the number of candidates.",
+    ):
         STV(simult_same_as_one_by_one_profile, m=8)
 
-    with pytest.raises(ValueError):  # needs tiebreak
+    with pytest.raises(
+        ValueError,
+        match="Cannot elect correct number of candidates without breaking ties.",
+    ):
         profile = PreferenceProfile(
             ballots=(
                 Ballot(ranking=(frozenset({"A"}),)),
@@ -332,8 +341,8 @@ def test_errors():
         # A and B are tied
         STV(profile, m=2, simultaneous=False)
 
-    with pytest.raises(ValueError):  # quota string
-        STV(PreferenceProfile(), m=1, quota="Drip")
+    with pytest.raises(ValueError, match="Misspelled or unknown quota type."):
+        STV(PreferenceProfile(candidates=["A"]), m=1, quota="Drip")
 
-    with pytest.raises(TypeError):  # need rankings
+    with pytest.raises(TypeError, match="Ballots must have rankings."):
         STV(PreferenceProfile(ballots=(Ballot(scores={"A": 4}),)))
