+++ src/votekit/elections/election_types/ranking/stv.py
@@ -61,7 +61,9 @@ class STV(RankingElection):
         self.m = m
         self.transfer = transfer
         self.quota = quota
-        self.threshold = 0  # for caching purposes
+        # Set to 0 initially so that first call to `get_threshold` returns the
+        # proper threshold.
+        self.threshold = 0
         self.threshold = self.get_threshold(profile.total_ballot_wt)
         self.simultaneous = simultaneous
         self.tiebreak = tiebreak
@@ -100,7 +102,6 @@ class STV(RankingElection):
             return self.threshold
 
     def _is_finished(self):
-        # when m candidates are elected
         elected_cands = [c for s in self.get_elected() for c in s]
 
         if len(elected_cands) == self.m:
@@ -125,7 +126,6 @@ class STV(RankingElection):
         """
         ranking_by_fpv = prev_state.remaining
 
-        # find ranking of cands above threshold
         elected = []
         for s in ranking_by_fpv:
             c = list(s)[0]  # all cands in set have same score
@@ -136,17 +136,12 @@ class STV(RankingElection):
             else:
                 break
 
-        # no tiebreaks needed here, since threshold is such that
-        # no more than m candidates can be elected
-
         ballots_by_fpv = ballots_by_first_cand(profile)
         new_ballots = [Ballot()] * profile.num_ballots
         ballot_index = 0
 
-        # for candidate sets above threshold
         for s in elected:
             for candidate in s:
-                # transfer ballots with candidate in first, using transfer method
                 transfer_ballots = self.transfer(
                     candidate,
                     prev_state.scores[candidate],
@@ -158,7 +153,6 @@ class STV(RankingElection):
                 ] = transfer_ballots
                 ballot_index += len(transfer_ballots)
 
-        # preserve remaining ballots
         for candidate in set([c for s in ranking_by_fpv for c in s]).difference(
             [c for s in elected for c in s]
         ):
@@ -168,7 +162,6 @@ class STV(RankingElection):
             ] = transfer_ballots
             ballot_index += len(transfer_ballots)
 
-        # remove elected from all ballots
         cleaned_ballots = remove_cand(
             [c for s in elected for c in s],
             tuple([b for b in new_ballots if b.ranking]),
@@ -204,7 +197,7 @@ class STV(RankingElection):
                 and whose third entry is the profile of ballots after transfers.
         """
         ranking_by_fpv = prev_state.remaining
-        # find one cand who crossed threshold with highest FPV
+
         elected, remaining, tiebreak = elect_cands_from_set_ranking(
             ranking_by_fpv, m=1, profile=profile, tiebreak=self.tiebreak
         )
@@ -219,7 +212,6 @@ class STV(RankingElection):
 
         elected_c = list(elected[0])[0]
 
-        # transfer ballots with elected candidate in first, using transfer method
         transfer_ballots = self.transfer(
             elected_c,
             prev_state.scores[elected_c],
@@ -231,7 +223,6 @@ class STV(RankingElection):
         ] = transfer_ballots
         ballot_index += len(transfer_ballots)
 
-        # preserve remaining ballots
         for s in remaining:
             for candidate in s:
                 transfer_ballots = tuple(ballots_by_fpv[candidate])
@@ -240,7 +231,6 @@ class STV(RankingElection):
                 ] = transfer_ballots
                 ballot_index += len(transfer_ballots)
 
-        # remove elected from all ballots
         cleaned_ballots = remove_cand(
             elected_c, tuple([b for b in new_ballots if b.ranking])
         )
@@ -258,6 +248,12 @@ class STV(RankingElection):
     ) -> PreferenceProfile:
         """
         Run one step of an election from the given profile and previous state.
[…]
