### 2024-08-08 4e8969e
final election file cleanup
