+++ src/votekit/elections/election_types.py
@@ -1,1119 +1,0 @@
-from fractions import Fraction
-import itertools as it
-import numpy as np
-from typing import Callable, Optional, Union
-from functools import lru_cache
-
-from .election_state import ElectionState
-from ..models import Election
-from ..graphs.pairwise_comparison_graph import PairwiseComparisonGraph
-from ..pref_profile import PreferenceProfile
-from .transfers import fractional_transfer
-from ..utils import (
-    compute_votes,
-    remove_cand,
-    scores_into_set_list,
-    tie_broken_ranking,
-    elect_cands_from_set_ranking,
-    first_place_votes,
-    compute_scores_from_vector,
-    validate_score_vector,
-    borda_scores,
-    ballots_by_first_cand,
-)
-
-
-class STV(Election):
-    """
-    Class for multi-winner STV elections.
-
-    Args:
-        profile (PreferenceProfile):   PreferenceProfile to run election on.
-        transfer (Callable): Transfer method (e.g. fractional transfer).
-        seats (int): Number of seats to be elected.
-        quota (str, optional): Formula to calculate quota. Accepts "droop" or "hare".
-            Defaults to "droop".
-        ballot_ties (bool, optional): Resolves input ballot ties if True, else assumes ballots have
-            no ties. Defaults to True.
-        tiebreak (Union[str, Callable], optional): Resolves procedural and final ties by specified
-            tiebreak. Can either be a custom tiebreak function or a string. Supported strings are
-            given in ``tie_broken_ranking`` documentation. The custom function must take as
-            input two named parameters; ``ranking``, a list-of-sets ranking of candidates and
-            ``profile``, the original ``PreferenceProfile``. It must return a list-of-sets
-            ranking of candidates with no ties. Defaults to random tiebreak.
-
-    Attributes:
-        _profile (PreferenceProfile):   PreferenceProfile to run election on.
-        state (ElectionState): Current state of the election.
-        transfer (Callable): Transfer method (e.g. fractional transfer).
-        seats (int): Number of seats to be elected.
-        quota (str): Formula to calculate quota.
-        tiebreak (Union[str, Callable]): Resolves procedural and final ties by specified
-            tiebreak.
-        threshold (int): Threshold number of votes to be elected.
-
-    """
-
-    def __init__(
-        self,
-        profile: PreferenceProfile,
-        transfer: Callable,
-        seats: int,
-        quota: str = "droop",
-        ballot_ties: bool = True,
-        tiebreak: Union[str, Callable] = "random",
-    ):
-        # let parent class handle the og profile and election state
-        super().__init__(profile, ballot_ties)
-
-        self.transfer = transfer
-        self.seats = seats
-        self.tiebreak = tiebreak
-        self.quota = quota.lower()
-        self.threshold = self.get_threshold()
-
-    # can cache since it will not change throughout rounds
-    def get_threshold(self) -> int:
-        """
-        Calculates threshold required for election.
-
-        Returns:
-            int: Value of the threshold.
-        """
-        quota = self.quota
-        if quota == "droop":
-            return int(self._profile.num_ballots() / (self.seats + 1) + 1)
-        elif quota == "hare":
-            return int(self._profile.num_ballots() / self.seats)
-        else:
-            raise ValueError("Misspelled or unknown quota type")
-
-    def next_round(self) -> bool:
-        """
-        Determines if the number of seats has been met to call an election.
-
-        Returns:
-            bool: True if number of seats has not been met, False otherwise.
-        """
-        cands_elected = 0
-        for s in self.state.winners():
-            cands_elected += len(s)
-        return cands_elected < self.seats
-
-    def run_step(self) -> ElectionState:
-        """
-        Simulates one round an STV election.
-
-        Returns:
-           ElectionState: An ElectionState object for a round.
-        """
[…]
