### 2024-08-08 e4f2171
clean ups from 3.0.0 edits
