+++ tests/test_pwc_graph.py
@@ -9,25 +9,27 @@ from votekit.pref_profile import PreferenceProfile
 
 
 ballot_list = [
-    Ballot(
-        id=None, ranking=[{"A"}, {"C"}, {"D"}, {"B"}, {"E"}], weight=Fraction(10, 1)
-    ),
-    Ballot(
-        id=None, ranking=[{"A"}, {"B"}, {"C"}, {"D"}, {"E"}], weight=Fraction(10, 1)
-    ),
-    Ballot(
-        id=None, ranking=[{"D"}, {"A"}, {"E"}, {"B"}, {"C"}], weight=Fraction(10, 1)
-    ),
-    Ballot(id=None, ranking=[{"A"}], weight=Fraction(24, 1)),
+    Ballot(ranking=[{"A"}, {"C"}, {"D"}, {"B"}, {"E"}], weight=Fraction(10, 1)),
+    Ballot(ranking=[{"A"}, {"B"}, {"C"}, {"D"}, {"E"}], weight=Fraction(10, 1)),
+    Ballot(ranking=[{"D"}, {"A"}, {"E"}, {"B"}, {"C"}], weight=Fraction(10, 1)),
+    Ballot(ranking=[{"A"}], weight=Fraction(24, 1)),
 ]
 TEST_PROFILE = PreferenceProfile(ballots=ballot_list)
 
 simple_ballot_list = [
-    Ballot(id=None, ranking=[{"C"}, {"B"}, {"A"}], weight=Fraction(10, 1)),
-    Ballot(id=None, ranking=[{"A"}, {"C"}, {"B"}], weight=Fraction(10, 1)),
-    Ballot(id=None, ranking=[{"B"}, {"A"}, {"C"}], weight=Fraction(10, 1)),
+    Ballot(ranking=[{"C"}, {"B"}, {"A"}], weight=Fraction(10, 1)),
+    Ballot(ranking=[{"A"}, {"C"}, {"B"}], weight=Fraction(10, 1)),
+    Ballot(ranking=[{"B"}, {"A"}, {"C"}], weight=Fraction(10, 1)),
 ]
 SIMPLE_TEST_PROFILE = PreferenceProfile(ballots=simple_ballot_list)
+
+EDGE_WEIGHT_0_PROFILE = PreferenceProfile(
+    ballots=[
+        Ballot(ranking=({"A"}, {"B"}, {"C"})),
+        Ballot(ranking=({"A"}, {"C"}, {"B"})),
+        Ballot(ranking=({"B"}, {"A"}, {"C"}), weight=2),
+    ]
+)
 
 
 def test_constructor_sequence():
@@ -40,19 +42,13 @@ def test_pwcg_ballot_fill():
     filled_ballots = pwcg.ballot_fill(TEST_PROFILE, 5)
 
     target_ballot_list = [
-        Ballot(
-            id=None, ranking=[{"A"}, {"C"}, {"D"}, {"B"}, {"E"}], weight=Fraction(10, 1)
-        ),
-        Ballot(
-            id=None, ranking=[{"A"}, {"B"}, {"C"}, {"D"}, {"E"}], weight=Fraction(10, 1)
-        ),
-        Ballot(
-            id=None, ranking=[{"D"}, {"A"}, {"E"}, {"B"}, {"C"}], weight=Fraction(10, 1)
-        ),
+        Ballot(ranking=[{"A"}, {"C"}, {"D"}, {"B"}, {"E"}], weight=Fraction(10, 1)),
+        Ballot(ranking=[{"A"}, {"B"}, {"C"}, {"D"}, {"E"}], weight=Fraction(10, 1)),
[…]
