+++ docs/user/tutorial/5_score_based.rst
@@ -1,0 +1,512 @@
+Score-based Voting
+==================
+
+As we saw in the introductory notebook, in addition to ranking-based
+voting, there are also a host of election systems that make use of
+score-based ballots. By the end of this section, you should be
+comfortable with score ballots, Rating elections, and Cumulative
+elections and generators.
+
+Score ballots
+-------------
+
+First, let’s revisit how to define score ballots.
+
+.. code:: ipython3
+
+    from votekit import Ballot
+    
+    score_ballot = Ballot(scores = {"A":4, "B": 3, "C":4}, weight = 3)
+    print(score_ballot)
+    print("ranking:", score_ballot.ranking)
+
+
+.. parsed-literal::
+
+    Scores
+    A: 4.00
+    B: 3.00
+    C: 4.00
+    Weight: 3
+    
+    ranking: None
+
+
+Notice that despite the scores inducing the ranking :math:`{A,C}>B`,
+the ballot only knows the scores. This is to conceptually separate score
+ballots from ranking ballots. If you want to convert a score ballot to a
+ranking, you can use the ``score_dict_to_ranking`` function from the
+``utils`` module.
+
+.. code:: ipython3
+
+    from votekit.utils import score_dict_to_ranking
+    
+    ranking=score_dict_to_ranking(score_ballot.scores)
+    print(ranking)
+    
+    ranked_ballot = Ballot(ranking = ranking, weight = score_ballot.weight)
+    print(ranked_ballot)
+
+
+.. parsed-literal::
+
+    (frozenset({'C', 'A'}), frozenset({'B'}))
+    Ranking
+    1.) C, A, (tie)
+    2.) B, 
+    Weight: 3
+    
+
+
+If you had an entire profile of score ballots and wanted to convert them
+all to ranked, you could do so as follows.
+
+.. code:: ipython3
+
+    from votekit import PreferenceProfile
+    score_profile = PreferenceProfile(ballots = [Ballot(scores = {"A":4, "B": 3, "C":4}, weight = 3),
+                                      Ballot(scores = {"A":2, "B": 3, "C":4}, weight = 2),
+                                      Ballot(scores = {"A":1, "B": 5, "C":4}, weight = 5),
+                                      Ballot(scores = {"A":0, "B": 2, "C":0}, weight = 3)])
+    
+    print("Score profile\n",score_profile)
+    
+    ranked_ballots = [Ballot(ranking = score_dict_to_ranking(b.scores),
+                            weight = b.weight) 
+                    for b in score_profile.ballots]
+    
+    ranked_profile = PreferenceProfile(ballots = ranked_ballots)
+    
+    
+    print("Ranked profile\n", ranked_profile)
+
+
+.. parsed-literal::
+
+    Score profile
+     Ranking                   Scores Weight
+         () (A:1.00, B:5.00, C:4.00)      5
+         () (A:4.00, B:3.00, C:4.00)      3
+         ()                (B:2.00,)      3
+         () (A:2.00, B:3.00, C:4.00)      2
+    Ranked profile
+                   Ranking Scores Weight
+                (B, C, A)     ()      5
+    ({'C', 'A'} (Tie), B)     ()      3
+                     (B,)     ()      3
+                (C, B, A)     ()      2
+
+
+Score ballots are flexible enough to allow any non-zero score, including
+negative scores. Scores of 0 are dropped from the dictionary. Note that
+not all election methods support negative scoring, but these elections
+in ``VoteKit`` validate your ballots and will raise a ``TypeError`` if
+an invalid score is passed.
+
+.. code:: ipython3
+
+    score_ballot = Ballot(scores = {"A":-1, "B": 3.14159, "C":0}, weight = 3)
+    print(score_ballot)
+
+
+.. parsed-literal::
+
+    Scores
+    A: -1.00
[…]
