### 2024-08-08 cc3e9bc
added missing DOIs
