+++ paper/paper.bib
@@ -68,14 +68,19 @@
            }
 }
 
-@misc{boehmer2024guidenumericalexperimentselections,
-      title={Guide to Numerical Experiments on Elections in Computational Social Choice}, 
-      author={Niclas Boehmer and Piotr Faliszewski and Łukasz Janeczko and Andrzej Kaczmarczyk and Grzegorz Lisowski and Grzegorz Pierczyński and Simon Rey and Dariusz Stolicki and Stanisław Szufa and Tomasz Wąs},
-      year={2024},
-      eprint={2402.11765},
-      archivePrefix={arXiv},
-      primaryClass={cs.GT},
-      url={https://arxiv.org/abs/2402.11765}, 
+@inproceedings{boehmer2024guidenumericalexperimentselections,
+  title     = {Guide to Numerical Experiments on Elections in Computational Social Choice},
+  author    = {Boehmer, Niclas and Faliszewski, Piotr and Janeczko, Łukasz and Kaczmarczyk, Andrzej and Lisowski, Grzegorz and Pierczyński, Grzegorz and Rey, Simon and Stolicki, Dariusz and Szufa, Stanisław and Wąs, Tomasz},
+  booktitle = {Proceedings of the Thirty-Third International Joint Conference on
+               Artificial Intelligence, {IJCAI-24}},
+  publisher = {International Joint Conferences on Artificial Intelligence Organization},
+  editor    = {Kate Larson},
+  pages     = {7962--7970},
+  year      = {2024},
+  month     = {8},
+  note      = {Survey Track},
+  doi       = {10.24963/ijcai.2024/881},
+  url       = {https://doi.org/10.24963/ijcai.2024/881},
 }
 
 @unpublished{RCV-Cruncher, 
@@ -92,19 +97,20 @@
 }
 
 @inproceedings{drawing-a-map,
-author = {Szufa, Stanis\l{}aw and Faliszewski, Piotr and Skowron, Piotr and Slinko, Arkadii and Talmon, Nimrod},
-title = {Drawing a Map of Elections in the Space of Statistical Cultures},
-year = {2020},
-isbn = {9781450375184},
-publisher = {International Foundation for Autonomous Agents and Multiagent Systems},
-address = {Richland, SC},
-abstract = {We consider the problem of forming a testbed of elections to be used for numerical experiments (such as testing algorithms or estimating the frequency of a given phenomenon). We seek elections that come from well-known statistical distributions and are as diverse as possible. To this end, we define a (pseudo)metric over elections, generate a set of election instances, and measure distances between them, to assess how diverse they are. Finally, we show how to use these elections to test election-related algorithms.},
-booktitle = {Proceedings of the 19th International Conference on Autonomous Agents and MultiAgent Systems},
-pages = {1341–1349},
-numpages = {9},
-keywords = {generative models, experimental social choice, elections, distances},
-location = {Auckland, New Zealand},
-series = {AAMAS '20}
+ author = {Szufa, Stanis\l{}aw and Faliszewski, Piotr and Skowron, Piotr and Slinko, Arkadii and Talmon, Nimrod},
+ title = {Drawing a Map of Elections in the Space of Statistical Cultures},
+ year = {2020},
+ isbn = {9781450375184},
+ doi={10.5555/3398761.3398916},
+ publisher = {International Foundation for Autonomous Agents and Multiagent Systems},
+ address = {Richland, SC},
[…]
