### 2024-08-08 eeeab2e
add pydantic
