+++ docs/requirements.txt
@@ -49,7 +49,7 @@ prompt_toolkit==3.0.45
 psutil==5.9.8
 ptyprocess==0.7.0
 pure-eval==0.2.2
-pydantic==1.10.15
+pydantic==2.8.2
 Pygments==2.18.0
 pyparsing==3.1.2
 python-dateutil==2.9.0.post0
