+++ pyproject.toml
@@ -9,7 +9,7 @@ readme = "README.md"
 python = "^3.9,<3.13"
 scipy = "^1.11.1"
 mypy = "^1.4.1"
-pydantic = "^1.10.9"
+pydantic = "^2.8.2"
 pandas-stubs = "^2.0.2.230605"
 pot = "0.9.3"
 networkx = "^3.1"
