+++ pyproject.toml
@@ -11,7 +11,7 @@ scipy = "^1.11.1"
 mypy = "^1.4.1"
 pydantic = "^2.8.2"
 pandas-stubs = "^2.0.2.230605"
-pot = "0.9.3"
+pot = "0.9.4"
 networkx = "^3.1"
 jinja2 = "^3.1.2"
 matplotlib = "^3.7.2"
