### 2024-08-09 c4ef3c8
try/except blocks
