+++ docs/user/tutorial/5_score_based.rst
@@ -28,12 +28,11 @@ First, let’s revisit how to define score ballots.
     B: 3.00
     C: 4.00
     Weight: 3
-    
     ranking: None
 
 
-Notice that despite the scores inducing the ranking :math:`{A,C}>B`,
-the ballot only knows the scores. This is to conceptually separate score
+Notice that despite the scores inducing the ranking :math:`A,C>B`, the
+ballot only knows the scores. This is to conceptually separate score
 ballots from ranking ballots. If you want to convert a score ballot to a
 ranking, you can use the ``score_dict_to_ranking`` function from the
 ``utils`` module.
@@ -51,12 +50,11 @@ ranking, you can use the ``score_dict_to_ranking`` function from the
 
 .. parsed-literal::
 
-    (frozenset({'C', 'A'}), frozenset({'B'}))
+    (frozenset({'A', 'C'}), frozenset({'B'}))
     Ranking
-    1.) C, A, (tie)
+    1.) A, C, (tie)
     2.) B, 
     Weight: 3
-    
 
 
 If you had an entire profile of score ballots and wanted to convert them
@@ -93,7 +91,7 @@ all to ranked, you could do so as follows.
     Ranked profile
                    Ranking Scores Weight
                 (B, C, A)     ()      5
-    ({'C', 'A'} (Tie), B)     ()      3
+    ({'A', 'C'} (Tie), B)     ()      3
                      (B,)     ()      3
                 (C, B, A)     ()      2
 
@@ -116,7 +114,6 @@ an invalid score is passed.
     A: -1.00
     B: 3.14
     Weight: 3
-    
 
 
 Rating Election
@@ -158,7 +155,7 @@ Let’s look at the score totals to convince ourselves B was the winner.
 
 .. parsed-literal::
 
-    {'B': Fraction(46, 1), 'C': Fraction(40, 1), 'A': Fraction(21, 1)}
+    {'A': Fraction(21, 1), 'B': Fraction(46, 1), 'C': Fraction(40, 1)}
 
 
 Now let’s see that the Rating election validates our profile before
@@ -170,52 +167,16 @@ running the election. All of these code blocks should raise
     ranking_profile = PreferenceProfile(ballots = [Ballot(ranking= [{"A"}, {"B"}, {"C"}])])
     
     # should raise a TypeError since this profile has no scores
-    election = Rating(ranking_profile, m = 1, L = 5)
+    try:
+        election = Rating(ranking_profile, m = 1, L = 5)
+    except Exception as e:
+        print(f"Found the following error:\n\t{e.__class__.__name__}: {e}")
 
 
-::
+.. parsed-literal::
 
-
-    ---------------------------------------------------------------------------
-
-    TypeError                                 Traceback (most recent call last)
-
-    Cell In[11], line 4
-          1 ranking_profile = PreferenceProfile(ballots = [Ballot(ranking= [{"A"}, {"B"}, {"C"}])])
-          3 # should raise a TypeError since this profile has no scores
-    ----> 4 election = Rating(ranking_profile, m = 1, L = 5)
-
-
-    File ~/PycharmProjects/VoteKit/src/votekit/elections/election_types/scores/rating.py:153, in Rating.__init__(self, profile, m, L, tiebreak)
-        146 def __init__(
-        147     self,
-        148     profile: PreferenceProfile,
-       (...)
-        151     tiebreak: Optional[str] = None,
-        152 ):
-    --> 153     super().__init__(profile, m=m, L=L, tiebreak=tiebreak)
-
-
-    File ~/PycharmProjects/VoteKit/src/votekit/elections/election_types/scores/rating.py:46, in GeneralRating.__init__(self, profile, m, L, k, tiebreak)
[…]
