+++ src/votekit/elections/election_types/ranking/plurality_veto.py
@@ -1,0 +1,179 @@
+from .abstract_ranking import RankingElection
+from ....pref_profile import PreferenceProfile
+from ....ballot import Ballot
+from ...election_state import ElectionState
+from ....utils import (
+    first_place_votes,
+    remove_cand,
+    score_dict_to_ranking,
+    tiebreak_set,
+)
+from fractions import Fraction
+import numpy as np
+from typing import Optional
+
+
+class PluralityVeto(RankingElection):
+    """
+    Scores each candidate by their plurality (number of first) place votes,
+    then in a randomized order it lets each voter decrement the score of their
+    least favorite candidate. The candidate with the largest score at the end is
+    then chosen as the winner.
+
+    Args:
+      profile (PreferenceProfile): PreferenceProfile to run election on. Note that
+        ballots must have integer weights to be considered valid for this mechanism.
+      m (int): Number of seats to elect.
+      tiebreak (str, optional): Tiebreak method to use. Options are None, 'random', and 'borda'.
+            Defaults to None, in which case a tie raises a ValueError.
+
+
+    """
+
+    def __init__(
+        self, profile: PreferenceProfile, m: int, tiebreak: Optional[str] = None
+    ):
+        self._pv_validate_profile(profile)
+
+        if m <= 0:
+            raise ValueError("m must be positive.")
+        elif m > len(profile.candidates):
+            raise ValueError(
+                "m must be less than or equal to the number of candidates."
+            )
+
+        self.m = m
+        self.tiebreak = tiebreak
+
+        # Decondense ballots
+        ballots = profile.ballots
+        new_ballots = [Ballot() for _ in range(int(profile.total_ballot_wt))]
+        bidx = 0
+        for b in ballots:
+            for _ in range(int(b.weight)):
+                new_ballots[bidx] = Ballot(b.ranking, weight=Fraction(1, 1))
+                bidx += 1
+
+        profile = PreferenceProfile(
+            ballots=tuple(new_ballots), candidates=profile.candidates
+        )
+
+        self.random_order = list(range(int(profile.num_ballots)))
+        np.random.shuffle(self.random_order)
+
+        super().__init__(profile, score_function=first_place_votes)
+
+    def _pv_validate_profile(self, profile: PreferenceProfile):
+        """
+        Validate that each ballot has a ranking and that each
+        ballot has integer weight.
+        """
+        for ballot in profile.ballots:
+            if not ballot.ranking or len(ballot.ranking) == 0:
+                raise TypeError("Ballots must have rankings.")
+            elif int(ballot.weight) != ballot.weight:
+                raise TypeError(f"Ballot {ballot} has non-integer weight.")
+
+    def _is_finished(self) -> bool:
+        cands_elected = [len(s) for s in self.get_elected()]
+        return sum(cands_elected) >= self.m
+
+    def _run_step(
+        self, profile: PreferenceProfile, prev_state: ElectionState, store_states=False
+    ) -> PreferenceProfile:
+        """
+        Run one step of an election from the given profile and previous state.
+        If m candidates have not yet been elected:
+        if the current round is 0, count plurality scores and remove all
+        candidates with a score of 0 to eliminated. Otherwise, this method
+        eliminates a single candidate by decrementing scores of candidates on
+        the current ticket until someone falls to a score of 0. Once we find
+        that there are exactly m candidates remaining on the ticket, elect all
+        of them.
+
+        Args:
+            profile (PreferenceProfile): Profile of ballots.
+            prev_state (ElectionState): The previous ElectionState.
+            store_states (bool, optional): True if `self.election_states` should be updated with the
+                ElectionState generated by this round. This should only be True when used by
+                `self._run_election()`. Defaults to False.
+
[…]
