+++ src/votekit/elections/election_types/ranking/__init__.py
@@ -12,3 +12,6 @@ from .top_two import TopTwo  # noqa
 from .dominating_sets import DominatingSets  # noqa
 from .condo_borda import CondoBorda  # noqa
 from .top_two import TopTwo  # noqa
+from .plurality_veto import PluralityVeto  # noqa
+from .random_dictator import RandomDictator  # noqa
+from .boosted_random_dictator import BoostedRandomDictator  # noqa
