### 2024-08-09 8e205b9
plurality veto tests
