### 2024-08-10 #7088 @mduchin
@samhforbes I think we fixed the affiliations, and the DOIs as much as possible.  (Many of the references are to white papers that do not have DOIs....)  Could you let us know if it looks ok?
