### 2024-08-14 aac570c
Fix plurality veto and add more comprehensive tests
