+++ tests/elections/election_types/ranking/test_plurality_veto.py
@@ -2,10 +2,21 @@ from votekit import Ballot, PreferenceProfile
 from votekit.elections import PluralityVeto
 import random
 import numpy as np
+import itertools
+from joblib import Parallel, delayed
+
+
+# TODO add tests for multiple winners
+
+
+def run_election_once(test_profile):
+    """Run one election and return the winner."""
+    election = PluralityVeto(test_profile, 1)
+    return list(election.get_elected()[0])[0]
 
 
 # TODO make tests in line with other election types
-def test_plurality_veto():
+def test_plurality_veto_simple_3_candidates():
     random.seed(919717)
 
     # simple 3 candidate election
@@ -25,12 +36,111 @@ def test_plurality_veto():
     # count the number of wins over a set of trials
     winner_counts = {c: 0 for c in candidates}
     trials = 2000
-    for t in range(trials):
-        election = PluralityVeto(test_profile, 1)
-        # election.run_election()
-        winner = list(election.get_elected()[0])[0]
-        # winner = list(election.state.winners()[0])[0]
-        winner_counts[winner] += 1
+
+    # Parallel execution
+    n_jobs = -1  # Use all available cores
+    results = Parallel(n_jobs=n_jobs)(
+        delayed(run_election_once)(test_profile) for _ in range(trials)
+    )
+
+    winner_counts = {c: results.count(c) for c in candidates}
 
     # check to make sure that the fraction of wins matches the true probability
-    assert np.allclose(1 / 3, winner_counts["A"] / trials, atol=1e-2)
+    assert np.allclose(1 / 3, winner_counts["A"] / trials, atol=5e-2)
+    assert np.allclose(1 / 3, winner_counts["B"] / trials, atol=5e-2)
+    assert np.allclose(1 / 3, winner_counts["C"] / trials, atol=5e-2)
+
+
+def test_plurality_veto_4_candidates_without_ties():
+    random.seed(919717)
+
+    candidates = ["A", "B", "C", "D"]
+
+    full_power = list(
+        list(
+            list(itertools.permutations(x))
+            for x in itertools.combinations(candidates, r)
+        )
+        for r in range(1, len(candidates) + 1)
+    )
+    powerset = [x for sublist in full_power for item in sublist for x in item]
+
+    ballots = list(map(lambda x: Ballot(ranking=list(set(y) for y in x)), powerset))
+
+    test_profile = PreferenceProfile(ballots=ballots, candidates=candidates)
+
+    trials = 5000
+
+    # Parallel execution
+    n_jobs = -1  # Use all available cores
+    results = Parallel(n_jobs=n_jobs)(
+        delayed(run_election_once)(test_profile) for _ in range(trials)
+    )
+
+    winner_counts = {c: results.count(c) for c in candidates}
+
+    assert np.allclose(1 / 4, winner_counts["A"] / trials, atol=5e-2)
+    assert np.allclose(1 / 4, winner_counts["B"] / trials, atol=5e-2)
+    assert np.allclose(1 / 4, winner_counts["C"] / trials, atol=5e-2)
+    assert np.allclose(1 / 4, winner_counts["D"] / trials, atol=5e-2)
+
+
+def run_election_once_with_ties(test_profile):
+    """Run one election and return the winner."""
+    election = PluralityVeto(test_profile, 1, tiebreak="random")
+    return list(election.get_elected()[0])[0]
+
+
+def test_plurality_veto_4_candidates_with_ties():
+    random.seed(919717)
+
+    candidates = ["A", "B", "C", "D"]
+
+    powerset = list(
[…]
