### 2024-08-14 bbd7c2c
Add more tests for boosted_random_dictator
