+++ tests/elections/election_types/ranking/test_boosted_random_dictator.py
@@ -2,14 +2,20 @@ from votekit import Ballot, PreferenceProfile
 from votekit.elections import BoostedRandomDictator
 import random
 import numpy as np
+from joblib import Parallel, delayed
+import itertools
+from votekit.utils import first_place_votes
 
 
-# TODO make tests in line with other election types
-def test_boosted_random_dictator():
-    # set seed for more predictable results
+def run_election_once(test_profile):
+    """Run one election and return the winner."""
+    election = BoostedRandomDictator(test_profile, 1)
+    return list(election.get_elected()[0])[0]
+
+
+def test_boosted_random_dictator_simple():
     random.seed(919717)
 
-    # simple 3 candidate election
     candidates = ["A", "B", "C"]
     ballots = [
         Ballot(ranking=[{"A"}, {"B"}, {"C"}], weight=3),
@@ -18,15 +24,196 @@ def test_boosted_random_dictator():
     ]
     test_profile = PreferenceProfile(ballots=ballots, candidates=candidates)
 
-    # count the number of wins over a set of trials
     winner_counts = {c: 0 for c in candidates}
-    trials = 3000
-    for t in range(trials):
-        election = BoostedRandomDictator(test_profile, 1)
-        winner = list(election.get_elected()[0])[0]
-        winner_counts[winner] += 1
+    trials = 10000
+
+    # Parallel execution
+    n_jobs = -1  # Use all available cores
+    results = Parallel(n_jobs=n_jobs)(
+        delayed(run_election_once)(test_profile) for _ in range(trials)
+    )
+
+    winner_counts = {c: results.count(c) for c in candidates}
 
     # check to make sure that the fraction of wins matches the true probability
     assert np.allclose(
         1 / 2 * 3 / 5 + 1 / 2 * 9 / 11, winner_counts["A"] / trials, atol=1e-2
+    )
+    assert np.allclose(
+        1 / 2 * 1 / 5 + 1 / 2 * 1 / 11, winner_counts["B"] / trials, atol=1e-2
+    )
+    assert np.allclose(
+        1 / 2 * 1 / 5 + 1 / 2 * 1 / 11, winner_counts["C"] / trials, atol=1e-2
+    )
+
+
+def test_boosted_random_dictator_4_candidates_without_ties():
+    random.seed(919717)
+
+    candidates = ["A", "B", "C", "D"]
+
+    full_power = list(
+        list(
+            list(itertools.permutations(x))
+            for x in itertools.combinations(candidates, r)
+        )
+        for r in range(1, len(candidates) + 1)
+    )
+
+    powerset = [x for sublist in full_power for item in sublist for x in item]
+
+    ballots = list(
+        map(
+            lambda x: Ballot(
+                ranking=list(set(y) for y in x), weight=3 if x[0] == "A" else 1
+            ),
+            powerset,
+        )
+    )
+
+    test_profile = PreferenceProfile(ballots=ballots, candidates=candidates)
+
+    fpv = first_place_votes(test_profile)
+    tot_fpv = sum(fpv.values())
+    tot_fpv_sq = sum(x**2 for x in fpv.values())
+    fpv_sq_dict = {c: v**2 / tot_fpv_sq for c, v in fpv.items()}
+    fpv = {c: v / tot_fpv for c, v in fpv.items()}
+
+    trials = 10000
+
+    # Parallel execution
+    n_jobs = -1  # Use all available cores
[…]
