### 2024-08-15 43bcfb6
Fix distributions for test_boosted_random_dictator
