### 2024-08-15 35c3afd
Fix mypy errors
