### 2024-08-15 efdad59
Prepare changelog, pyproject, and tests for full release
