+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "votekit"
-version = "2.0.1"
+version = "3.0.0"
 description = "A Swiss Army Knife for computational social choice research"
 authors = ["MGGG <engineering@mggg.org>"]
 readme = "README.md"
@@ -23,7 +23,7 @@ numpy = "^1.26.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.0.275"
-black = "^23.3.0"
+black = "^24.8.0"
 pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
 pre-commit = "^3.3.3"
