### 2024-08-23 #7088 @britta-wstnr
Hello @cdonnay 👋
Before inviting reviewers, I have a couple of points for you to address 👷 🙂
1. You can see that our Editorial Bot now flags a few DOIs as faulty (missing or invalid), see [this comment](https://github.com/openjournals/joss-reviews/issues/7088#issuecomment-2303969751). Can you have a look and see if they're fixable?
2. Reading your paper, I saw that it is pretty long. Please note that our word limit is around 1000 words. Can you have a look if the paper can be shortened? [See here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) for an overview of what should be included in your paper (although I think you adhere to this pretty well). Maybe you could move some of the information to your documentation?
If you have any concerns about this, please let me know!

Thanks a lot! 🙏
