### 2024-09-03 7cf3ce9
Update paper.bib
remove extraneous ref for "guide to numerical.."
