+++ paper/paper.bib
@@ -90,12 +90,6 @@
 }
 
 
-@unpublished{GuideExperiments, 
-	author = {Boehmer, Niclas and  Faliszewski, Piotr and   Janeczko, Łukasz and   Kaczmarczyk, Andrzej and   Lisowski, Grzegorz and   Pierczyński, Grzegorz and   Rey, Simon and   Stolicki, Dariusz and   Szufa, Stanisław and   Wąs, Tomasz},
-	title = {Guide to Numerical Experiments on Elections in Computational Social Choice}, 
-	note = {ArXiV preprint. \url{https://arxiv.org/pdf/2402.11765}}
-}
-
 @inproceedings{drawing-a-map,
  author = {Szufa, Stanis\l{}aw and Faliszewski, Piotr and Skowron, Piotr and Slinko, Arkadii and Talmon, Nimrod},
  title = {Drawing a Map of Elections in the Space of Statistical Cultures},
