### 2024-09-03 d3f3caf
Update paper.md
update "guide to numerical" citation
