### 2024-09-03 3506d33
Update paper.bib
update doi for "drawing a map"
