+++ paper/paper.bib
@@ -95,7 +95,7 @@
  title = {Drawing a Map of Elections in the Space of Statistical Cultures},
  year = {2020},
  isbn = {9781450375184},
- doi={10.5555/3398761.3398916},
+ url={https://dl.acm.org/doi/10.5555/3398761.3398916},
  publisher = {International Foundation for Autonomous Agents and Multiagent Systems},
  address = {Richland, SC},
  abstract = {We consider the problem of forming a testbed of elections to be used for numerical experiments (such as testing algorithms or estimating the frequency of a given phenomenon). We seek elections that come from well-known statistical distributions and are as diverse as possible. To this end, we define a (pseudo)metric over elections, generate a set of election instances, and measure distances between them, to assess how diverse they are. Finally, we show how to use these elections to test election-related algorithms.},
