### 2024-09-03 #7088 @cdonnay
Hi @britta-wstnr.

For length, we are thinking of cutting the last two sections: 1) Area of need: resources for research and 2) projects. Do you think that would be appropriate?

We have edited the suggested doi and the invalid doi, but the rest of the articles have no DOI.
