### 2024-09-06 #7088 @britta-wstnr
Hi @cdonnay,

from my point of view, mentioning projects that use your software is quite useful, it's also in our [paper guidelines](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain). Maybe you can list them a little more concisely.
I think you could save some text by making some of your examples (both theoretic and code examples) a little less verbose. You could for example consider moving the details to your documentation and referring there. But I am also curious what @samhforbes thinks about this!

---
Just as a note because I happened to come across this right now: in the last paper link you generated above, some references appear as `?`, e.g. in footnote 4 on page 2 and in line 119.
