### 2024-09-06 #7088 @samhforbes
Hi @cdonnay I'd agree with @britta-wstnr here - I think it's useful and even recommended to have some examples of current use.
It's obviously your paper, but if it were me I'd look to economise on each of the areas of need sections in particular. I think they could certainly go somewhere in software documentation, but the details you provide in this paper don't need to be full ones.
