### 2024-09-09 d58964e
update broken citations
