### 2024-09-18 1b37793
update votekit version requirements
