+++ docs/requirements.txt
@@ -82,6 +82,6 @@ traitlets==5.14.3
 types-pytz==2024.1.0.20240417
 typing_extensions==4.12.1
 urllib3==2.2.1
-votekit==2.0.0
+votekit==3.0.0
 wcwidth==0.2.13
 webencodings==0.5.1
