### 2024-09-18 f73562f
update POT requirement docs
