+++ docs/requirements.txt
@@ -44,7 +44,7 @@ parso==0.8.4
 pexpect==4.9.0
 pillow==10.3.0
 platformdirs==4.2.2
-POT==0.9.0
+POT==0.9.4
 prompt_toolkit==3.0.45
 psutil==5.9.8
 ptyprocess==0.7.0
