### 2024-09-23 #7088 @cdonnay
Howdy! We have shortened our draft as much as we feel comfortable;  how does it look on your end?
