### 2024-10-02 #7088 @britta-wstnr
Hey @cdonnay - paper looks good now! Thanks for shortening it considerably.
I will go ahead and start inviting reviewers now - are there any reviewers you would suggest? That would help a lot - if you want to suggest reviewers, please do so _without_ @-ing them. See [this comment](https://github.com/openjournals/joss-reviews/issues/7088#issue-2455484496) from the editorial 🤖 for more information on suggesting reviewers.
