### 2024-11-13 #7477 @editorialbot
[REVIEW]: VoteKit: A Python package for computational social choice research
**Submitting author:** @cdonnay (<a href="http://orcid.org/0000-0003-4498-4067">Moon Duchin</a>)
**Repository:** https://github.com/mggg/votekit
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 3.1.0
