### 2024-11-23 #7477 @FATelarico
## General checks

The authors clearly state the problems the software is designed to solve. Also, the paper provides a clear high-level description of the functionality and purpose of the software, and it is accessible to a diverse, non-specialist audience.

## Software paper

Overall, the paper is well-written and does not require significant edits for structure, language, or quality. Minor improvements in phrasing and additional context for visual aids could enhance its presentation, but these are not critical issues.

Some minor improvements that may benefit the paper:

* Some phrases, such as references to "gaps" (rows 31, 40, 46)  in computational social choice software, are repeated multiple times. This could be streamlined to avoid redundancy.

* The included figures are relevant but could be accompanied by more descriptive captions to enhance their standalone interpretability.

* While the examples provided are helpful, including a step-by-step demonstration or explicit code snippets would make the descriptions more concrete.

A  citation lack complete details (on p. 1):

> Endriss and Simon Rey, n.d.
