### 2024-12-09 #7477 @britta-wstnr
Hi @andrewheiss, gentle reminder of your open review here at JOSS. Would you mind creating the checklist so that we can see that everything is working as it should? Thanks! 🙏

_Also sent via email in case Github notifications are off._
