### 2024-12-10 mggg/VoteKit#154 @andrewheiss
JOSS review thread
Hello! I'm reviewing this package/paper for JOSS (https://github.com/openjournals/joss-reviews/issues/7477), and following their recommendation, I'll leave comments and suggestions as comments here.
