### 2024-12-10 mggg/VoteKit#154 @andrewheiss
Super minor thing (and I can't find the LaTeX file that generates the paper in this repo—it likely lives elsewhere): Figures 1 and 2 in the paper aren't vector based. This is fine for Figure 1 since it uses a high enough resolution that it looks okay, but Figure 2 appears quite pixelated (almost unreadable) when printed. Including SVGs or PDFs in the paper manuscript (or higher resolution PNGs) would be helpful.
