### 2024-12-10 mggg/VoteKit#154 @andrewheiss
Everything installed just fine for me both via pip and through `poetry`—it was my first time using poetry and it was really neat that it automatically created a virtualenv and installed everything

I did get an error initially, since my system python is 3.13, and the project settings look for less than 3.13:

https://github.com/mggg/VoteKit/blob/1d739fef11699bcd8821c4d08a5958637b8d9eca/pyproject.toml#L9

…but poetry ended up automatically installing a 3.9 virtualenv for me. Perhaps allowing 3.13 could be helpful
