### 2024-12-10 mggg/VoteKit#154 @andrewheiss
The current repo uses Sphinx to build https://votekit.readthedocs.io/en/latest/, which seems to be the actual documentation. It uses the comprehensive ipynbs in https://github.com/mggg/VoteKit/tree/main/notebooks to provide a helpful tutorial, and for my part, running those notebooks interactively on my own was the best way for me to get familiar with the package and its functionality and output. These are great and should be kept as the main documentation tutorial.

There's a link to an older, less-organized and less-polished version of the documentation (https://mggg.github.io/VoteKit/) that I accidentally used as my initial entry to the documentation and it took a while to figure out why it didn't match what was in the current repo:

https://github.com/mggg/VoteKit/blob/1d739fef11699bcd8821c4d08a5958637b8d9eca/README.md?plain=1#L22
