### 2024-12-10 mggg/VoteKit#154 @andrewheiss
JOSS recommends having clear contributor guidelines:

> **Community guidelines**: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

The project currently has a [code of conduct](CODE_OF_CONDUCT.md), and the README states that "bug reports and feature requests, as well as contributions, are welcome", but it doesn't provide guidelines for how to do that (like issue templates, recommendations for contributors, style preferences, etc.)

It might be helpful to include a CONTRIBUTING.md file—see https://github.com/jessesquires/.github/blob/main/CONTRIBUTING.md for a good generic one that I've adapted from several times, or [this shorter R-specific one](https://github.com/andrewheiss/compassionate-clam/blob/main/CONTRIBUTING.md), for examples.
