### 2024-12-10 #7477 @andrewheiss
I've included all my issues and suggestions as comments in https://github.com/mggg/VoteKit/issues/154

In general, this is a really neat and comprehensive package for election analysis and simulation, and I'm happy with the documentation (beyond [a super minor old link that points to older documentation](https://github.com/mggg/VoteKit/issues/154#issuecomment-2530429946)). As noted in https://github.com/mggg/VoteKit/issues/154, all my issues are minor—adding guidelines for contributors, cleaning up some of the graphics, and so on.

I also agree with @FATelarico that the graphics need additional context and explanation in the JOSS paper. They're well-explained in [`notebooks/3_viz.ipynb`](https://github.com/mggg/VoteKit/blob/main/notebooks/3_viz.ipynb).
