### 2024-12-13 #7477 @tuvavbardal
The package serves as a valuable contribution to the range of COMSOC tools used to study elections. The paper is well-written and clearly situates the software within the broader landscape of software developed in the field, while offering a good, accessible overview of high-level functionality.

I generally found the package to be well-documented, but I will echo @andrewheiss in that guidelines for contributors should be added.

For the paper, I agree with the other reviewers that figures would benefit from more context. Furthermore, I would add URLs to access tools and websites to the bibliography (e.g., https://comsoc-community.org/ in line 188).
