### 2024-12-20 mggg/VoteKit#154 @cdonnay
Thank you for the super detailed feedback! We really appreciate it.
