+++ paper/paper.bib
@@ -6,6 +6,7 @@
 	note = {Python Library. \url{https://github.com/mggg/VoteKit}},
   year      = {2023},
   month     = {8},
+  url = {https://github.com/mggg/VoteKit}
 }
 
 @unpublished{preflibtools, 
@@ -13,6 +14,7 @@
 	title = {PrefLib-Tools}, 
 	note = {Python Library. \url{https://github.com/PrefLib/preflibtools}},
   year      = {2022},
+  url = {https://github.com/PrefLib/preflibtools}
 }
 
 @unpublished{prefvoting, 
@@ -20,12 +22,14 @@
 	title = {pref_voting}, 
 	note = {Python Library. \url{https://github.com/voting-tools/pref_voting}},
   year      = {2022},
+  url = {https://github.com/voting-tools/pref_voting}
 }
 
 @unpublished{votelib, 
 	author = {{Jan \v{S}imbera}},
 	title = {Votelib}, 
 	note = {Python Library. \url{https://github.com/simberaj/votelib}},
+  url = {https://github.com/simberaj/votelib},
   year      = {2020},
 }
 
@@ -34,6 +38,7 @@
 	title = {apportionment}, 
 	note = {Python Library. \url{https://github.com/martinlackner/apportionmentb}},
   year      = {2022},
+  url = {https://github.com/martinlackner/apportionmentb}
 }
 
 
@@ -55,7 +60,9 @@
 @unpublished{Scot-Elex, 
 	author = {{MGGG Redistricting Lab}},
 	title = {Scottish STV Election Repo}, 
-	note = {Data Repository. \url{https://github.com/mggg/scot-elex}}
+  year = {2024},
+	note = {Data Repository. \url{https://github.com/mggg/scot-elex}},
+  url = {https://github.com/mggg/scot-elex}
 }
 
 
@@ -65,7 +72,8 @@
 	note = {GitHub Repository and Community Resources. 
        \url{https://github.com/comsoc-community/comsoc-website}, 
         \url{https://comsoc-community.org/}
-           }
+           },
+  url = {https://comsoc-community.org/tools}
 }
 
 @inproceedings{boehmer2024guidenumericalexperimentselections,
@@ -86,7 +94,9 @@
 @unpublished{RCV-Cruncher, 
 	author = {{FairVote}},
 	title = {RCV Cruncher}, 
-	note = {Python Library. \url{https://github.com/fairvotereform/rcv_cruncher/}}
+  year      = {2024},
+	note = {Python Library. \url{https://github.com/fairvotereform/rcv_cruncher/}},
+  url = {https://github.com/fairvotereform/rcv_cruncher/}
 }
 
 
