+++ paper/paper.bib
@@ -60,7 +60,7 @@
 @unpublished{Scot-Elex, 
 	author = {{MGGG Redistricting Lab}},
 	title = {Scottish STV Election Repo}, 
-  year = {2024},
+  year = {2023},
 	note = {Data Repository. \url{https://github.com/mggg/scot-elex}},
   url = {https://github.com/mggg/scot-elex}
 }
@@ -69,6 +69,7 @@
 @unpublished{ComSoc, 
 	author = {{Ulle Endriss and Simon Rey}},
 	title = {COMSOC Community Site}, 
+  year = {2024},
 	note = {GitHub Repository and Community Resources. 
        \url{https://github.com/comsoc-community/comsoc-website}, 
         \url{https://comsoc-community.org/}
