+++ src/votekit/new-feature.py
@@ -1,1 +1,0 @@
-print("i made a new feature!")
