+++ pyproject.toml
@@ -28,14 +28,10 @@ pytest = "^7.4.0"
 pytest-cov = "^4.1.0"
 pre-commit = "^3.3.3"
 ipython = "^8.17.2"
-mkdocs = "^1.5.3"
-mkdocs-autorefs = "^0.5.0"
-mkdocs-click = "^0.8.1"
-mkdocs-material = "^9.4.10"
-mkdocs-material-extensions = "^1.3"
-mkdocs-table-reader-plugin = "^2.0.3"
-mkdocstrings = "^0.24.0"
-mkdocstrings-python = "^1.7.4"
+sphinx = "^7.4.7"
+sphinx-copybutton = "^0.5.2"
+recommonmark = "^0.7.1"
+sphinx-rtd-theme = "^3.0.2"
 
 [build-system]
 requires = ["poetry-core"]
