+++ docs/package_info/contributing.rst
@@ -21,22 +21,22 @@ While it is not feasible to test every single aspect of a package, adding tests 
 of open-source software. You can use our `current tests <https://github.com/mggg/VoteKit/blob/main/tests>`_  as a starting place 
 to show you how these work. Depending on the scale of your feature, you may need multiple tests.
 At minimum, your tests should 
+
 - check the basic functionality of your feature, such as a test case of an algorithm or an instantiation of a class, and 
 - ensure that any error messages are correctly raised.
 
 **Documentation**: For any new features that you add, please make sure to include
 a comprehensive docstring. We have a defined format for docstrings that we use
-throughout the `codebase <https://github.com/mggg/VoteKit/blob/main/src/votekit>`_ codebase, so please make sure that any additions are consistent
+throughout the `codebase <https://github.com/mggg/VoteKit/blob/main/src/votekit>`_, so please make sure that any additions are consistent
 with that format.
 
 Poetry
 =============
 
-VoteKit uses Poetry to manage the package and all of its dependencies for developers.
-This is not needed for standard users.
-Within your `virtual environment <../user/install.rst>`_, run ``pip install poetry``.
+VoteKit uses Poetry to manage the package and all of its dependencies; Poetry is not needed for standard users, but developers will find it useful.
+To install Poetry, within your `virtual environment <../user/install.rst>`_, run ``pip install poetry``.
 Then, run ``poetry install`` from within the root of your VoteKit project directory to install all dependencies. 
-This will add the necessary dependencies to a `.venv` directory.
+This will add the necessary dependencies to a ``.venv`` directory.
 Once you've run ``poetry install``,  run ``poetry run pre-commit install``, which will install code linting hooks that will run on every commit. 
 This helps ensure code quality, like automatically checking formatting.
 
@@ -57,7 +57,7 @@ Quick Start via Command Line
 4. **Create a new branch** for your contribution ``git checkout -b my-new-feature``.
 5. **Make your changes** and commit them ``git commit -am 'Add some feature'``.
 6. **Run tests and linters** with ``poetry run pytest``, ``poetry run ruff src tests``, 
-``poetry run mypy src``, and ``black src``.
+    ``poetry run mypy src``, and ``black src``.
 7. **Pull the latest changes from upstream** and rebase your branch if necessary.
 8. **Push your branch** to GitHub ``git push origin my-new-feature``.
 9. **Open a Pull Request** on GitHub.
@@ -65,30 +65,65 @@ Quick Start via Command Line
 Quick Start via Desktop GUI
 ============================
 
+1. **Fork the repository** on GitHub. 
 
-1. **Fork the repository** on GitHub.
     .. image:: ../_static/assets/github_desktop/fork_repo.png
         :align: center
-2. **Clone the repository locally**.
-3. **Create a new branch** for your feature and publish the branch.
-4. **Make your changes** and commit them to your branch.
+
+2. **Clone the repository locally**. Copy the url to the repository from GitHub.
+
+    .. image:: ../_static/assets/github_desktop/clone_repo_1.png
+        :align: center
+
+    Then from the "File" menu on the GitHub desktop GUI, select "Clone Repository" and provide the url.
+
+3. **Create a new branch** for your feature by selecting "New Branch" from the "Branch" menu.
+
+    .. image:: ../_static/assets/github_desktop/new_branch_1.png
+        :align: center
+
+Publish the branch.
+
+    .. image:: ../_static/assets/github_desktop/publish_branch.png
+        :align: center
+
+4. **Make your changes** to the VoteKit code base and commit them to your branch.
+
+    .. image:: ../_static/assets/github_desktop/commit.png
+        :align: center
+
 5. **Run tests and linters** in the command line at the root of your VoteKit directory
[…]
