+++ docs/package_info/issues.rst
@@ -1,15 +1,20 @@
-================
-Reporting Issues
-================
+====================================
+Reporting Issues or Feature Requests
+====================================
 
 Sometimes, you may encounter a problem with VoteKit, and that is okay! We would
 love to hear about it so that we can fix it. If you can provide a minimal example that
 reproduces the issue, please feel free to report it on the `GitHub issue tracker
 <https://github.com/mggg/VoteKit/issues>`_.
+If you believe you have a fix for a current bug, feel free to open a pull request, following our
+guidelines for `contributing <contributing.rst>`_.
 
 In the event that you would prefer not to post the issue publicly, you can also email
 us at "code[at]mggg[dot]org". We will do our best to respond to your issue in a timely
 manner. Thank you for your help in making VoteKit better! 
 
-If you believe you have a fix for a current bug, feel free to open a pull request, following our
+You can also use the `GitHub issue tracker
+<https://github.com/mggg/VoteKit/issues>`_ to submit feature requests. VoteKit is open-source, so you can also submit code for new features, following our
 guidelines for `contributing <contributing.rst>`_.
+
+
