+++ .github/workflows/test.yml
@@ -22,7 +22,9 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: install Poetry
-      uses: snok/install-poetry@v1.3.3
+      uses: snok/install-poetry@v1
+      with:
+        version: 1.8.5
     - name: install dependencies
       run: poetry install
     - name: lint with ruff
