+++ pyproject.toml
@@ -1,8 +1,8 @@
 [tool.poetry]
 name = "votekit"
 version = "3.0.0"
-description = "A Swiss Army Knife for computational social choice research"
-authors = ["MGGG <engineering@mggg.org>"]
+description = "A Swiss army knife for computational social choice research."
+authors = ["MGGG <code@mggg.org>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
