### 2025-01-06 0afb0b2
add doc string example
