+++ pyproject.toml
@@ -1,12 +1,12 @@
 [tool.poetry]
 name = "votekit"
-version = "3.0.0"
+version = "3.1.0"
 description = "A Swiss army knife for computational social choice research."
 authors = ["MGGG <code@mggg.org>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.9,<3.13"
+python = "^3.9,<3.14"
 scipy = "^1.11.1"
 mypy = "^1.4.1"
 pydantic = "^2.8.2"
