+++ paper/paper.bib
@@ -9,6 +9,13 @@
   url = {https://github.com/mggg/VoteKit}
 }
 
+@unpublished{VoteKitDocs, 
+	author = {{MGGG Redistricting Lab}},
+	title = {VoteKit Documentation}, 
+  year      = {2024},
+  url = {https://votekit.readthedocs.io/en/latest/}
+}
+
 @unpublished{preflibtools, 
 	author = {{Nicholas Mattei and Simon Rey}},
 	title = {PrefLib-Tools}, 
