+++ paper/paper.md
@@ -165,7 +165,7 @@ Reform advocates also need to describe voting mechanisms and their
 likely outcomes effectively to members of their communities. The end-to-end pipeline provided by `VoteKit` allows advocates to toggle different system settings and compare expected outcomes. For example, 
 Figure 1 is reprinted from a report on reform proposals for the chambers of the Washington state legislature, with Systems 0-3 as paired bicameral systems and Systems 4-5 as unicameral solutions.  Using the codebase that formed the foundation of `VoteKit`, researchers compared the expected outcomes for minority representation under these six systems.
 
-![A comparison of a variety of electoral systems and their effect on minority representation in a [case study](https://mggg.org/washington) of reform proposals for the Washington state legislature [@washington_leg]. It can be seen that certain systems, like System 0, are projected to be less successful for minority represenation, while other systems, like System 1, tend to secure proportional representation for minorities.](./figures/WA_poc_seats_chartsystem_compare_pared.png){width=100%}
+![A comparison of a variety of electoral systems and their effect on minority representation in a [case study](https://mggg.org/washington) of reform proposals for the Washington state legislature [@washington_leg]. It can be seen that certain systems, like System 0, are projected to be less successful for minority representation, while other systems, like System 1, tend to secure proportional representation for minorities.](./figures/WA_poc_seats_chartsystem_compare_pared.png){width=100%}
 
 
 ## Area of need: Resources for research
