### 2025-01-09 0aaf6ea
create bar plot
