+++ src/votekit/plots/__init__.py
@@ -1,2 +1,2 @@
 from .mds import plot_MDS, compute_MDS  # noqa
-from .profile_plots import plot_summary_stats  # noqa
+from .profile_plots import plot_summary_stats, bar_plot  # noqa
