+++ src/votekit/plots/profile_plots.py
@@ -1,22 +1,467 @@
 from matplotlib import pyplot as plt  # type: ignore
 from votekit.pref_profile import PreferenceProfile
 from votekit.utils import first_place_votes, mentions, COLOR_LIST, borda_scores
+from matplotlib.axes import Axes
+from typing import Optional, Union, Tuple, Literal
+import matplotlib.patches as mpatches
+from copy import deepcopy
+
+
+def _validate_bar_plot_args(
+    data: list[dict[str, float]],
+    data_set_labels: list[str],
+    x_label_ordering: list[str],
+):
+    """
+    Validates bar plot arguments. Raises ValueError.
+
+    Args:
+        data (list[dict[str, float]]): Categorical data to be plotted. The value of the dict
+            should be the frequency of the key which is the category name.
+        data_set_labels (list[str], optional): List of labels for data sets. Must be on per data
+            set and must be unique.
+        x_label_ordering (list[str]): Ordering of x labels.
+
+    """
+    if len(set(data_set_labels)) != len(data_set_labels):
+        raise ValueError("Data set labels must be unique.")
+
+    if len(data_set_labels) != len(data):
+        raise ValueError("There must be one data set label for each data set.")
+
+    if set(x_label_ordering) != set(k for data_dict in data for k in data_dict.keys()):
+        raise ValueError("x_label_ordering must match the keys of the data.")
+
+
+def _set_default_bar_plot_args(
+    data: Union[list[dict[str, float]], dict[str, float]],
+    data_set_labels: Optional[list[str]],
+    data_set_to_color: Optional[dict],
+    bar_width: Optional[float],
+    x_label_ordering: Optional[list[str]],
+    ax: Optional[Axes],
+) -> Tuple[list[dict[str, float]], list[str], dict, float, list[str], Axes]:
+    """
+    Handles setting default arguments.
+
+    Args:
+        data (Union[list[dict[str, float]], dict[str, float]]): Categorical data to be plotted.
+            If a list is provided, will plot all data sets on same plot. The value of the dict
+            should be the frequency of the key which is the category name.
+        data_set_labels (list[str], optional): List of labels for data sets.
+        data_set_to_color (dict, optional): Dictionary mapping data set labels to colors. Defaults
+            to subset of ``COLOR_LIST`` from ``utils`` module.
+        bar_width (float, optional): Width of bars, defaults to None which computes the bar width
+            as 0.7 divided by the number of data sets.
+        x_label_ordering (list[str], optional): Ordering of x labels. Defaults to order retrieved
+            from data dictionary.
+        ax (Axes, optional): A matplotlib axes object to plot the figure on. Defaults to None, in
+            which case the function creates and returns a new axes.
+
+    Returns:
+        Tuple[list[dict[str, float]], list[str], dict, float, list[str], Axes]: data,
+        data_set_labels, data_set_to_color, bar_width, x_label_ordering, ax
+    """
+    if not isinstance(data, list):
+        data = [data]
+
+    if not data_set_labels:
+        data_set_labels = [f"Data Set {i+1}" for i in range(len(data))]
+
+    if not data_set_to_color:
+        data_set_to_color = {d: COLOR_LIST[i] for i, d in enumerate(data_set_labels)}
+
+    if not bar_width:
+        bar_width = 0.7 / len(data)
+
+    if ax is None:
+        fig, ax = plt.subplots()
+
+    if not x_label_ordering:
+        x_label_ordering = list(
+            set([k for data_dict in data for k in data_dict.keys()])
+        )
+
+    return data, data_set_labels, data_set_to_color, bar_width, x_label_ordering, ax
+
+
[…]
