+++ src/votekit/utils.py
@@ -7,20 +7,45 @@ from .ballot import Ballot
 from .pref_profile import PreferenceProfile
 
 COLOR_LIST = [
-    (0.55, 0.71, 0.0),
-    (0.82, 0.1, 0.26),
-    (0.44, 0.5, 0.56),
-    (1.0, 0.75, 0.0),
-    (1.0, 0.77, 0.05),
-    (0.0, 0.42, 0.24),
-    (0.13, 0.55, 0.13),
-    (0.9, 0.13, 0.13),
-    (0.08, 0.38, 0.74),
-    (0.41, 0.21, 0.61),
-    (1.0, 0.72, 0.77),
-    (1.0, 0.66, 0.07),
-    (1.0, 0.88, 0.21),
-    (0.55, 0.82, 0.77),
+    "#0099cd",
+    "#ffca5d",
+    "#00cd99",
+    "#99cd00",
+    "#cd0099",
+    "#9900cd",
+    "#8dd3c7",
+    "#bebada",
+    "#fb8072",
+    "#80b1d3",
+    "#fdb462",
+    "#b3de69",
+    "#fccde5",
+    "#bc80bd",
+    "#ccebc5",
+    "#ffed6f",
+    "#ffffb3",
+    "#a6cee3",
+    "#1f78b4",
+    "#b2df8a",
+    "#33a02c",
+    "#fb9a99",
+    "#e31a1c",
+    "#fdbf6f",
+    "#ff7f00",
+    "#cab2d6",
+    "#6a3d9a",
+    "#b15928",
+    "#64ffda",
+    "#00B8D4",
+    "#A1887F",
+    "#76FF03",
+    "#DCE775",
+    "#B388FF",
+    "#FF80AB",
+    "#D81B60",
+    "#26A69A",
+    "#FFEA00",
+    "#6200EA",
 ]
 
 
