+++ src/votekit/plots/profile_plots.py
@@ -5,12 +5,43 @@ from matplotlib.axes import Axes
 from typing import Optional, Union, Tuple, Literal
 import matplotlib.patches as mpatches
 from copy import deepcopy
+import warnings
+from matplotlib.legend import Legend
+
+
+def add_null_keys(data: list[dict[str, float]]) -> list[dict[str, float]]:
+    """
+    Prepares list of dictionaries to be passed to ``bar_plot()``. If a key is missing from
+    a dictionary, this function adds the key with value 0.
+
+    Args:
+        data (list[dict[str, float]]): Categorical data to be cleaned. The value of each dict
+            should be the frequency of the key which is the category name.
+
+    Returns:
+        list[dict[str, float]]: Cleaned data.
+    """
+    data = deepcopy(data)
+
+    x_labels = []
+    for data_dict in data:
+        for x_label in data_dict.keys():
+            if x_label not in x_labels:
+                x_labels.append(x_label)
+
+    for x_label in x_labels:
+        for data_dict in data:
+            if x_label not in data_dict:
+                data_dict[x_label] = 0
+
+    return data
 
 
 def _validate_bar_plot_args(
     data: list[dict[str, float]],
     data_set_labels: list[str],
     x_label_ordering: list[str],
+    bar_width: float,
 ):
     """
     Validates bar plot arguments. Raises ValueError.
@@ -20,17 +51,39 @@ def _validate_bar_plot_args(
             should be the frequency of the key which is the category name.
         data_set_labels (list[str], optional): List of labels for data sets. Must be on per data
             set and must be unique.
-        x_label_ordering (list[str]): Ordering of x labels.
+        x_label_ordering (list[str]): Ordering of x labels. Must match data keys.
+        bar_width (float): Width of bars. Must be between 0 and 1/len(data).
 
     """
+    if any(set(data_dict.keys()) != set(data[0].keys()) for data_dict in data):
+        raise ValueError("All data dictionaries must have the same keys.")
+
     if len(set(data_set_labels)) != len(data_set_labels):
-        raise ValueError("Data set labels must be unique.")
+        raise ValueError("Each data set must have a unique label.")
 
     if len(data_set_labels) != len(data):
         raise ValueError("There must be one data set label for each data set.")
 
-    if set(x_label_ordering) != set(k for data_dict in data for k in data_dict.keys()):
-        raise ValueError("x_label_ordering must match the keys of the data.")
+    sym_dif = set(x_label_ordering) ^ set(
+        k for data_dict in data for k in data_dict.keys()
+    )
+    if len(sym_dif) != 0:
+        raise ValueError(
+            (
+                "x_label_ordering must match the keys of the data. Here is the symmetric "
+                f"difference of the two sets: {sym_dif}"
+            )
+        )
+
+    if bar_width <= 0:
+        raise ValueError("Bar width must be positive.")
+
+    elif bar_width > 1 / len(data):
+        warnings.warn(
+            "Bar width must be less than 1. Bar width has now been set to 1.",
+            category=UserWarning,
+        )
+        bar_width = 1 / len(data)
 
 
 def _set_default_bar_plot_args(
@@ -39,8 +92,11 @@ def _set_default_bar_plot_args(
     data_set_to_color: Optional[dict],
     bar_width: Optional[float],
     x_label_ordering: Optional[list[str]],
+    legend_font_size: Optional[float],
     ax: Optional[Axes],
-) -> Tuple[list[dict[str, float]], list[str], dict, float, list[str], Axes]:
+) -> Tuple[
+    list[dict[str, float]], list[str], dict, float, list[str], float, float, Axes
+]:
     """
[…]
