### 2025-01-13 #7477 @britta-wstnr
Hi @cdonnay once the you are happy with your revisions, the reviewers will have a look. Once they are happy too (which is usually an iterative process), I will have another look at the paper (checking for formatting and such). Then we have to get some admin done (e.g., creating an archive). Then I will hand off to the EiC :) ✨
Does that make the process clearer for you?
