+++ src/votekit/plots/profile_plots.py
@@ -2,7 +2,7 @@ from matplotlib import pyplot as plt  # type: ignore
 from votekit.pref_profile import PreferenceProfile
 from votekit.utils import first_place_votes, mentions, COLOR_LIST, borda_scores
 from matplotlib.axes import Axes
-from typing import Optional, Union, Tuple, Literal
+from typing import Optional, Union, Tuple, Literal, Callable
 import matplotlib.patches as mpatches
 from copy import deepcopy
 import warnings
@@ -35,55 +35,6 @@ def add_null_keys(data: list[dict[str, float]]) -> list[dict[str, float]]:
                 data_dict[x_label] = 0
 
     return data
-
-
-def _validate_bar_plot_args(
-    data: list[dict[str, float]],
-    data_set_labels: list[str],
-    x_label_ordering: list[str],
-    bar_width: float,
-):
-    """
-    Validates bar plot arguments. Raises ValueError.
-
-    Args:
-        data (list[dict[str, float]]): Categorical data to be plotted. The value of the dict
-            should be the frequency of the key which is the category name.
-        data_set_labels (list[str], optional): List of labels for data sets. Must be on per data
-            set and must be unique.
-        x_label_ordering (list[str]): Ordering of x labels. Must match data keys.
-        bar_width (float): Width of bars. Must be between 0 and 1/len(data).
-
-    """
-    if any(set(data_dict.keys()) != set(data[0].keys()) for data_dict in data):
-        raise ValueError("All data dictionaries must have the same keys.")
-
-    if len(set(data_set_labels)) != len(data_set_labels):
-        raise ValueError("Each data set must have a unique label.")
-
-    if len(data_set_labels) != len(data):
-        raise ValueError("There must be one data set label for each data set.")
-
-    sym_dif = set(x_label_ordering) ^ set(
-        k for data_dict in data for k in data_dict.keys()
-    )
-    if len(sym_dif) != 0:
-        raise ValueError(
-            (
-                "x_label_ordering must match the keys of the data. Here is the symmetric "
-                f"difference of the two sets: {sym_dif}"
-            )
-        )
-
-    if bar_width <= 0:
-        raise ValueError("Bar width must be positive.")
-
-    elif bar_width > 1 / len(data):
-        warnings.warn(
-            "Bar width must be less than 1. Bar width has now been set to 1.",
-            category=UserWarning,
-        )
-        bar_width = 1 / len(data)
 
 
 def _set_default_bar_plot_args(
@@ -154,6 +105,55 @@ def _set_default_bar_plot_args(
         legend_font_size,
         ax,
     )
+
+
+def _validate_bar_plot_args(
+    data: list[dict[str, float]],
+    data_set_labels: list[str],
+    x_label_ordering: list[str],
+    bar_width: float,
+):
+    """
+    Validates bar plot arguments. Raises ValueError.
+
+    Args:
+        data (list[dict[str, float]]): Categorical data to be plotted. The value of the dict
+            should be the frequency of the key which is the category name.
+        data_set_labels (list[str], optional): List of labels for data sets. Must be on per data
+            set and must be unique.
+        x_label_ordering (list[str]): Ordering of x labels. Must match data keys.
+        bar_width (float): Width of bars. Must be between 0 and 1/len(data).
+
+    """
+    if any(set(data_dict.keys()) != set(data[0].keys()) for data_dict in data):
+        raise ValueError("All data dictionaries must have the same keys.")
+
+    if len(set(data_set_labels)) != len(data_set_labels):
+        raise ValueError("Each data set must have a unique label.")
[…]
