+++ paper/paper.md
@@ -68,23 +68,21 @@ Users can:
     visualizations* to compare and analyze profiles and election
     outcomes.
 
-A tutorial that includes step-by-step example code for all of the aforementioned topics can be found in the VoteKit documentation [@VoteKitDocs].
+A tutorial that includes step-by-step example code can be found in the VoteKit documentation [@VoteKitDocs].
 
 # Statement of need
 
 <!-- Social choice theory grew out of welfare economics in the mid-twentieth
 century and has been recognized as a deep and highly applicable area of
 economic theory, forming part of the basis for at least four Nobel Prize
-awards.[^2] Since the 1990s, a new fusion of economics and computer
+awards.[^2]--> 
+Since the 1990s, a fusion of economics and computer
 science has emerged under the name of *computational social choice*,
 studying questions of complexity and design and further advancing the
-axiomatic study of elections.[^3] But most of these innovations have
+axiomatic study of elections. But most of these innovations have
 been highly abstract, and there has been a significant gap in the
 literature---and in the landscape of software---between the theory and
-the practice of democracy. -->
-There has been a significant gap in the
-literature of computational social choice---and in the landscape of software---between 
-the theory and the practice of democracy.
+the practice of democracy. 
 On the software side, researchers have built a multitude of different
 packages for generating and analyzing elections.[^4] 
 Most packages, to our knowledge, handle just one part of the research arc; for instance, 
@@ -109,10 +107,10 @@ given election style; profiles can be sampled from a generative model to
 produce simulated or synthetic elections. Having sources of rich,
 varied, and realistic data is essential to an empirically grounded
 research program to probe the properties of voting rules. Good
-generative models are also essential to advise reformers deciding
-between options in a new locality, as they enable generation of
-synthetic profiles keyed to the scale, demographics, and election specs
-of that specific place. 
+generative models are also a fundamental tool to advise reformers deciding
+between alternative electoral systems in a new locality, as they enable generation of
+synthetic profiles keyed to the scale, demographics, and election styles
+considered for that specific place. 
 <!-- But most of the models in the literature, like
 the Impartial Culture model (all permutations of candidates are equally
 likely) or the Impartial Anonymous Culture model (sampling proportional
@@ -125,30 +123,30 @@ describe rankings in actual elections\" [@Tideman2010TheSO]. They
 therefore recommend *spatial models* instead, which themselves are of
 dubious realism for the selection of political candidates.[^5] -->
 
-`VoteKit` implements many of the models typically used in computational social choice, as
-well as newer mathematical models that give users the ability to
+`VoteKit` implements many of the models typically used in computational social choice research,[^5] as
+well as newer parametrized models that give users the ability to
 generate profiles that are designed to comport with real-world ranking
 behavior and particularly to generate polarized elections. Two leading
 choices are based on classic statistical ranking mechanisms, called the
 Plackett--Luce (PL) and Bradley--Terry (BT) models; another model called
 the Cambridge Sampler (CS) draws from historical ranking data in
 Cambridge, MA city council elections. These models have flexible
-parameters, allowing users to vary voting bloc proportions, candidate
-strength within slates, and polarization between blocs. These parameters
-can be specified or randomly sampled.
+parameters---allowing users to vary voting bloc proportions, candidate
+strength within slates, and polarization between blocs---that can be 
+specified or randomly sampled.
 
 ## Area of need: Comparison and communication
 
-In the realm of democracy reform, groups of stakeholders often ask
[…]
