+++ paper/paper.bib
@@ -230,8 +230,8 @@ WA School District},
 
 
 @article{duchin_tapp_24,
-  author = "Duchin and Tapp",
-  title  = "Learning blocs and slates from observed elections",
+  author = "Duchin and Shmoys and Tapp",
+  title  = "Learning blocs and slates from ranked-choice ballots",
   year   = "2024",
   journal = "Preprint"
 }
