### 2025-01-14 #7477 @cdonnay
Alright, we have updated the paper with the revisions and are ready for another look from the reviewers! I have included a summary of our responses to each comment.

- Some phrases, such as references to “gaps” (rows 31, 40, 46) in computational social choice software, are repeated multiple times. This could be streamlined to avoid redundancy. -> We have done so.
- The included figures are relevant but could be accompanied by more descriptive captions to enhance their standalone interpretability. They are well explained in jupyter notebook 3_viz. -> We have done so.
- While the examples provided are helpful, including a step-by-step demonstration or explicit code snippets would make the descriptions more concrete. -> We have added a link to our tutorials in the introduction. We did this to save space.
- A citation lack complete details (on p. 1): Endriss and Simon Rey, n.d.-> We have fixed this.
- a super minor old link that points to older documentation : https://github.com/mggg/VoteKit/issues/154#issuecomment-2530429946 -> PR that fixes this has been merged
- guidelines for contributors should be added -> we have added this
- add URLs to access tools and websites to the bibliography -> we have added this
- figure 1 and 2 should be vector based images -> we made fig 2 higher resolution
- add support for python 3.13 -> will be added in the next release of VoteKit
