### 2025-01-20 #7477 @FATelarico
Sorry for the delay. I took some additional time to try the package again. As far as I am concerned the paper's good to go.

Congrats!
