### 2025-01-21 #7477 @britta-wstnr
@andrewheiss and @FATelarico: thanks a lot for your confirmation and your work on this review! I appreciate it! 🙏
@andrewheiss could you have a quick look at your checklist? For me, the point "Community Guidelines" appears to still be open.
