+++ src/votekit/plots/profile_plots.py
@@ -10,7 +10,6 @@ from matplotlib.legend import Legend
 from matplotlib.lines import Line2D
 
 VALID_STATS = ["fpv", "mentions", "borda", "first place votes"]
-# Valid_stat = Literal[*VALID_STATS]
 Valid_stat = Literal["fpv", "mentions", "borda", "first place votes"]
 DEFAULT_LINE_KWDS = {
     "ls": "-",
@@ -53,11 +52,20 @@ def _set_default_bar_plot_args(
     bar_width: Optional[float],
     x_label_ordering: Optional[list[str]],
     legend_font_size: Optional[float],
-    threshold_values: Optional[list[float] | float],
-    threshold_kwds: Optional[list[dict[str, str]] | dict[str, str]],
+    threshold_values: Optional[Union[list[float], float]],
+    threshold_kwds: Optional[Union[list[dict], dict]],
     ax: Optional[Axes],
 ) -> Tuple[
-    list[dict[str, float]], list[str], dict, float, list[str], float, float, Axes
+    list[dict[str, float]],
+    list[str],
+    dict,
+    float,
+    list[str],
+    float,
+    float,
+    Optional[list[float]],
+    Optional[list[dict]],
+    Axes,
 ]:
     """
     Handles setting default arguments.
@@ -75,15 +83,16 @@ def _set_default_bar_plot_args(
             from first data dictionary.
         threshold_values (Union[list[float], float], optional): List of values to plot horizontal
             lines at. Can be provided as a list or a single float.
-        threshold_kwds (Union[list[dict[str, str]], dict[str, str]], optional): List of plotting
+        threshold_kwds (Union[list[dict], dict], optional): List of plotting
             keywords for the horizontal lines. Can be a list or single dictionary.
         ax (Axes, optional): A matplotlib axes object to plot the figure on. Defaults to None, in
             which case the function creates and returns a new axes.
 
     Returns:
-        Tuple[list[dict[str, float]], list[str], dict, float, list[str], float, float, Axes]: data,
+        Tuple[list[dict[str, float]], list[str], dict, float, list[str], float, float,
+        Optional[list[float]], Optional[list[dict]], Axes]: data,
         data_set_labels, data_set_to_color, bar_width, x_label_ordering, FONT_SIZE,
-        legend_font_size, ax
+        legend_font_size, threshold_values, threshold_kwds, ax
     """
     if not isinstance(data, list):
         data = [data]
@@ -102,25 +111,20 @@ def _set_default_bar_plot_args(
     if not x_label_ordering:
         x_label_ordering = list(data[0].keys())
 
-    FONT_SIZE = len(x_label_ordering) + 10
+    FONT_SIZE = len(x_label_ordering) + 10.0
 
     if legend_font_size is None:
         legend_font_size = FONT_SIZE
 
-    if threshold_values and not isinstance(threshold_values, list):
+    if threshold_values is not None and not isinstance(threshold_values, list):
         threshold_values = [threshold_values]
 
-    if threshold_kwds and not isinstance(threshold_kwds, list):
+    if threshold_kwds is not None and not isinstance(threshold_kwds, list):
         threshold_kwds = [threshold_kwds]
 
-    if threshold_values and not threshold_kwds:
-        threshold_kwds = [
-            DEFAULT_LINE_KWDS.update(
-                {"label": f"Line {i}"} for i in range(len(threshold_values))
-            )
-        ]
-
-    elif threshold_values and threshold_kwds:
+    if isinstance(threshold_values, list):
+        if not isinstance(threshold_kwds, list):
+            threshold_kwds = [dict() for _ in range(len(threshold_values))]
[…]
