### 2025-01-21 e514cff
edit bar_plot tests for thresholds
