### 2025-01-21 f26aa3e
add tests plot_cand_summary
