### 2025-01-27 98fc42d
add Peter edge cases
