### 2025-01-27 c157032
update tie scoring conventions
