+++ src/votekit/utils.py
@@ -1,5 +1,5 @@
 from fractions import Fraction
-from typing import Union, Sequence, Optional, TypeVar, cast
+from typing import Union, Sequence, Optional, TypeVar, cast, Literal
 from itertools import permutations
 import math
 import random
@@ -249,14 +249,13 @@ def score_profile_from_rankings(
     profile: PreferenceProfile,
     score_vector: Sequence[Union[float, Fraction]],
     to_float: bool = False,
+    tie_convention: Literal["high", "average", "low"] = "low",
 ) -> Union[dict[str, Fraction], dict[str, float]]:
     """
     Score the candidates based on a score vector. For example, the vector (1,0,...) would
     return the first place votes for each candidate. Vectors should be non-increasing and
     non-negative. Vector should be as long as the number of candidates. If it is shorter,
-    we add 0s. Candidates tied in a position receive an average of the points they would have
-    received had it been untied. Any candidates not listed on a ballot are considered tied in last
-    place (and thus receive an average of any remaining points).
+    we add 0s.
 
 
     Args:
@@ -266,6 +265,12 @@ def score_profile_from_rankings(
             If it is shorter, we add 0s.
         to_float (bool, optional): If True, compute scores as floats instead of Fractions.
             Defaults to False.
+        tie_convention (Literal["high", "average", "low"], optional): How to award points for
+            tied rankings. Defaults to "low", where any candidates tied receive the lowest possible
+            points for their position, eg three people tied for 3rd would each receive the points
+            for 5th. "high" awards the highest possible points, so in the previous example, they
+            would each receive the points for 3rd. "average" averages the points, so they would each
+            receive the points for 4th place.
 
     Returns:
         Union[dict[str, Fraction], dict[str, float]]:
@@ -276,8 +281,6 @@ def score_profile_from_rankings(
     max_length = len(profile.candidates)
     if len(score_vector) < max_length:
         score_vector = list(score_vector) + [0] * (max_length - len(score_vector))
-
-    profile = add_missing_cands(profile)
 
     scores = {c: Fraction(0) for c in profile.candidates}
     for ballot in profile.ballots:
@@ -292,7 +295,20 @@ def score_profile_from_rankings(
                 local_score_vector = score_vector[
                     current_ind : current_ind + position_size
                 ]
-                allocation = sum(local_score_vector) / position_size
+
+                if tie_convention == "high":
+                    allocation = max(local_score_vector)
+                elif tie_convention == "low":
+                    allocation = min(local_score_vector)
+                elif tie_convention == "average":
+                    allocation = sum(local_score_vector) / position_size
+                else:
+                    raise ValueError(
+                        (
+                            "tie_convention must be one of 'high', 'low', 'average', "
+                            f"not {tie_convention}"
+                        )
+                    )
                 for c in s:
                     scores[c] += Fraction(allocation) * ballot.weight
                 current_ind += position_size
@@ -303,7 +319,9 @@ def score_profile_from_rankings(
 
 
 def first_place_votes(
-    profile: PreferenceProfile, to_float: bool = False
+    profile: PreferenceProfile,
+    to_float: bool = False,
+    tie_convention: Literal["high", "average", "low"] = "average",
 ) -> Union[dict[str, Fraction], dict[str, float]]:
     """
     Computes first place votes for all candidates in a ``PreferenceProfile``.
@@ -312,6 +330,9 @@ def first_place_votes(
         profile (PreferenceProfile): The profile to compute first place votes for.
         to_float (bool): If True, compute first place votes as floats instead of Fractions.
             Defaults to False.
[…]
