+++ tests/test_utils.py
@@ -213,13 +213,13 @@ def test_validate_score_vector():
     validate_score_vector([3, 3, 3, 3])
 
 
-def test_score_profile_from_rankings():
+def test_score_profile_from_rankings_low():
     true_scores = {
-        "A": Fraction(105, 4),
-        "B": Fraction(73, 4),
-        "C": Fraction(77, 4),
-        "D": Fraction(45, 4),
-        "E": Fraction(30, 4),
+        "A": Fraction(25),
+        "B": Fraction(17),
+        "C": Fraction(17),
+        "D": Fraction(6),
+        "E": Fraction(1),
     }
 
     comp_scores = score_profile_from_rankings(profile_with_missing, [5, 4, 3, 2, 1])
@@ -227,8 +227,46 @@ def test_score_profile_from_rankings():
     assert isinstance(comp_scores["A"], Fraction)
 
 
+def test_score_profile_from_rankings_high():
+    true_scores = {
+        "A": Fraction(27.5),
+        "B": Fraction(19.5),
+        "C": Fraction(18.5),
+        "D": Fraction(7.5),
+        "E": Fraction(1),
+    }
+
+    comp_scores = score_profile_from_rankings(
+        profile_with_missing, [5, 4, 3, 2, 1], tie_convention="high"
+    )
+    assert comp_scores == true_scores
+    assert isinstance(comp_scores["A"], Fraction)
+
+
+def test_score_profile_from_rankings_avg():
+    true_scores = {
+        "A": Fraction(26.25),
+        "B": Fraction(18.25),
+        "C": Fraction(17.75),
+        "D": Fraction(6.75),
+        "E": Fraction(1),
+    }
+
+    comp_scores = score_profile_from_rankings(
+        profile_with_missing, [5, 4, 3, 2, 1], tie_convention="average"
+    )
+    assert comp_scores == true_scores
+    assert isinstance(comp_scores["A"], Fraction)
+
+
 def test_score_profile_from_rankings_to_float():
-    true_scores = {"A": 105 / 4, "B": 73 / 4, "C": 77 / 4, "D": 45 / 4, "E": 30 / 4}
+    true_scores = {
+        "A": 25,
+        "B": 17,
+        "C": 17,
+        "D": 6,
+        "E": 1,
+    }
 
     comp_scores = score_profile_from_rankings(
         profile_with_missing, [5, 4, 3, 2, 1], to_float=True
@@ -257,6 +295,13 @@ def test_score_profile_from_rankings_errors():
         score_profile_from_rankings(
             PreferenceProfile(ballots=(Ballot(ranking=({"A"}, frozenset(), {"B"})),)),
             [3, 2, 1],
+        )
+    with pytest.raises(
+        ValueError,
+        match=("tie_convention must be one of 'high', 'low', 'average', " "not highlo"),
+    ):
+        score_profile_from_rankings(
+            profile_no_ties, [5, 4, 3, 2, 1], tie_convention="highlo"
         )
 
 
@@ -301,7 +346,7 @@ def test_mentions_errors():
 
 
 def test_borda_no_ties():
-    true_borda = {"A": Fraction(15, 2), "B": Fraction(9), "C": Fraction(21, 2)}
+    true_borda = {"A": Fraction(15, 2), "B": Fraction(9), "C": Fraction(19, 2)}
 
     borda = borda_scores(profile_no_ties)
 
@@ -310,7 +355,7 @@ def test_borda_no_ties():
 
 
 def test_borda_no_ties_to_float():
-    true_borda = {"A": 15 / 2, "B": 9, "C": 21 / 2}
+    true_borda = {"A": 15 / 2, "B": 9, "C": 19 / 2}
 
[…]
