### 2025-01-27 1be9ee2
fix defaults of fpv
