+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "votekit"
-version = "3.0.0"
+version = "3.1.0"
 description = "A Swiss army knife for computational social choice research."
 authors = ["MGGG <code@mggg.org>"]
 readme = "README.md"
