+++ pyproject.toml
@@ -1,26 +1,28 @@
-[tool.poetry]
+[project]
 name = "votekit"
 version = "3.1.0"
 description = "A Swiss army knife for computational social choice research."
-authors = ["MGGG <code@mggg.org>"]
+authors = [{name = "MGGG", email =  "code@mggg.org"}]
 readme = "README.md"
+requires-python = ">=3.10, <3.13"
+dependencies = [
+    "scipy (>=1.11.1, <2.0)",
+    "mypy (>=1.4.1, <2.0)",
+    "pydantic (>=2.8.2, <3.0)",
+    "pandas-stubs (>=2.0.2.230605, <2.1)",
+    "pot (==0.9.4)",
+    "networkx (>=3.1, <4.0)",
+    "jinja2 (>=3.1.2, <4.0)",
+    "matplotlib (>=3.7.2, <4.0)",
+    "pandas (>=1.5.3, <2.0)",
+    "apportionment (>=1.0, <2.0)",
+    "scikit-learn (>=1.3.2, <2.0)",
+    "numpy (>=1.26.0, <2.0)",
+]
 
-[tool.poetry.dependencies]
-python = "^3.9,<3.13"
-scipy = "^1.11.1"
-mypy = "^1.4.1"
-pydantic = "^2.8.2"
-pandas-stubs = "^2.0.2.230605"
-pot = "0.9.4"
-networkx = "^3.1"
-jinja2 = "^3.1.2"
-matplotlib = "^3.7.2"
-pandas = "^1.5.3"
-apportionment = "^1.0"
-scikit-learn = "^1.3.2"
-numpy = "^1.26.0"
-
-
+#[tool.poetry.dependencies]
+#python = "
+ 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.0.275"
 black = "^24.8.0"
@@ -33,9 +35,9 @@ sphinx-copybutton = "^0.5.2"
 recommonmark = "^0.7.1"
 sphinx-rtd-theme = "^3.0.2"
 
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+#[build-system]
+#requires = ["poetry-core"]
+#build-backend = "poetry.core.masonry.api"
 
 [tool.ruff]
 line-length = 100
