### 2025-01-31 #7477 @britta-wstnr
Hi @cdonnay - the reviewers are happy and all checkboxes have been checked! 🙂
Thanks a lot @FATelarico, @tuvavbardal, @andrewheiss for your reviewing work, JOSS and I appreciate it! 🙏

The next step is now that I check the paper again, @cdonnay, I will get back to you once I have done that. Then we will move on to our post-review / pre-acceptance checklist.
