### 2025-02-03 #7477 @britta-wstnr
Hi @cdonnay,
I went over the paper again and have the following points regarding the format:
- [ ] the affiliations should be stated as Institute Name, Country. See here: https://joss.readthedocs.io/en/latest/example_paper.html
- [ ] If available, please add ORCIDs for the authors
- [ ] line 19: I think it would be better to add the URL for the toolbox directly into the parentheses. Right now it looks as if `VoteKit` has been previously published.
- [ ] `PrefVoting` has been recently published: https://joss.theoj.org/papers/10.21105/joss.07020
- [ ] throughout the paper, you sometimes cite papers by citing the first and last name of the authors. I believe that this comes down to the use of double curly braces and syntax in your bib-file, compare `{{Nicholas Mattei and Simon Rey}}` to `{Reynolds, Andrew and Reilly, Ben and Ellis, Andrew}`. Have a look at https://joss.readthedocs.io/en/latest/ to see how references should be entered.
- [ ] lines 65-66: do the models need references?
- [ ] line 88: From what I can see, the figure is _redrawn_ and the publication organ is not a Publishing House, which might require permission for reprinting figures. However, you could consider rewording this sentences to make it clearer that you are not reprinting under copyright law (or adding a license note to the figure caption).
- [ ] lines 92 + 98: The citations Elkind as well as Szufa should be in-text
- [ ] lines 120 + 122: Previously, CS referred to Cambridge Sampler - is this still the case here or do you mean computer science (or something else)? In any case, I think it would be good to (re-)introduce the abbreviation
- [ ] reference Benadè et al 2021 (and others): if a reference has a DOI, this is preferred over other websites to ensure long-lasting availability, please double-check
- [ ] reference Duchin et al. 2024 is stated as "preprint" but has no DOI or URL
- [ ] please double-check the paper example for the citation of software packages on GitHub: https://joss.readthedocs.io/en/latest/example_paper.html

I'll have a closer look at the references (URLs, DOIs) once you adapted them.
Thanks! 🙏
