### 2025-02-04 a9d0006
update citation format and affils
