+++ paper/paper.bib
@@ -17,20 +17,24 @@
 }
 
 @unpublished{preflibtools, 
-	author = {{Nicholas Mattei and Simon Rey}},
+	author = {Nicholas Mattei and Simon Rey},
 	title = {PrefLib-Tools}, 
 	note = {Python Library. \url{https://github.com/PrefLib/preflibtools}},
   year      = {2022},
   url = {https://github.com/PrefLib/preflibtools}
 }
 
-@unpublished{prefvoting, 
-	author = {{Eric Pacuit and Wesley H. Holliday}},
-	title = {pref_voting}, 
-	note = {Python Library. \url{https://github.com/voting-tools/pref_voting}},
-  year      = {2022},
-  url = {https://github.com/voting-tools/pref_voting}
-}
+@article{prefvoting, 
+doi = {10.21105/joss.07020}, 
+url = {https://doi.org/10.21105/joss.07020}, 
+year = {2025}, 
+publisher = {The Open Journal}, 
+volume = {10}, 
+number = {105}, 
+pages = {7020}, 
+author = {Wesley H. Holliday and Eric Pacuit}, 
+title = {pref_voting: The Preferential Voting Tools package for Python}, 
+journal = {Journal of Open Source Software} }
 
 @unpublished{votelib, 
 	author = {{Jan \v{S}imbera}},
@@ -74,7 +78,7 @@
 
 
 @unpublished{ComSoc, 
-	author = {{Ulle Endriss and Simon Rey}},
+	author = {Ulle Endriss and Simon Rey},
 	title = {COMSOC Community Site}, 
   year = {2024},
 	note = {GitHub Repository and Community Resources. 
@@ -86,7 +90,7 @@
 
 @inproceedings{boehmer2024guidenumericalexperimentselections,
   title     = {Guide to Numerical Experiments on Elections in Computational Social Choice},
-  author    = {Boehmer, Niclas and Faliszewski, Piotr and Janeczko, Łukasz and Kaczmarczyk, Andrzej and Lisowski, Grzegorz and Pierczyński, Grzegorz and Rey, Simon and Stolicki, Dariusz and Szufa, Stanisław and Wąs, Tomasz},
+  author    = {Niclas Boehmer and Piotr Faliszewski and Łukasz Janeczko and Andrzej Kaczmarczyk and Grzegorz Lisowski and Grzegorz Pierczyński and Simon Rey and Dariusz Stolicki and Stanisław Szufa and Tomasz Wąs},
   booktitle = {Proceedings of the Thirty-Third International Joint Conference on
                Artificial Intelligence, {IJCAI-24}},
   publisher = {International Joint Conferences on Artificial Intelligence Organization},
@@ -109,7 +113,7 @@
 
 
 @inproceedings{drawing-a-map,
- author = {Szufa, Stanis\l{}aw and Faliszewski, Piotr and Skowron, Piotr and Slinko, Arkadii and Talmon, Nimrod},
+ author = {Stanis\l{}aw Szufa and Piotr Faliszewski and Piotr Skowron and Arkadii Slinko and Nimrod Talmon},
  title = {Drawing a Map of Elections in the Space of Statistical Cultures},
  year = {2020},
  isbn = {9781450375184},
@@ -127,7 +131,7 @@
 
 @inproceedings{distance-elex,
   title     = {Understanding Distance Measures Among Elections},
-  author    = {Boehmer, Niclas and Faliszewski, Piotr and Niedermeier, Rolf and Szufa, Stanisław and Wąs, Tomasz},
+  author    = {Niclas Boehmer and Piotr Faliszewski and Rolf Niedermeier and Stanisław Szufa and Tomasz Wąs},
   booktitle = {Proceedings of the Thirty-First International Joint Conference on
[…]
