### 2025-02-04 #7477 @cdonnay
Hi @britta-wstnr, working my way through these! Can you clarify what you mean by "he citations Elkind as well as Szufa should be in-text" ? Thanks!
