+++ paper/paper.md
@@ -132,7 +132,7 @@ behavior and particularly to generate polarized elections. Two leading
 choices are based on classic statistical ranking mechanisms, called the
 Plackett--Luce (PL) and Bradley--Terry (BT) models; another model called
 the Cambridge Sampler (CS) draws from historical ranking data in
-Cambridge, MA city council elections. These models have flexible
+Cambridge, MA city council elections [@benade_donnay_duchin_weighill_24]. These models have flexible
 parameters---allowing users to vary voting bloc proportions, candidate
 strength within slates, and polarization between blocs---that can be 
 specified or randomly sampled.
@@ -161,9 +161,9 @@ references.  In addition, `VoteKit` is flexible enough to allow users to write c
 
 Reform advocates also need to describe voting mechanisms and their
 likely outcomes effectively to members of their communities. The end-to-end pipeline provided by `VoteKit` allows advocates to toggle different system settings and compare expected outcomes. For example, 
-Figure 1 is reprinted from a report on reform proposals for the chambers of the Washington state legislature.  Using the codebase that formed the foundation of `VoteKit`, researchers compared the expected outcomes for minority representation under six possible electoral systems.
+Figure 1 is reprinted with permission from a report on reform proposals for the chambers of the Washington state legislature.  Using the codebase that formed the foundation of `VoteKit`, researchers compared the expected outcomes for minority representation under six possible electoral systems.
 
-![A comparison of a variety of electoral systems and their effect on minority representation in a [case study](https://mggg.org/washington) of reform proposals for the Washington state legislature [@washington_leg]. Even within ranked-choice proposals, certain options, like System 0 (based on single-member districts), are projected to be less successful for minority representation, while other systems, like System 1 (based on multi-member districts), predict that candidates of choice for people of color ("POC") are elected more in line with the POC share of population or citizen voting age population ("CVAP").](./figures/WA_poc_seats_chartsystem_compare_pared.png){width=100%}
+![A comparison of a variety of electoral systems and their effect on minority representation, reprinted with permission, from a [case study](https://mggg.org/washington) of reform proposals for the Washington state legislature [@washington_leg]. Even within ranked-choice proposals, certain options, like System 0 (based on single-member districts), are projected to be less successful for minority representation, while other systems, like System 1 (based on multi-member districts), predict that candidates of choice for people of color ("POC") are elected more in line with the POC share of population or citizen voting age population ("CVAP").](./figures/WA_poc_seats_chartsystem_compare_pared.png){width=100%}
 
 
 ## Area of need: Resources for research
@@ -235,11 +235,11 @@ the following.
 -   A peer-reviewed article for an election law audience on the impact
     of STV elections on minority representation [@Benade2021];
 
--   A peer-reviewed article for a CS/econ audience that probes whether
+-   A peer-reviewed article for a computer science and econ audience that probes whether
     STV delivers proportional representation
     [@benade_donnay_duchin_weighill_24]; and
 
--   A peer-reviewed article for an CS/operations research audience on
+-   A peer-reviewed article for a computer science and operations research audience on
     optimizing to "learn\" blocs and slates in real-world elections
     [@duchin_tapp_24].
 
