+++ paper/paper.bib
@@ -1,11 +1,12 @@
 
 
-@unpublished{VoteKit, 
+@misc{VoteKit, 
 	author = {{MGGG Redistricting Lab}},
 	title = {VoteKit}, 
-	note = {Python Library. \url{https://github.com/mggg/VoteKit}},
   year      = {2023},
   month     = {8},
+  publisher = {Github},
+  journal = {Github repository},
   url = {https://github.com/mggg/VoteKit}
 }
 
@@ -16,10 +17,11 @@
   url = {https://votekit.readthedocs.io/en/latest/}
 }
 
-@unpublished{preflibtools, 
+@misc{preflibtools, 
 	author = {Nicholas Mattei and Simon Rey},
 	title = {PrefLib-Tools}, 
-	note = {Python Library. \url{https://github.com/PrefLib/preflibtools}},
+	publisher = {Github},
+  journal = {Github repository},
   year      = {2022},
   url = {https://github.com/PrefLib/preflibtools}
 }
@@ -36,18 +38,20 @@ author = {Wesley H. Holliday and Eric Pacuit},
 title = {pref_voting: The Preferential Voting Tools package for Python}, 
 journal = {Journal of Open Source Software} }
 
-@unpublished{votelib, 
+@misc{votelib, 
 	author = {{Jan \v{S}imbera}},
 	title = {Votelib}, 
-	note = {Python Library. \url{https://github.com/simberaj/votelib}},
+	publisher = {Github},
+  journal = {Github repository},
   url = {https://github.com/simberaj/votelib},
   year      = {2020},
 }
 
-@unpublished{apportionment, 
+@misc{apportionment, 
 	author = {{Martin Lackner}},
 	title = {apportionment}, 
-	note = {Python Library. \url{https://github.com/martinlackner/apportionmentb}},
+  publisher = {Github},
+  journal = {Github repository},
   year      = {2022},
   url = {https://github.com/martinlackner/apportionmentb}
 }
@@ -68,11 +72,12 @@ journal = {Journal of Open Source Software} }
 
 
 
-@unpublished{Scot-Elex, 
+@misc{Scot-Elex, 
 	author = {{MGGG Redistricting Lab}},
 	title = {Scottish STV Election Repo}, 
   year = {2023},
-	note = {Data Repository. \url{https://github.com/mggg/scot-elex}},
+	publisher = {Github},
+  journal = {Github repository},
   url = {https://github.com/mggg/scot-elex}
 }
 
@@ -103,11 +108,12 @@ journal = {Journal of Open Source Software} }
   url       = {https://doi.org/10.24963/ijcai.2024/881},
 }
 
-@unpublished{RCV-Cruncher, 
+@misc{RCV-Cruncher, 
 	author = {{FairVote}},
 	title = {RCV Cruncher}, 
   year      = {2024},
-	note = {Python Library. \url{https://github.com/fairvotereform/rcv_cruncher/}},
+	publisher = {Github},
+  journal = {Github repository},
   url = {https://github.com/fairvotereform/rcv_cruncher/}
 }
 
