### 2025-02-05 1bad43f
standardize use of max_ballot_length in scoring
