### 2025-02-05 8d8fc7d
edit comments to reflect max_ballot_length
