+++ paper/paper.bib
@@ -27,16 +27,17 @@
 }
 
 @article{prefvoting, 
-doi = {10.21105/joss.07020}, 
-url = {https://doi.org/10.21105/joss.07020}, 
-year = {2025}, 
-publisher = {The Open Journal}, 
-volume = {10}, 
-number = {105}, 
-pages = {7020}, 
-author = {Wesley H. Holliday and Eric Pacuit}, 
-title = {pref_voting: The Preferential Voting Tools package for Python}, 
-journal = {Journal of Open Source Software} }
+  doi = {10.21105/joss.07020}, 
+  url = {https://doi.org/10.21105/joss.07020}, 
+  year = {2025}, 
+  publisher = {The Open Journal}, 
+  volume = {10}, 
+  number = {105}, 
+  pages = {7020}, 
+  author = {Wesley H. Holliday and Eric Pacuit}, 
+  title = {pref_voting: The Preferential Voting Tools package for Python}, 
+  journal = {Journal of Open Source Software},
+ }
 
 @misc{votelib, 
 	author = {{Jan \v{S}imbera}},
