+++ src/votekit/matrices/__init__.py
@@ -1,0 +1,1 @@
+from .mentions import mention, comention, comention_above, comentions_matrix  # noqa
