+++ tests/matrices/test_convert_dict_to_mat.py
@@ -1,4 +1,4 @@
-from votekit.matrices._convert_dict_to_mat import _convert_dict_to_matrix
+from votekit.matrices._utils import _convert_dict_to_matrix
 import numpy as np
 import pytest
 
@@ -17,12 +17,23 @@ def test_convert():
     assert mat[2][1] == -1
 
 
-def test_convert_error():
+def test_convert_float_error():
     data = {
         "Chris": {"Chris": 2, "Peter": "hi", "Moon": 4},
         "Peter": {"Chris": 1, "Peter": 5, "Moon": 2},
         "Moon": {"Chris": 0, "Peter": -1, "Moon": 0.25},
     }
 
-    with pytest.raises(ValueError, match="could not convert (.*?) to float"):
+    with pytest.raises(ValueError, match="Unable to parse (.*?) at position"):
+        _convert_dict_to_matrix(data)
+
+
+def test_convert_size_error():
+    data = {
+        "Chris": {"Chris": 2, "Peter": 1, "Moon": 4},
+        "Peter": {"Chris": 1, "Peter": 5, "Moon": 2, "David": 5},
+        "Moon": {"Chris": 0, "Peter": -1, "Moon": 0.25},
+    }
+
+    with pytest.raises(AssertionError, match="Inner keys do not match across all rows"):
         _convert_dict_to_matrix(data)
