+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
     hooks:
       - id: ruff
   - repo: https://github.com/ambv/black
-    rev: 22.10.0
+    rev: 24.8.0
     hooks:
       - id: black
   # - repo: https://github.com/pre-commit/mirrors-mypy
