### 2025-02-11 ad79b2c
remove mentions function
