### 2025-02-12 5eac6d8
check for 0 denom
