### 2025-02-12 03ff5fa
reorganize matrix code
