### 2025-02-12 114d847
add seaborn to requirements
