+++ pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "apportionment (>=1.0, <2.0)",
     "scikit-learn (>=1.3.2, <2.0)",
     "numpy (>=1.26.0, <2.0)",
+    "seaborn (>=0.13.2,<0.14.0)",
 ]
 
 #[tool.poetry.dependencies]
