+++ src/votekit/matrices/heatmap.py
@@ -1,0 +1,344 @@
+import numpy as np
+from matplotlib.axes import Axes
+from typing import Optional, Dict, List, Tuple, Union
+from matplotlib.colors import Colormap, TwoSlopeNorm
+import matplotlib.pyplot as plt
+import seaborn as sns
+import matplotlib.patches as mpatches
+
+
+def _validate_heatmap_inputs(
+    *,
+    matrix: np.ndarray,
+    row_labels: Optional[List[str]] = None,
+    column_labels: Optional[List[str]] = None,
+    row_legend: Optional[Dict[str, str]] = None,
+    column_legend: Optional[Dict[str, str]] = None,
+) -> None:
+    """
+    Used internally for the `matrix_heatmap` function to validate the inputs to the function.
+
+    Args:
+        matrix (np.ndarray): A 2D numpy array containing the data to be plotted.
+        row_labels (Optional[List[str]]): A list of strings containing the labels for the rows
+            of the heatmap. Defaults to None.
+        column_labels (Optional[List[str]]): A list of strings containing the labels for the
+            columns of the heatmap. Defaults to None.
+        row_legend (Optional[Dict[str, str]]): A dictionary mapping row labels to legend
+            descriptions. Defaults to None.
+        column_legend (Optional[Dict[str, str]]): A dictionary mapping column labels to legend
+            descriptions. Defaults to None.
+
+    Raises:
+        ValueError: If the matrix is not a 2D numpy array.
+        ValueError: If the row labels are not of the correct length.
+        ValueError: If the column labels are not of the correct length.
+        ValueError: If the keys of the row legend are do not match the row labels.
+        ValueError: If the keys of the column legend are do not match the column labels.
+        ValueError: If the legend descriptions in the row legend and column legend are in conflict.
+    """
+    if matrix.ndim != 2:
+        raise ValueError(
+            f"Please provide a 2D matrix to plot. Found a {matrix.ndim}-D matrix."
+        )
+
+    n_rows = matrix.shape[0]
+    n_cols = matrix.shape[1]
+
+    if row_labels is not None:
+        if n_rows != len(row_labels):
+            raise ValueError(
+                f"Please provide {n_rows} labels for the rows of the "
+                f"matrix. Found {len(row_labels)} labels."
+            )
+
+    if column_labels is not None:
+        if n_cols != len(column_labels):
+            raise ValueError(
+                f"Please provide {n_cols} labels for the columns of the "
+                f"matrix. Found {len(column_labels)} labels."
+            )
+
+    if row_legend is not None and column_legend is not None:
+        for label, description in column_legend.items():
+            desc = row_legend.get(label, description)
+            if desc != description:
+                raise ValueError(
+                    f"Conflicting legend descriptions for '{label}': "
+                    f"got '{description}' and '{desc}'."
+                )
+
+    if row_legend is not None and list(row_legend.keys()) != row_labels:
+        raise ValueError("Row labels do not match row legend keys.")
+
+    if column_legend is not None and list(column_legend.keys()) != column_labels:
+        raise ValueError("Column labels do not match column legend keys.")
+
+
+def _add_text_to_heatmap(
+    *,
+    heatmap: Axes,
+    n_decimals_to_display: int,
+    matrix: np.ndarray,
+    cell_font_size: Optional[int] = None,
+) -> Axes:
+    """
+    Adds the text values to the heatmap cells. This function dynamically determines
+    the font size based on the number of cells in the figure, the figure size, and
+    the length of the numbers to ensure readability.
+
+    Args:
[…]
