+++ pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
     "scikit-learn (>=1.3.2, <2.0)",
     "numpy (>=1.26.0, <2.0)",
     "seaborn (>=0.13.2,<0.14.0)",
+    "types-seaborn (>=0.13.2.20250111,<0.14.0.0)",
 ]
 
 #[tool.poetry.dependencies]
