### 2025-02-13 1a43712
add bar plot
