+++ src/votekit/plots/__init__.py
@@ -1,2 +1,2 @@
 from .mds import plot_MDS, compute_MDS  # noqa
-from .profile_plots import plot_candidate_summary_stats, bar_plot  # noqa
+from .bar_plot import bar_plot  # noqa
