+++ src/votekit/plots/bar_plot.py
@@ -1,0 +1,628 @@
+from matplotlib import pyplot as plt  # type: ignore
+from votekit.utils import COLOR_LIST
+from matplotlib.axes import Axes
+from typing import Optional, Union, Tuple, Literal, Any
+import matplotlib.patches as mpatches
+import warnings
+from matplotlib.legend import Legend
+from matplotlib.lines import Line2D
+
+VALID_STATS = ["fpv", "mentions", "borda", "first place votes"]
+Valid_stat = Literal["fpv", "mentions", "borda", "first place votes"]
+DEFAULT_LINE_KWDS = {
+    "linestyle": "-",
+    "linewidth": 2,
+}
+
+
+def add_null_keys(data: dict[str, dict[str, float]]) -> dict[str, dict[str, float]]:
+    """
+    Prepares dictionary of dictionaries to be passed to ``bar_plot()``. If a key is missing from
+    a dictionary, this function adds the key with value 0.
+
+    Args:
+        data (dict[str, dict[str, float]]): Categorical data to be cleaned. The value of each dict
+            should be the frequency of the key which is the category name.
+
+    Returns:
+        dict[str, dict[str, float]]: Cleaned data.
+    """
+
+    x_labels = []
+    for data_dict in data.values():
+        for x_label in data_dict.keys():
+            if x_label not in x_labels:
+                x_labels.append(x_label)
+
+    for x_label in x_labels:
+        for data_dict in data.values():
+            if x_label not in data_dict:
+                data_dict[x_label] = 0
+
+    return data
+
+
+def _set_default_bar_plot_args(
+    *,
+    data: dict[str, dict[str, float]],
+    data_set_colors: Optional[dict],
+    bar_width: Optional[float],
+    category_ordering: Optional[list[str]],
+    legend_font_size: Optional[float],
+    threshold_values: Optional[Union[list[float], float]],
+    threshold_kwds: Optional[Union[list[dict], dict]],
+    ax: Optional[Axes],
+) -> dict[str, Any]:
+    """
+    Handles setting default arguments.
+
+    Args:
+        data (dict[str, dict[str, float]]): Categorical data to be plotted. Top level keys are
+            data set labels. Inner keys are categories, and inner values are the height of the bars.
+        data_set_colors (dict, optional): Dictionary mapping data set labels or data set indices to
+            colors. Defaults to None, in which case we use a subset of ``COLOR_LIST`` from ``utils``
+            module. Dictionary keys can be a subset of the data sets.
+        bar_width (float, optional): Width of bars, defaults to None which computes the bar width
+            as 0.7 divided by the number of data sets.
+        category_ordering (list[str], optional): Ordering of x labels. Defaults to order retrieved
+            from first data dictionary.
+        threshold_values (Union[list[float], float], optional): List of values to plot horizontal
+            lines at. Can be provided as a list of values or a single float.
+        threshold_kwds (Union[list[dict], dict], optional): List of plotting
+            keywords for the horizontal lines. Can be a list of dictionaries or single dictionary.
+        ax (Axes, optional): A matplotlib axes object to plot the figure on. Defaults to None, in
+            which case the function creates and returns a new axes.
+
+    Returns:
+        dict[str, Any]: data_set_to_color, bar_width, category_ordering, font_size,
+        legend_font_size, threshold_values, threshold_kwds, ax
+    """
+
+    data_set_to_color = {d: COLOR_LIST[i] for i, d in enumerate(data.keys())}
+
+    if data_set_colors:
+        for data_set, color in data_set_colors.items():
+            data_set_to_color[data_set] = color
+
[…]
