### 2025-02-14 #7477 @cdonnay
Hi @britta-wstnr, we went ahead and addressed all of your comments except two.

1. The references that have no DOI will unfortunately have to stay that way. Most of them are white papers from our lab, and it would take well over a month to work with our university to get them DOIs.
2. We are working on getting a URL/DOI for the Duchin et al. 2024 is stated as "preprint", but that might take a week or so from now. Happy to ping again we have that!
