### 2025-02-17 #7477 @britta-wstnr
Hi @cdonnay - sure, no problem! It was really meant as "if they have a DOI" if they don't then the URL is fine. But good to hear you will add either of the two for Duchin et al.!
