### 2025-02-17 #7477 @britta-wstnr
Whoops, today, editorialbot commands don't seem to be my strong suit 🙈
