### 2025-02-17 #7477 @britta-wstnr
Thanks for adapting the paper - looks almost good, @cdonnay!
Two small things:
1. Note that also double curly braces on single authors in the `.bib` file lead to unwanted results. These are the examples I saw: `author = {{Martin Lackner}}`, `{{Jan \v{S}imbera}}` (should be instead `{Jan {\v{S}imbera}}` to treat the character correctly, I believe), and `{{P. Emerson}}`.
Added like this will make the first and last name unseperable, leading to citations such as "P. Emerson, 2013".
2. I think something funky is going on with line 82, which I might have overlooked before, sorry: "See generally (Amorós et al., 2016; McCune et al., 2023; Nicolaus Tideman, 1995; P. Emerson, 2013; Reynolds et al., 2008) for references."
I am not sure what was meant - maybe "See Amorós et al., 2016; McCune et al., 2023; Nicolaus Tideman, 1995; P. Emerson, 2013; Reynolds et al., 2008 for references." -- but then it might be more elegant to do "... Alaska or the sequential RCV in Utah local elections (Amorós et al., 2016; McCune et al., 2023; Nicolaus Tideman, 1995; P. Emerson, 2013; Reynolds et al., 2008)"
