+++ paper/paper.bib
@@ -40,7 +40,7 @@
  }
 
 @misc{votelib, 
-	author = {{Jan \v{S}imbera}},
+	author = {Jan {\v{S}imbera}},
 	title = {Votelib}, 
 	publisher = {Github},
   journal = {Github repository},
@@ -49,7 +49,7 @@
 }
 
 @misc{apportionment, 
-	author = {{Martin Lackner}},
+	author = {Martin Lackner},
 	title = {apportionment}, 
   publisher = {Github},
   journal = {Github repository},
@@ -257,7 +257,7 @@ WA School District},
 
 @article{
     STV,
-    Author = {{Nicolaus Tideman}},
+    Author = {Nicolaus Tideman},
     Title = {The Single Transferable Vote},
     Journal = {Journal of Economic Perspectives},
     Volume = {9},
@@ -272,7 +272,7 @@ WA School District},
 @article{Burden,
  ISSN = {00925853, 15405907},
  URL = {http://www.jstor.org/stable/2960485},
- author = {{Barry C. Burden}},
+ author = {Barry C. Burden},
  journal = {American Journal of Political Science},
  number = {4},
  pages = {1150--1169},
@@ -300,7 +300,7 @@ WA School District},
 
 @article{
     Borda,
-    Author = {{P. Emerson}},
+    Author = {P. Emerson},
     Title = {The original Borda count and partial voting},
     Journal = {Social Choice and Welfare},
     Volume = {40},
@@ -327,7 +327,7 @@ WA School District},
  url={http://dx.doi.org/10.1007/978-94-009-4225-7}, 
  DOI={10.1007/978-94-009-4225-7}, 
  publisher={Springer Netherlands}, 
- author={{Duncan Black}}, 
+ author={Duncan Black}, 
  year={1986} 
  }
 
