### 2025-02-17 #7477 @cdonnay
Hi @britta-wstnr!

For 2) all we are trying to say is that those are the references for the various election methods listed above.  So adding them to the end of the Utah local elections sentence isn't quite the right place for it. I tried just moving the original sentence up so that it is clearer what it references. Let me know what you think!

Also made all the changes for 1).

Everything should be updated in the JOSS branch!
