+++ src/votekit/plots/bar_plot.py
@@ -7,12 +7,7 @@ import warnings
 from matplotlib.legend import Legend
 from matplotlib.lines import Line2D
 
-VALID_STATS = ["fpv", "mentions", "borda", "first place votes"]
-Valid_stat = Literal["fpv", "mentions", "borda", "first place votes"]
-DEFAULT_LINE_KWDS = {
-    "linestyle": "-",
-    "linewidth": 2,
-}
+DEFAULT_LINE_KWDS = {"linestyle": "-", "linewidth": 2, "color": "black"}
 
 
 def add_null_keys(data: dict[str, dict[str, float]]) -> dict[str, dict[str, float]]:
@@ -107,10 +102,6 @@ def _set_default_bar_plot_args(
         if not isinstance(threshold_kwds, list):
             threshold_kwds = [dict() for _ in range(len(threshold_values))]
         for i, kwd_dict in enumerate(threshold_kwds):
-            if "color" not in kwd_dict:
-                kwd_dict["color"] = COLOR_LIST[-(i + 1)]
-            if "label" not in kwd_dict:
-                kwd_dict["label"] = f"Line {i+1}"
             for k, v in DEFAULT_LINE_KWDS.items():
                 if k not in kwd_dict:
                     kwd_dict[k] = v
@@ -154,7 +145,10 @@ def _validate_bar_plot_args(
     Raises:
         ValueError: All bar heights must be positive.
         ValueError: All sub dictionaries in data must have the same keys.
-        ValueError: category_ordering must match the keys of data (unordered).
+        ValueError: Cannot plot more than len(COLOR_LIST) data sets.
+        ValueError: category_ordering must be the same length as the keys of sub-dictionaries.
+        ValueError: category_ordering must have no extraneous labels.
+        ValueError: category_ordering must have no missing labels.
         ValueError: Bar width must be positive.
         UserWarning: Bar width must be less than 1.
         ValueError: threshold_values and kwds must have the same length.
@@ -172,13 +166,29 @@ def _validate_bar_plot_args(
     ):
         raise ValueError("All data dictionaries must have the same categorical keys.")
 
-    sym_dif = set(category_ordering) ^ set(next(iter(data.values())).keys())
-    if len(sym_dif) != 0:
+    if len(data) > len(COLOR_LIST):
+        raise ValueError(f"Cannot plot more than {len(COLOR_LIST)} data sets.")
+
+    true_categories = next(iter(data.values())).keys()
+
+    if len(category_ordering) != len(true_categories):
+        raise ValueError(
+            "category_ordering must be the same length as sub-dictionaries."
+        )
+
+    if set(category_ordering).difference(true_categories) != set():
         raise ValueError(
             (
-                "category_ordering must match the keys of the data sub-dictionaries. "
-                "Here is the symmetric "
-                f"difference of the two sets: {sym_dif}"
+                "category_ordering has extraneous labels: "
+                f"{set(category_ordering).difference(true_categories)}"
+            )
+        )
+
+    if set(true_categories).difference(category_ordering) != set():
+        raise ValueError(
+            (
+                "category_ordering has missing labels: "
+                f"{set(true_categories).difference(category_ordering)}"
             )
         )
 
@@ -287,7 +297,7 @@ def _plot_datasets_on_bar_plot(
     """
 
     for i, data_set in enumerate(y_data):
-        bar_shift = bar_width * i - len(y_data) / 2 * bar_width
+        bar_shift = bar_width * (i - len(y_data) / 2)
         bar_centers = [x + bar_shift for x in range(len(category_ordering))]
 
         align: Literal["center", "edge"] = "edge" if len(y_data) % 2 == 0 else "center"
[…]
