+++ tests/plots/test_bar_plot.py
@@ -1,227 +1,0 @@
-from votekit.plots import bar_plot
-from votekit.plots.profile_plots import (
-    add_null_keys,
-    _normalize_data,
-    _set_default_bar_plot_args,
-    _prepare_data_bar_plot,
-)
-import pytest
-from matplotlib.axes import Axes
-from votekit.utils import COLOR_LIST
-import matplotlib.pyplot as plt
-
-test_data_1 = {"Moon": 4.0, "Peter": 0.0, "Chris": 2.0}
-test_data_2 = {"Moon": 4.0, "Peter": 1.0, "Chris": 1.0}
-test_data_3 = {"Mala": 4.0, "David": 1.0, "Jeanne": 1.0}
-
-
-def test_add_null_keys():
-    data = add_null_keys([test_data_1, {"Peter": 1.0, "Chris": 1.0, "Moon": 4.0}])
-    assert data[0].keys() == data[1].keys()
-    assert list(data[0].keys()) == ["Moon", "Peter", "Chris"]
-
-    data = add_null_keys([test_data_1, test_data_3])
-    assert data[0].keys() == data[1].keys()
-    assert list(data[0].keys()) == ["Moon", "Peter", "Chris", "Mala", "David", "Jeanne"]
-    assert data[0]["Jeanne"] == 0 and data[1]["Moon"] == 0
-
-
-def test_bar_plot_simple():
-    assert isinstance(bar_plot(test_data_1), Axes)
-
-
-def test_bar_plot_multi_sets():
-    assert isinstance(
-        bar_plot([test_data_1, test_data_2], data_set_labels=["1", "2"]), Axes
-    )
-
-    data = add_null_keys([test_data_1, test_data_3])
-    assert isinstance(bar_plot(data, data_set_labels=["1", "3"]), Axes)
-
-
-def test_bar_plot_threshold_lines():
-    assert isinstance(bar_plot(test_data_1, threshold_values=[2, 3]), Axes)
-
-
-def test_bar_plot_validation_errors():
-    with pytest.raises(
-        ValueError, match="All data dictionaries must have the same keys."
-    ):
-        bar_plot([test_data_1, test_data_3])
-
-    with pytest.raises(ValueError, match="Each data set must have a unique label."):
-        bar_plot([test_data_1, test_data_2], data_set_labels=["1", "1"])
-
-    with pytest.raises(
-        ValueError, match="There must be one data set label for each data set."
-    ):
-        bar_plot([test_data_1, test_data_2], data_set_labels=["1"])
-
-    with pytest.raises(
-        ValueError, match="There must be one data set label for each data set."
-    ):
-        bar_plot(test_data_1, data_set_labels=["1", "2"])
-
-    with pytest.raises(
-        ValueError,
-        match=(
-            "x_label_ordering must match the keys of the data."
-            " Here is the symmetric difference of the two sets:"
-        ),
-    ):
-        bar_plot(test_data_1, x_label_ordering=["Chris"])
-
-    with pytest.raises(
-        ValueError,
-        match=(
-            "x_label_ordering must match the keys of the data."
-            " Here is the symmetric difference of the two sets:"
-        ),
-    ):
-        bar_plot(test_data_1, x_label_ordering=["Chris", "Moon", "Peter", "Mala"])
-
-    with pytest.raises(
-        ValueError,
-        match=(
-            "x_label_ordering must match the keys of the data."
-            " Here is the symmetric difference of the two sets:"
-        ),
-    ):
-        bar_plot(
-            add_null_keys([test_data_1, test_data_3]),
-            x_label_ordering=["Chris", "Moon", "Peter", "Mala"],
[…]
