### 2025-02-19 #7477 @britta-wstnr
Hi @cdonnay, thanks!
I am _really_ nitpicking now, sorry, but for me this is was/is not clear. What about rephrasing it the way you explained it in the comment to me? That I found very clear 🙂
Taking a stab at it, maybe something like: "For references for the listed election methods, see for example: Amorós et al., 2016; Emerson, 2013; McCune et al., 2023; Reynolds et al., 2008; Tideman, 1995"?

Let me know once the DOI is in. In the meanwhile, you can already look at the **Author Tasks** of [this](https://github.com/openjournals/joss-reviews/issues/7477#issuecomment-2668200870) post-review checklist as well. Thanks a lot! 🙏
