### 2025-02-20 54325a9
bar_plot wrapper
