+++ src/votekit/plots/bar_plot.py
@@ -656,6 +656,80 @@ def multi_bar_plot(
     return ax  # type: ignore[return-value]
 
 
-# TODO
-def bar_plot():
-    pass
+def bar_plot(
+    data: dict[str, float],
+    *,
+    data_set_label: str = "Data set",
+    normalize: bool = False,
+    data_set_color: str = COLOR_LIST[0],
+    bar_width: Optional[float] = None,
+    category_ordering: Optional[list[str]] = None,
+    x_axis_name: Optional[str] = None,
+    y_axis_name: Optional[str] = None,
+    title: Optional[str] = None,
+    show_data_set_legend: bool = False,
+    categories_legend: Optional[dict[str, str]] = None,
+    threshold_values: Optional[Union[list[float], float]] = None,
+    threshold_kwds: Optional[Union[list[dict], dict]] = None,
+    legend_font_size: Optional[float] = None,
+    ax: Optional[Axes] = None,
+) -> Axes:
+    """
+    Plots bar plot of a single categorical data set. Wrapper for ``multi_bar_plot``.
+
+    Args:
+        data (dict[str, float]): Categorical data set to be plotted. Keys are categories, and
+            values are the height of the bars.
+        data_set_label(str, optional): Label for data set. Defaults to "Data set".
+        normalize (bool, optional): Whether or not to normalize data. Defaults to False.
+        data_set_color (str, optional): Color of data set.
+            Defaults to the first color from ``COLOR_LIST`` from ``utils`` module.
+        bar_width (float, optional): Width of bars. Defaults to None which computes the bar width
+            as 0.7 divided by the number of data sets. Must be in the interval :math:`(0,1]`.
+        category_ordering (list[str], optional): Ordering of x-labels. Defaults to order retrieved
+            from data dictionary.
+        x_axis_name (str, optional): Name of x-axis. Defaults to None, which does not plot a name.
+        y_axis_name (str, optional): Name of y-axis. Defaults to None, which does not plot a name.
+        title (str, optional): Title for the figure. Defaults to None, which does not plot a title.
+        show_data_set_legend (bool, optional): Whether or not to plot the data set legend.
+            Defaults to False. Is automatically shown if any threshold lines have the keyword
+            "label" passed through ``threshold_kwds``.
+        categories_legend (dict[str, str], optional): Dictionary mapping data categories
+            to description. Defaults to None. If provided, generates a second legend for data
+            categories.
+        threshold_values (Union[list[float], float], optional): List of values to plot horizontal
+            lines at. Can be provided as a list or a single float.
+        threshold_kwds (Union[list[dict], dict], optional): List of plotting
+            keywords for the horizontal lines. Can be a list or single dictionary. These will be
+            passed to plt.axhline(). Common keywords include "linestyle", "linewidth", and "label".
+            If "label" is passed, automatically plots the data set legend with the labels.
+        legend_font_size (float, optional): The font size to use for the legend. Defaults to 10.0
+            + the number of categories.
+        legend_loc (str, optional): The location parameter to pass to ``Axes.legend(loc=)``.
+            Defaults to "center left".
+        legend_bbox_to_anchor (Tuple[float, float], otptional): The bounding box to anchor
+            the legend to. Defaults to (1, 0.5).
+        ax (Axes, optional): A matplotlib axes object to plot the figure on. Defaults to None, in
+            which case the function creates and returns a new axes. The figure height is 6 inches
+            and the figure width is 3 inches times the number of categories.
+
+    Returns:
+        Axes: A ``matplotlib`` axes with a bar plot of the given data.
+    """
+
+    return multi_bar_plot(
+        {data_set_label: data},
[…]
