+++ src/votekit/plots/__init__.py
@@ -1,2 +1,2 @@
 from .mds import plot_MDS, compute_MDS  # noqa
-from .bar_plot import multi_bar_plot  # noqa
+from .bar_plot import multi_bar_plot, bar_plot  # noqa
