### 2025-02-20 7ae2122
remove profile plots
