+++ src/votekit/plots/profile_plots.py
@@ -1,953 +1,0 @@
-from matplotlib import pyplot as plt  # type: ignore
-from votekit.pref_profile import PreferenceProfile
-from votekit.utils import first_place_votes, mentions, COLOR_LIST, borda_scores
-from matplotlib.axes import Axes
-from typing import Optional, Union, Tuple, Literal, Callable
-import matplotlib.patches as mpatches
-from copy import deepcopy
-import warnings
-from matplotlib.legend import Legend
-from matplotlib.lines import Line2D
-
-VALID_STATS = ["fpv", "mentions", "borda", "first place votes"]
-Valid_stat = Literal["fpv", "mentions", "borda", "first place votes"]
-DEFAULT_LINE_KWDS = {
-    "ls": "-",
-    "linewidth": 2,
-}
-
-
-def add_null_keys(data: list[dict[str, float]]) -> list[dict[str, float]]:
-    """
-    Prepares list of dictionaries to be passed to ``bar_plot()``. If a key is missing from
-    a dictionary, this function adds the key with value 0.
-
-    Args:
-        data (list[dict[str, float]]): Categorical data to be cleaned. The value of each dict
-            should be the frequency of the key which is the category name.
-
-    Returns:
-        list[dict[str, float]]: Cleaned data.
-    """
-    data = deepcopy(data)
-
-    x_labels = []
-    for data_dict in data:
-        for x_label in data_dict.keys():
-            if x_label not in x_labels:
-                x_labels.append(x_label)
-
-    for x_label in x_labels:
-        for data_dict in data:
-            if x_label not in data_dict:
-                data_dict[x_label] = 0
-
-    return data
-
-
-def _set_default_bar_plot_args(
-    data: Union[list[dict[str, float]], dict[str, float]],
-    data_set_labels: Optional[list[str]],
-    data_set_to_color: Optional[dict],
-    bar_width: Optional[float],
-    x_label_ordering: Optional[list[str]],
-    legend_font_size: Optional[float],
-    threshold_values: Optional[Union[list[float], float]],
-    threshold_kwds: Optional[Union[list[dict], dict]],
-    ax: Optional[Axes],
-) -> Tuple[
-    list[dict[str, float]],
-    list[str],
-    dict,
-    float,
-    list[str],
-    float,
-    float,
-    Optional[list[float]],
-    Optional[list[dict]],
-    Axes,
-]:
-    """
-    Handles setting default arguments.
-
-    Args:
-        data (Union[list[dict[str, float]], dict[str, float]]): Categorical data to be plotted.
-            If a list is provided, will plot all data sets on same plot. The value of the dict
-            should be the frequency of the key which is the category name.
-        data_set_labels (list[str], optional): List of labels for data sets.
-        data_set_to_color (dict, optional): Dictionary mapping data set labels to colors. Defaults
-            to subset of ``COLOR_LIST`` from ``utils`` module.
-        bar_width (float, optional): Width of bars, defaults to None which computes the bar width
-            as 0.7 divided by the number of data sets.
-        x_label_ordering (list[str], optional): Ordering of x labels. Defaults to order retrieved
-            from first data dictionary.
-        threshold_values (Union[list[float], float], optional): List of values to plot horizontal
-            lines at. Can be provided as a list or a single float.
-        threshold_kwds (Union[list[dict], dict], optional): List of plotting
-            keywords for the horizontal lines. Can be a list or single dictionary.
-        ax (Axes, optional): A matplotlib axes object to plot the figure on. Defaults to None, in
-            which case the function creates and returns a new axes.
-
-    Returns:
[…]
