### 2025-02-20 ae4a3e8
add ballot_length function
