### 2025-02-20 d9339f4
add profile barplot wrapper and tests
