+++ src/votekit/plots/profiles/profile_bar_plot.py
@@ -1,0 +1,191 @@
+from typing import Callable, Optional, Union, Any
+from ...pref_profile import PreferenceProfile
+from matplotlib.axes import Axes
+from ...utils import (
+    first_place_votes,
+    borda_scores,
+    mentions,
+    ballot_lengths,
+    COLOR_LIST,
+)
+from functools import partial
+from ..bar_plot import add_null_keys, multi_bar_plot
+
+plot_functions = {
+    "first place votes": first_place_votes,
+    "mentions": mentions,
+    "borda": borda_scores,
+    "ballot length": ballot_lengths,
+}
+
+
+def _partial_stat_function(
+    stat_function: Union[str, Callable[[PreferenceProfile], dict[str, float]]],
+    stat_function_kwds: Optional[dict],
+):
+    """
+    Takes the stat function string and returns the corresponding callable with the correct
+    keywords.
+
+    Args:
+        stat_function (Union[str, Callable[[PreferenceProfile], dict[str, float]]]): Which stat
+            to use for the bar plot. Can be one of "first place votes", "mentions", "borda", or
+            "ballot_length", or it can be a custom callable that takes a profile and returns
+            a dict with str keys and float values.
+        stat_function_kwds (dict[str, Any], optional): Keyword arguments to pass to
+            ``stat_function``.
+    """
+    if isinstance(stat_function, str):
+        stat_function = plot_functions[stat_function]  # type: ignore[assignment]
+
+        if stat_function_kwds and callable(stat_function):
+            stat_function = partial(stat_function, **stat_function_kwds)
+
+    return stat_function
+
+
+def _validate_inputs(
+    stat_function: Union[str, Callable[[PreferenceProfile], dict[str, float]]]
+):
+    """
+    Validate inputs to bar plot.
+
+    Args:
+        stat_function (Union[str, Callable[[PreferenceProfile], dict[str, float]]]): Which stat
+            to use for the bar plot. Can be one of "first place votes", "mentions", "borda", or
+            "ballot_length", or it can be a custom callable that takes a profile and returns
+            a dict with str keys and float values.
+
+    Raises:
+        ValueError: if stat_function str is not one of the defined stats.
+    """
+
+    if isinstance(stat_function, str):
+        if stat_function not in plot_functions:
+            raise ValueError(
+                (
+                    f"stat_function string {stat_function} not an available statistic."
+                    f" Available stats include: {list(plot_functions.keys())}"
+                )
+            )
+
+
+def profile_multi_bar_plot(
+    profile_dict: dict[str, PreferenceProfile],
+    stat_function: Union[str, Callable[[PreferenceProfile], dict[str, float]]],
+    stat_function_kwds: Optional[dict[str, Any]] = None,
+    normalize: bool = False,
+    profile_colors: Optional[dict[str, str]] = None,
+    bar_width: Optional[float] = None,
+    category_ordering: Optional[list[str]] = None,
+    x_axis_name: Optional[str] = None,
+    y_axis_name: Optional[str] = None,
+    title: Optional[str] = None,
+    show_profile_legend: bool = False,
+    categories_legend: Optional[dict[str, str]] = None,
+    threshold_values: Optional[Union[list[float], float]] = None,
+    threshold_kwds: Optional[Union[list[dict], dict]] = None,
+    legend_font_size: Optional[float] = None,
+    ax: Optional[Axes] = None,
+) -> Axes:
+    """
+    Plot a multi bar plot over a set of preference profiles and some statistic about the profile,
+    like ballot length, first place votes for candidate, etc.
+
+    Args:
+        profile_dict (dict[str, PreferenceProfile]): Keys are profile labels and values are
+            profiles to plot statistics for.
[…]
