+++ src/votekit/plots/profiles/__init__.py
@@ -1,0 +1,1 @@
+from .profile_bar_plot import profile_multi_bar_plot  # noqa
