### 2025-02-20 ad47df3
fix color list length error
