+++ tests/representation_scores/test_r_representation_score.py
@@ -1,6 +1,12 @@
 from votekit.representation_scores import r_representation_score
 from votekit import PreferenceProfile, Ballot
+from votekit.cvr_loaders import load_csv
+from votekit.utils import remove_cand
+from pathlib import Path
 import pytest
+
+BASE_DIR = Path(__file__).resolve().parent.parent
+CSV_DIR = BASE_DIR / "data/csv/"
 
 profile = PreferenceProfile(
     ballots=(
@@ -36,3 +42,27 @@ def test_r_rep_score_multi_cand():
 def test_r_rep_score_error():
     with pytest.raises(ValueError, match="r \(0\) must be at least 1."):
         r_representation_score(PreferenceProfile(), 0, [])
+
+
+def test_r_rep_score_warning():
+    with pytest.warns(
+        UserWarning, match="are not found in the profile's candidate list:"
+    ):
+        r_representation_score(profile, 3, ["David", "Chris", "Jeanne", "Mala"])
+
+
+def test_r_rep_score_portland():
+    profile = load_csv(
+        CSV_DIR / "Portland_D1_Condensed.csv", rank_cols=[1, 2, 3, 4, 5, 6]
+    )
+    clean_profile = remove_cand("skipped", profile)
+
+    assert (
+        round(
+            r_representation_score(
+                clean_profile, 3, ["Candace Avalos", "Loretta Smith", "Jamie Dunphy"]
+            ),
+            3,
+        )
+        == 0.709
+    )
