+++ pyproject.toml
@@ -14,7 +14,7 @@ dependencies = [
     "networkx (>=3.1, <4.0)",
     "jinja2 (>=3.1.2, <4.0)",
     "matplotlib (>=3.7.2, <4.0)",
-    "pandas (==2.2)",
+    "pandas (>2.2)",
     "apportionment (>=1.0, <2.0)",
     "scikit-learn (>=1.3.2, <2.0)",
     "numpy (>=1.26.0, <2.0)",
