### 2025-02-25 3f10931
bar_plot edge cut off fixed
