### 2025-02-25 9a7d864
update cat legend behavior
