+++ src/votekit/plots/profiles/__init__.py
@@ -1,1 +1,1 @@
-from .profile_bar_plot import profile_multi_bar_plot  # noqa
+from .profile_bar_plot import profile_multi_bar_plot, profile_bar_plot  # noqa
