+++ src/votekit/plots/profiles/profile_bar_plot.py
@@ -186,6 +186,90 @@ def profile_multi_bar_plot(
             raise e
 
 
-# TODO
-def profile_bar_plot():
-    pass
+def profile_bar_plot(
+    profile: PreferenceProfile,
+    stat_function: Union[str, Callable[[PreferenceProfile], dict[str, float]]],
+    *,
+    profile_label: str = "Profile",
+    stat_function_kwds: Optional[dict[str, Any]] = None,
+    normalize: bool = False,
+    profile_color: str = COLOR_LIST[0],
+    bar_width: Optional[float] = None,
+    category_ordering: Optional[list[str]] = None,
+    x_axis_name: Optional[str] = None,
+    y_axis_name: Optional[str] = None,
+    title: Optional[str] = None,
+    show_profile_legend: bool = False,
+    categories_legend: Optional[dict[str, str]] = None,
+    threshold_values: Optional[Union[list[float], float]] = None,
+    threshold_kwds: Optional[Union[list[dict], dict]] = None,
+    legend_font_size: Optional[float] = None,
+    ax: Optional[Axes] = None,
+) -> Axes:
+    """
+    Plots bar plot of single profile. Wrapper for ``profile_multi_bar_plot``.
+
+    Args:
+        profile (PreferenceProfile): Profile to plot statistics for.
+        stat_function (Union[str, Callable[[PreferenceProfile], dict[str, float]]]): Which stat
+            to use for the bar plot. Can be one of "first place votes", "mentions", "borda", or
+            "ballot_length", or it can be a custom callable that takes a profile and returns
+            a dict with str keys and float values.
+        profile_label (str, optional): Label for profile. Defaults to "Profile".
+        stat_function_kwds (dict[str, Any], optional): Keyword arguments to pass to
+            ``stat_function``. Defaults to None, in which case default values for ``stat_function``
+            are used.
+        normalize (bool, optional): Whether or not to normalize data. Defaults to False.
+        profile_color (str, optional): Color to plot. Defaults to the first color from
+            ``COLOR_LIST`` from ``utils`` module.
+        bar_width (float, optional): Width of bars. Defaults to None which computes the bar width
+            as 0.7 divided by the number of data sets. Must be in the interval :math:`(0,1]`.
+        category_ordering (list[str], optional): Ordering of x-labels. Defaults to order retrieved
+            from data dictionary.
+        x_axis_name (str, optional): Name of x-axis. Defaults to None, which does not plot a name.
+        y_axis_name (str, optional): Name of y-axis. Defaults to None, which does not plot a name.
+        title (str, optional): Title for the figure. Defaults to None, which does not plot a title.
+        show_profile_legend (bool, optional): Whether or not to plot the profile legend.
+            Defaults to False. Is automatically shown if any threshold lines have the keyword
+            "label" passed through ``threshold_kwds``.
+        categories_legend (dict[str, str], optional): Dictionary mapping data categories
+            to description. Defaults to None. If provided, generates a second legend for data
+            categories.
+        threshold_values (Union[list[float], float], optional): List of values to plot horizontal
+            lines at. Can be provided as a list or a single float.
+        threshold_kwds (Union[list[dict], dict], optional): List of plotting
+            keywords for the horizontal lines. Can be a list or single dictionary. These will be
+            passed to plt.axhline(). Common keywords include "linestyle", "linewidth", and "label".
+            If "label" is passed, automatically plots the data set legend with the labels.
+        legend_font_size (float, optional): The font size to use for the legend. Defaults to 10.0
+            + the number of categories.
+        legend_loc (str, optional): The location parameter to pass to ``Axes.legend(loc=)``.
+            Defaults to "center left".
[…]
