+++ src/votekit/plots/__init__.py
@@ -1,3 +1,3 @@
 from .mds import plot_MDS, compute_MDS  # noqa
 from .bar_plot import multi_bar_plot, bar_plot  # noqa
-from .profiles import profile_multi_bar_plot  # noqa
+from .profiles import profile_multi_bar_plot, profile_bar_plot  # noqa
