+++ tests/plots/bar_plot/profile_bar_plot/test_profile_multi_bar_plot.py
@@ -1,9 +1,10 @@
-from votekit.plots.profiles import profile_multi_bar_plot
+from votekit.plots.profiles import multi_profile_bar_plot
 from votekit.pref_profile import PreferenceProfile
 from votekit.utils import COLOR_LIST
 from votekit.ballot import Ballot
 from matplotlib.axes import Axes
 import pytest
+import matplotlib.pyplot as plt
 
 ballot_1 = Ballot(
     ranking=(frozenset({"Chris"}), frozenset({"Peter"}), frozenset({"Moon"})), weight=1
@@ -17,38 +18,43 @@ profile_2 = PreferenceProfile(ballots=(ballot_1, ballot_2, ballot_4))
 
 
 def test_barplot_with_defaults():
-    ax = profile_multi_bar_plot(
+    ax = multi_profile_bar_plot(
         {"Profile 1": profile_1, "Profile 2": profile_2},
         stat_function="first place votes",
     )
     assert isinstance(ax, Axes)
+    plt.close()
 
-    ax = profile_multi_bar_plot(
+    ax = multi_profile_bar_plot(
         {"Profile 1": profile_1, "Profile 2": profile_2}, stat_function="mentions"
     )
     assert isinstance(ax, Axes)
+    plt.close()
 
-    ax = profile_multi_bar_plot(
+    ax = multi_profile_bar_plot(
         {"Profile 1": profile_1, "Profile 2": profile_2}, stat_function="borda"
     )
     assert isinstance(ax, Axes)
+    plt.close()
 
-    ax = profile_multi_bar_plot(
-        {"Profile 1": profile_1, "Profile 2": profile_2}, stat_function="ballot length"
+    ax = multi_profile_bar_plot(
+        {"Profile 1": profile_1, "Profile 2": profile_2}, stat_function="ballot lengths"
     )
     assert isinstance(ax, Axes)
+    plt.close()
 
 
 def test_barplot_with_callable():
-    ax = profile_multi_bar_plot(
+    ax = multi_profile_bar_plot(
         {"Profile 1": profile_1, "Profile 2": profile_2},
         stat_function=lambda x: {"1": 1},
     )
     assert isinstance(ax, Axes)
+    plt.close()
 
 
 def test_barplot_with_no_defaults():
-    ax = profile_multi_bar_plot(
+    ax = multi_profile_bar_plot(
         {"Profile 1": profile_1, "Profile 2": profile_2},
         stat_function="first place votes",
         stat_function_kwds={"to_float": True},
@@ -66,13 +72,14 @@ def test_barplot_with_no_defaults():
         legend_font_size=4,
     )
     assert isinstance(ax, Axes)
+    plt.close()
 
 
 def test_wrapped_errors():
     with pytest.raises(
         ValueError, match=f"Cannot plot more than {len(COLOR_LIST)} profiles."
     ):
-        profile_multi_bar_plot(
+        multi_profile_bar_plot(
             {f"Profile_{i}": profile_1 for i in range(len(COLOR_LIST) + 2)},
             stat_function="borda",
         )
@@ -80,7 +87,7 @@ def test_wrapped_errors():
     with pytest.raises(
         ValueError, match="category_ordering must be the same length as the dictionary"
     ):
-        profile_multi_bar_plot(
+        multi_profile_bar_plot(
             {f"Profile_{i}": profile_1 for i in range(3)},
             stat_function="borda",
             category_ordering=["Chris"],
