+++ src/votekit/plots/__init__.py
@@ -1,3 +1,14 @@
 from .mds import plot_MDS, compute_MDS  # noqa
 from .bar_plot import multi_bar_plot, bar_plot  # noqa
-from .profiles import profile_multi_bar_plot, profile_bar_plot  # noqa
+from .profiles import (
+    multi_profile_bar_plot,  # noqa
+    multi_profile_mentions_plot,  # noqa
+    multi_profile_ballot_lengths_plot,  # noqa
+    multi_profile_borda_plot,  # noqa
+    multi_profile_fpv_plot,  # noqa
+    profile_bar_plot,  # noqa
+    profile_borda_plot,  # noqa
+    profile_ballot_lengths_plot,  # noqa
+    profile_fpv_plot,  # noqa
+    profile_mentions_plot,  # noqa
+)
