+++ src/votekit/plots/profiles/__init__.py
@@ -1,1 +1,12 @@
-from .profile_bar_plot import profile_multi_bar_plot, profile_bar_plot  # noqa
+from .profile_bar_plot import (
+    multi_profile_bar_plot,  # noqa
+    multi_profile_ballot_lengths_plot,  # noqa
+    multi_profile_borda_plot,  # noqa
+    multi_profile_fpv_plot,  # noqa
+    multi_profile_mentions_plot,  # noqa
+    profile_bar_plot,  # noqa
+    profile_borda_plot,  # noqa
+    profile_mentions_plot,  # noqa
+    profile_fpv_plot,  # noqa
+    profile_ballot_lengths_plot,  # noqa
+)
