+++ src/votekit/plots/profiles/profile_bar_plot.py
@@ -15,7 +15,7 @@ plot_functions = {
     "first place votes": first_place_votes,
     "mentions": mentions,
     "borda": borda_scores,
-    "ballot length": ballot_lengths,
+    "ballot lengths": ballot_lengths,
 }
 
 
@@ -70,7 +70,7 @@ def _validate_inputs(
             )
 
 
-def profile_multi_bar_plot(
+def multi_profile_bar_plot(
     profile_dict: dict[str, PreferenceProfile],
     stat_function: Union[str, Callable[[PreferenceProfile], dict[str, float]]],
     stat_function_kwds: Optional[dict[str, Any]] = None,
@@ -186,6 +186,440 @@ def profile_multi_bar_plot(
             raise e
 
 
+def multi_profile_borda_plot(
+    profile_dict: dict[str, PreferenceProfile],
+    borda_kwds: Optional[dict[str, Any]] = None,
+    normalize: bool = False,
+    profile_colors: Optional[dict[str, str]] = None,
+    bar_width: Optional[float] = None,
+    candidate_ordering: Optional[list[str]] = None,
+    x_axis_name: Optional[str] = None,
+    y_axis_name: Optional[str] = None,
+    title: Optional[str] = None,
+    show_profile_legend: bool = False,
+    candidate_legend: Optional[dict[str, str]] = None,
+    relabel_candidates_with_int: bool = False,
+    threshold_values: Optional[Union[list[float], float]] = None,
+    threshold_kwds: Optional[Union[list[dict], dict]] = None,
+    legend_font_size: Optional[float] = None,
+    ax: Optional[Axes] = None,
+) -> Axes:
+    """
+    Plot the borda scores for a collection of profiles. Wrapper for ``multi_profile_bar_plot``.
+
+    Args:
+        profile_dict (dict[str, PreferenceProfile]): Keys are profile labels and values are
+            profiles to plot statistics for.
+        borda_kwds (dict[str, Any], optional): Keyword arguments to pass to
+            ``borda_scores``. Defaults to None, in which case default values for ``borda_scores``
+            are used.
+        normalize (bool, optional): Whether or not to normalize data. Defaults to False.
+        profile_colors (dict[str, str], optional): Dictionary mapping profile labels
+            to colors. Defaults to None, in which case we use a subset of ``COLOR_LIST``
+            from ``utils`` module. Dictionary keys can be a subset of the profiles.
+        bar_width (float, optional): Width of bars. Defaults to None which computes the bar width
+            as 0.7 divided by the number of data sets. Must be in the interval :math:`(0,1]`.
+        candidate_ordering (list[str], optional): Ordering of x-labels. Defaults to decreasing
+            borda scores from the first profile.
+        x_axis_name (str, optional): Name of x-axis. Defaults to None, which does not plot a name.
+        y_axis_name (str, optional): Name of y-axis. Defaults to None, which does not plot a name.
+        title (str, optional): Title for the figure. Defaults to None, which does not plot a title.
+        show_profile_legend (bool, optional): Whether or not to plot the profile legend.
+            Defaults to False. Is automatically shown if any threshold lines have the keyword
+            "label" passed through ``threshold_kwds``.
+        candidate_legend (dict[str, str], optional): Dictionary mapping candidates
+            to relableing. Defaults to None. If provided, generates a second legend for data
+            categories.
+        relabel_candidates_with_int (bool, optional): Relabel the candidates with integer labels.
+            Defaults to False. If ``candidate_legend`` is passed, those labels supercede.
+        threshold_values (Union[list[float], float], optional): List of values to plot horizontal
+            lines at. Can be provided as a list or a single float.
+        threshold_kwds (Union[list[dict], dict], optional): List of plotting
+            keywords for the horizontal lines. Can be a list or single dictionary. These will be
[…]
