+++ src/votekit/plots/profiles/multi_profile_bar_plot.py
@@ -1,7 +1,5 @@
 from typing import Callable, Optional, Union, Any
 from ...pref_profile import PreferenceProfile
-from .multi_profile_bar_plot import multi_profile_bar_plot
-from ..bar_plot import bar_plot
 from matplotlib.axes import Axes
 from ...utils import (
     first_place_votes,
@@ -10,15 +8,42 @@ from ...utils import (
     ballot_lengths,
     COLOR_LIST,
 )
+from functools import partial
+from ..bar_plot import add_null_keys, multi_bar_plot
 
 
-def profile_bar_plot(
-    profile: PreferenceProfile,
+def _create_data_dict(
+    profile_dict: dict[str, PreferenceProfile],
     stat_function: Callable[[PreferenceProfile], dict[str, float]],
-    *,
-    profile_label: str = "Profile",
+) -> dict[str, dict[str, float]]:
+    """
+    Create the correctly formatted dict to pass to ``multi_bar_plot``. Ensures each
+    subdictionary has the same keys, and uses the default value 0 if a key is missing.
+
+    Args:
+        profile_dict (dict[str, PreferenceProfile]): Keys are profile labels and values are
+            profiles to plot statistics for.
+        stat_function (Callable[[PreferenceProfile], dict[str, float]]): Which stat
+            to use for the bar plot. Must be a callable that takes a profile and returns
+            a dict with str keys and float values.
+
+    Returns:
+        dict[str, dict[str, float]]: Data dictionary for ``multi_bar_plot``.
+
+    """
+
+    data_dict = {
+        label: stat_function(profile) for label, profile in profile_dict.items()
+    }
+
+    return add_null_keys(data_dict)
+
+
+def multi_profile_bar_plot(
+    profile_dict: dict[str, PreferenceProfile],
+    stat_function: Callable[[PreferenceProfile], dict[str, float]],
     normalize: bool = False,
-    profile_color: str = COLOR_LIST[0],
+    profile_colors: Optional[dict[str, str]] = None,
     bar_width: Optional[float] = None,
     category_ordering: Optional[list[str]] = None,
     x_axis_name: Optional[str] = None,
@@ -32,17 +57,19 @@ def profile_bar_plot(
     ax: Optional[Axes] = None,
 ) -> Axes:
     """
-    Plots bar plot of single profile. Wrapper for ``multi_profile_bar_plot``.
+    Plot a multi bar plot over a set of preference profiles and some statistic about the profile,
+    like ballot length, first place votes for candidate, etc.
 
     Args:
-        profile (PreferenceProfile): Profile to plot statistics for.
+        profile_dict (dict[str, PreferenceProfile]): Keys are profile labels and values are
+            profiles to plot statistics for.
         stat_function (Callable[[PreferenceProfile], dict[str, float]]): Which stat
             to use for the bar plot. Must be a callable that takes a profile and returns
             a dict with str keys and float values.
-        profile_label (str, optional): Label for profile. Defaults to "Profile".
         normalize (bool, optional): Whether or not to normalize data. Defaults to False.
-        profile_color (str, optional): Color to plot. Defaults to the first color from
-            ``COLOR_LIST`` from ``utils`` module.
+        profile_colors (dict[str, str], optional): Dictionary mapping profile labels
+            to colors. Defaults to None, in which case we use a subset of ``COLOR_LIST``
+            from ``utils`` module. Dictionary keys can be a subset of the profiles.
         bar_width (float, optional): Width of bars. Defaults to None which computes the bar width
             as 0.7 divided by the number of data sets. Must be in the interval :math:`(0,1]`.
         category_ordering (list[str], optional): Ordering of x-labels. Defaults to order retrieved
@@ -75,33 +102,48 @@ def profile_bar_plot(
     Returns:
[…]
