### 2025-03-03 27e24b1
update notebook3 rst
