+++ docs/user/tutorial/3_viz.rst
@@ -16,15 +16,15 @@ are far less realistic and flexible than the models we ran in previous
 sections!
 
 Impartial Culture (IC) is essentially the :math:`\alpha=\infty` extreme
-of the family of Dirichlet measures :ref:`from
-earlier<user/tutorial/2_real_and_simulated_profiles:Dirichlet Distribution>`;
+of the family of Dirichlet measures `from
+earlier <2_real_and_simulated_profiles.html#dirichlet-distribution>`__;
 that is, the probability of each ranking is set exactly equal. Impartial
 Anonymous Culture (IAC) is the :math:`\alpha=1` (“all bets are off”)
 case.
 
 .. code:: ipython3
 
-    from votekit.plots import plot_summary_stats
+    from votekit.plots import multi_profile_fpv_plot, profile_fpv_plot
     import votekit.ballot_generator as bg
     
     # generate a profile to work with first
@@ -48,60 +48,71 @@ case.
 
     IC profile:
       Ranking Scores Weight
-    (A, C, B)     ()    203
-    (C, B, A)     ()    182
-    (B, C, A)     ()    168
+    (C, B, A)     ()    194
+    (A, B, C)     ()    169
+    (A, C, B)     ()    166
+    (C, A, B)     ()    160
     (B, A, C)     ()    157
-    (A, B, C)     ()    149
-    (C, A, B)     ()    141
+    (B, C, A)     ()    154
     
     IAC profile:
       Ranking Scores Weight
-    (A, C, B)     ()    378
-    (C, B, A)     ()    197
-    (B, A, C)     ()    144
-    (A, B, C)     ()    140
-    (C, A, B)     ()    123
-    (B, C, A)     ()     18
+    (C, B, A)     ()    593
+    (A, B, C)     ()    235
+    (B, C, A)     ()     90
+    (A, C, B)     ()     49
+    (B, A, C)     ()     30
+    (C, A, B)     ()      3
 
 
-Now we’ll plot some summary statistics for the generated elections.
+Now we’ll plot some summary statistics for the generated profiles.
 
--  ``first place votes`` will measure how many first place votes each
-   candidate received.
+- ``first place votes`` will measure how many first place votes each
+  candidate received.
 
--  ``borda`` reports the **Borda score** of each candidate. If there are
-   :math:`n` candidates on a ballot, the first place candidate gets
-   :math:`n` points, the second :math:`n-1`, and so on.
+- ``borda`` reports the **Borda score** of each candidate. If there are
+  :math:`n` candidates on a ballot, the first place candidate gets
+  :math:`n` points, the second :math:`n-1`, and so on.
 
--  ``mentions`` simply counts the number of times candidates were listed
-   at all. Note that if we use generative methods that produce complete
-   rankings, everyone will necessarily have the same number of mentions!
+- ``mentions`` simply counts the number of times candidates were listed
+  at all. Note that if we use generative methods that produce complete
+  rankings, everyone will necessarily have the same number of mentions!
+
+- ``ballot lengths`` is the distribution of ballot lengths in the
+  profile. Again, if we use generative methods that produce complete
+  rankings, every ballot will be the same length.
+
+We can plot first place votes for one profile or for multiple profiles
+as follows.
 
 .. code:: ipython3
 
-    fig1 = plot_summary_stats(profile1, "first place votes", multi_color = False, 
[…]
