### 2025-03-03 9b5c237
add nbconvert to dev requirements
