+++ pyproject.toml
@@ -36,6 +36,7 @@ recommonmark = "^0.7.1"
 sphinx-rtd-theme = "^3.0.2"
 seaborn = "^0.13.2"
 ipykernel = "^6.29.5"
+nbconvert = "^7.16.6"
 
 [tool.ruff]
 line-length = 100
