+++ docs/requirements.txt
@@ -1,87 +1,119 @@
 alabaster==0.7.16
+annotated-types==0.7.0
 appnope==0.1.4
 apportionment==1.0
-asttokens==2.4.1
-attrs==23.2.0
-Babel==2.15.0
-beautifulsoup4==4.12.3
-bleach==6.1.0
-certifi==2024.6.2
-charset-normalizer==3.3.2
+asttokens==3.0.0
+attrs==25.1.0
+babel==2.16.0
+beautifulsoup4==4.13.3
+black==24.10.0
+bleach==6.2.0
+certifi==2024.12.14
+cfgv==3.4.0
+charset-normalizer==3.4.1
+click==8.1.8
 comm==0.2.2
 commonmark==0.9.1
-contourpy==1.2.1
+contourpy==1.3.1
+coverage==7.6.10
 cycler==0.12.1
-debugpy==1.8.1
+debugpy==1.8.12
 decorator==5.1.1
 defusedxml==0.7.1
-docutils==0.20.1
-executing==2.0.1
-fastjsonschema==2.19.1
-fonttools==4.53.0
-idna==3.7
+distlib==0.3.9
+docutils==0.21.2
+executing==2.2.0
+fastjsonschema==2.21.1
+filelock==3.17.0
+fonttools==4.55.6
+identify==2.6.6
+idna==3.10
 imagesize==1.4.1
-jedi==0.19.1
-Jinja2==3.1.4
+iniconfig==2.0.0
+ipykernel==6.29.5
+ipython==8.31.0
+jedi==0.19.2
+Jinja2==3.1.5
 joblib==1.4.2
-jsonschema==4.22.0
-jsonschema-specifications==2023.12.1
-jupyter==1.0.0
-kiwisolver==1.4.5
-MarkupSafe==2.1.5
-matplotlib==3.9.0
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+jupyter_client==8.6.3
+jupyter_core==5.7.2
+jupyterlab_pygments==0.3.0
+kiwisolver==1.4.8
+MarkupSafe==3.0.2
+matplotlib==3.10.0
 matplotlib-inline==0.1.7
-mistune==3.0.2
-mypy==1.10.0
+mistune==3.1.2
+mypy==1.15.0
 mypy-extensions==1.0.0
-networkx==3.2
+nbclient==0.10.2
+nbconvert==7.16.6
+nbformat==5.10.4
+nest-asyncio==1.6.0
+networkx==3.4.2
+nodeenv==1.9.1
 numpy==1.26.4
-packaging==24.0
-pandas==1.5.3
-pandas-stubs==2.2.2.240514
+packaging==24.2
+pandas==2.2.0
+pandas-stubs==2.0.3.230814
+pandoc==2.4
 pandocfilters==1.5.1
 parso==0.8.4
+pathspec==0.12.1
 pexpect==4.9.0
-pillow==10.3.0
-platformdirs==4.2.2
+pillow==11.1.0
+platformdirs==4.3.6
+pluggy==1.5.0
+plumbum==1.9.0
+ply==3.11
 POT==0.9.4
-prompt_toolkit==3.0.45
-psutil==5.9.8
+pre-commit==3.8.0
+prompt_toolkit==3.0.50
+psutil==7.0.0
 ptyprocess==0.7.0
-pure-eval==0.2.2
-pydantic==2.8.2
-Pygments==2.18.0
-pyparsing==3.1.2
+pure_eval==0.2.3
+pydantic==2.10.6
+pydantic_core==2.27.2
+Pygments==2.19.1
+pyparsing==3.2.1
+pytest==7.4.4
[…]
