### 2025-03-03 b1597da
update notebook 3 rst
