### 2025-03-07 74cce7f
update PWCG class, split out pairwise dict
