+++ src/votekit/graphs/pairwise_comparison_graph.py
@@ -1,197 +1,201 @@
 from fractions import Fraction
-from itertools import permutations, combinations
+from itertools import combinations
 import matplotlib.pyplot as plt  # type: ignore
+from matplotlib.axes import Axes
+import matplotlib.patches as mpatches
 import networkx as nx  # type: ignore
 from functools import cache
 from typing import Optional
-
-from ..ballot import Ballot
-from .base_graph import Graph
 from ..pref_profile import PreferenceProfile
 
 
-class PairwiseComparisonGraph(Graph):
+def pairwise_dict(
+    profile: PreferenceProfile,
+) -> dict[tuple[str, ...], list[Fraction]]:
+    """
+    Computes a dictionary whose keys are candidate pairs (A,B) and whose values are lists [a,b]
+    where a denotes the number of times A beats B head to head, and b is the reverse.
+
+    Args:
+        profile (PreferenceProfile): Profile to compute dict on.
+
+    Returns:
+        dict[tuple[str, ...], list[Fraction]]: Pairwise comparison dictionary.
+
+    """
+    pairwise_dict = {
+        tuple(sorted((c1, c2))): [Fraction(0), Fraction(0)]
+        for c1, c2 in combinations(profile.candidates, 2)
+    }
+
+    candidate_set = set(profile.candidates)
+
+    mentioned_pairs_dict: dict[str, list[tuple[str, str]]] = {
+        c: [] for c in profile.candidates
+    }
+    for c1, c2 in pairwise_dict.keys():
+        mentioned_pairs_dict[c1].append((c1, c2))
+        mentioned_pairs_dict[c2].append((c1, c2))
+
+    sets_mentioned_pairs_dict = {
+        c: set(mentioned_pairs_dict[c]) for c in mentioned_pairs_dict
+    }
+
+    for _, row in profile.df.iterrows():
+        ranking = row["Ranking"]
+
+        mentioned_so_far = set()
+        for i in range(len(ranking)):
+            cand = ranking[i]
+            if cand in mentioned_so_far:
+                continue
+
+            mentioned_so_far.add(cand)
+
+            remaining_candidates = candidate_set - mentioned_so_far
+            for c1, c2 in sets_mentioned_pairs_dict[cand]:
+                if c1 in remaining_candidates or c2 in remaining_candidates:
+                    if c1 == cand:
+                        pairwise_dict[(c1, c2)][0] += row["Weight"]
+                    else:
+                        pairwise_dict[(c1, c2)][1] += row["Weight"]
+    return pairwise_dict
+
+
+def restrict_pairwise_dict_to_subset(
+    cand_subset: list[str] | tuple[str] | set[str],
+    pairwise_dict: dict[tuple[str, ...], list[Fraction]],
+) -> dict[tuple[str, ...], list[Fraction]]:
+    """
+    Restricts the full pairwise dictionary to a subset of candidates. The pairwise dictionary is a
+    dictionary whose keys are candidate pairs (A,B) and whose values are lists [a,b]
+    where a denotes the number of times A beats B head to head, and b is the reverse.
+
+    Args:
+        cands (list[str] | tuple[str] | set[str]): Candidate subset to restrict to.
+        pairwise_dict (dict[tuple[str, ...], tuple[int,int]]): Full pairwise comparison dictionary.
+
+    Returns:
+        dict[tuple[str, ...], list[Fraction]]: Pairwise dict restricted to the provided
+            candidates.
+
+    Raises:
+        ValueError: cand_subset must be at least length 2.
+        ValueError: cand_subset must be a subset of the candidates in the dictionary.
+    """
+    if len(cand_subset) < 2:
+        raise ValueError(
[…]
