+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+
+
+## Changed 
+- renamed `cleaning.deduplicate_profile` to `cleaning.remove_repeated_candidates`.
+- moved `remove_cand` from `utils` to `cleaning`, and removed the older function `remove_noncands`.
+
+## Fixed
 
 ## [3.1.0] - 2025-03-03
 ## Added
