### 2025-03-18 #7477 @britta-wstnr
Hi @cdonnay - just a gentle check-in on the DOI. I know things like this take time, so no rush from my side - just want to pop in here and see if there is anything you would need help with etc. 🌱
