### 2025-03-19 9f108d0
update condense_profile to include voter_set
