+++ tests/test_pref_profile.py
@@ -194,13 +194,24 @@ def test_to_scores_dict():
 def test_condense_profile_ranking():
     profile = PreferenceProfile(
         ballots=(
-            Ballot(ranking=({"A"}, {"B"}, {"C"}), weight=Fraction(1)),
             Ballot(ranking=({"A"}, {"B"}, {"C"}), weight=Fraction(2)),
+            Ballot(
+                ranking=({"A"}, {"B"}, {"C"}), weight=Fraction(1), voter_set={"Chris"}
+            ),
+            Ballot(
+                ranking=({"A"}, {"B"}, {"C"}),
+                weight=Fraction(2),
+                voter_set={"Moon", "Peter"},
+            ),
         ),
         candidates=("A", "B", "C", "D"),
     )
     pp = profile.condense_ballots()
-    assert pp.ballots[0] == Ballot(ranking=({"A"}, {"B"}, {"C"}), weight=Fraction(3))
+    assert pp.ballots[0] == Ballot(
+        ranking=({"A"}, {"B"}, {"C"}),
+        weight=Fraction(5),
+        voter_set={"Chris", "Moon", "Peter"},
+    )
     assert set(pp.candidates) == set(profile.candidates)
 
 
@@ -211,19 +222,37 @@ def test_condense_profile_scores():
                 ranking=({"A"}, {"B"}, {"C"}),
                 scores={"A": 3, "B": 2},
                 weight=Fraction(1),
+                voter_set={"Chris"},
             ),
             Ballot(ranking=({"A"}, {"B"}, {"C"}), weight=Fraction(2)),
             Ballot(
                 ranking=({"A"}, {"B"}, {"C"}),
                 scores={"A": 3, "B": 2},
                 weight=Fraction(2),
+                voter_set={"Peter", "Moon"},
             ),
         )
     )
     pp = profile.condense_ballots()
-    assert pp.ballots[0] == Ballot(
-        ranking=({"A"}, {"B"}, {"C"}), scores={"A": 3, "B": 2}, weight=Fraction(3)
+
+    for b in pp.ballots:
+        print(b.ranking)
+        print(b.scores)
+        print(b.voter_set)
+        print(b.weight)
+        print()
+
+    assert (
+        Ballot(
+            ranking=({"A"}, {"B"}, {"C"}),
+            scores={"A": 3, "B": 2},
+            weight=Fraction(3),
+            voter_set={"Chris", "Moon", "Peter"},
+        )
+        in pp.ballots
     )
+
+    assert Ballot(ranking=({"A"}, {"B"}, {"C"}), weight=Fraction(2)) in pp.ballots
 
 
 def test_profile_equals():
