### 2025-03-19 1b3035a
rename condense_ballots to group_ballots
