+++ src/votekit/utils.py
@@ -167,7 +167,7 @@ def remove_cand(
             )
 
         if condense:
-            clean_profile = clean_profile.condense_ballots()
+            clean_profile = clean_profile.group_ballots()
 
         return cast(COB, clean_profile)
 
@@ -184,7 +184,7 @@ def remove_cand(
             )
 
         if condense:
-            clean_profile = clean_profile.condense_ballots()
+            clean_profile = clean_profile.group_ballots()
 
         return cast(COB, clean_profile.ballots[0])
     else:
@@ -200,7 +200,7 @@ def remove_cand(
             )
 
         if condense:
-            clean_profile = clean_profile.condense_ballots()
+            clean_profile = clean_profile.group_ballots()
 
         return cast(COB, clean_profile.ballots)
 
@@ -242,7 +242,7 @@ def add_missing_cands(profile: PreferenceProfile) -> PreferenceProfile:
 
     return PreferenceProfile(
         ballots=tuple(new_ballots), candidates=tuple(candidates)
-    ).condense_ballots()
+    ).group_ballots()
 
 
 def validate_score_vector(score_vector: Sequence[Union[float, Fraction]]):
@@ -695,7 +695,7 @@ def resolve_profile_ties(profile: PreferenceProfile) -> PreferenceProfile:
     new_ballots = tuple(
         [b for ballot in profile.ballots for b in expand_tied_ballot(ballot)]
     )
-    return PreferenceProfile(ballots=new_ballots).condense_ballots()
+    return PreferenceProfile(ballots=new_ballots).group_ballots()
 
 
 def score_profile_from_ballot_scores(
