### 2025-03-19 e7ef835
rename condense_ballots to group_ballots
