### 2025-03-19 db656ca
add ballot id and ranking to ballot functions
