### 2025-03-19 0331af0
add ballot id and voterset to remove_cand
