### 2025-03-19 6174bb6
finish notebook code
