### 2025-03-20 3d11a73
adjust cleaning module, align format of all functions
