+++ src/votekit/cleaning.py
@@ -1,274 +1,281 @@
 from fractions import Fraction
-from functools import reduce
-from typing import Callable, Union, TypeVar, cast
+from functools import partial
+from typing import Callable, Union, overload, Literal
 
 from .pref_profile import PreferenceProfile
 from .ballot import Ballot
 
 
-def remove_empty_ballots(
-    pp: PreferenceProfile, keep_candidates: bool = False
-) -> PreferenceProfile:
-    """
-    Removes empty ballots from a PreferenceProfile.
-
-    Args:
-        pp (PreferenceProfile): A PreferenceProfile to clean.
-        keep_candidates (bool, optional): If True, keep all of the candidates
-            from the original PreferenceProfile in the returned PreferenceProfile, even if
-            they got no votes. Defaults to False.
-
-    Returns:
-        PreferenceProfile: A cleaned PreferenceProfile.
-    """
-
-    ballots_nonempty = tuple([ballot for ballot in pp.ballots if ballot.ranking])
-    if keep_candidates:
-        old_cands = pp.candidates
-        pp_clean = PreferenceProfile(ballots=ballots_nonempty, candidates=old_cands)
-    else:
-        pp_clean = PreferenceProfile(ballots=ballots_nonempty)
-    return pp_clean
+@overload
+def clean_profile(
+    profile: PreferenceProfile,
+    clean_ballot_func: Callable[[Ballot], Ballot],
+    return_adjusted_count: Literal[False],
+    remove_empty_ballots: bool = True,
+    remove_zero_weight_ballots: bool = True,
+) -> PreferenceProfile: ...
+@overload
+def clean_profile(
+    profile: PreferenceProfile,
+    clean_ballot_func: Callable[[Ballot], Ballot],
+    return_adjusted_count: Literal[True],
+    remove_empty_ballots: bool = True,
+    remove_zero_weight_ballots: bool = True,
+) -> tuple[PreferenceProfile, Fraction]: ...
 
 
 def clean_profile(
-    pp: PreferenceProfile, clean_ballot_func: Callable[[Ballot], Ballot]
-) -> PreferenceProfile:
+    profile: PreferenceProfile,
+    clean_ballot_func: Callable[[Ballot], Ballot],
+    return_adjusted_count: Literal[True] | Literal[False] = False,
+    remove_empty_ballots: bool = True,
+    remove_zero_weight_ballots: bool = True,
+) -> PreferenceProfile | tuple[PreferenceProfile, Fraction]:
     """
     Allows user-defined cleaning rules for PreferenceProfile. Input function
     that applies modification to a single ballot.
 
     Args:
-        pp (PreferenceProfile): A PreferenceProfile to clean.
+        profile (PreferenceProfile): A PreferenceProfile to clean.
         clean_ballot_func (Callable[[Ballot], Ballot]): Function that
             takes a ``Ballot`` and returns a cleaned ``Ballot``.
+        return_adjusted_count (bool, optional): Whether or not to return the number of ballots
+            adjusted. An adjusted ballot is a ballot that was not spoiled by the cleaning rule,
+            but did have some adjustment. Count is computed using original ballot weights.
+            Defaults to False. If True, function returns a tuple (cleaned_profile, count).
+        remove_empty_ballots (bool, optional): Whether or not to remove ballots that have no
+            ranking or scores as a result of cleaning. Defaults to True.
+        remove_zero_weight_ballots (bool, optional): Whether or not to remove ballots that have no
+            weight as a result of cleaning. Defaults to True.
 
     Returns:
-        PreferenceProfile: A cleaned ``PreferenceProfile``.
+        Union[PreferenceProfile, tuple[PreferenceProfile, Fraction]]:
+            A cleaned ``PreferenceProfile``. If return_adjusted_count, function returns the profile
[…]
