+++ tests/cleaning/test_remove_repeated_cands.py
@@ -9,20 +9,16 @@ def test_remove_repeated_candidates():
     ballot = Ballot(ranking=[{"A"}, {"A"}, {"B"}, {"C"}], weight=Fraction(1))
     ballot_tuple = (ballot, ballot)
     profile = PreferenceProfile(ballots=ballot_tuple)
-    cleaned_ballot = remove_repeated_candidates(ballot)
-    cleaned_ballot_tuple = remove_repeated_candidates(ballot_tuple)
     cleaned_profile = remove_repeated_candidates(profile)
 
-    assert isinstance(cleaned_ballot, Ballot)
-    assert isinstance(cleaned_ballot_tuple, tuple)
     assert isinstance(cleaned_profile, PreferenceProfile)
 
     true_cleaned_ballot = Ballot(
         ranking=(frozenset({"A"}), frozenset({"B"}), frozenset({"C"}))
     )
-    assert cleaned_ballot == true_cleaned_ballot
-    assert cleaned_ballot_tuple == (true_cleaned_ballot, true_cleaned_ballot)
-    assert cleaned_profile.ballots == cleaned_ballot_tuple
+
+    assert cleaned_profile.ballots[0] == true_cleaned_ballot
+    assert cleaned_profile != profile
 
 
 def test_remove_repeated_candidates_ties():
@@ -52,9 +48,11 @@ def test_remove_repeated_candidates_ties():
 
 def test_remove_repeated_cands_errors():
     with pytest.raises(TypeError, match="Ballot must have rankings:"):
-        remove_repeated_candidates(Ballot())
+        remove_repeated_candidates(PreferenceProfile(ballots=(Ballot(),)))
 
     with pytest.raises(TypeError, match="Ballot must only have rankings, not scores:"):
         remove_repeated_candidates(
-            Ballot(ranking=(frozenset({"Chris"}),), scores={"Chris": 1})
+            PreferenceProfile(
+                ballots=(Ballot(ranking=(frozenset({"Chris"}),), scores={"Chris": 1}),)
+            )
         )
