### 2025-03-21 f5bca51
adjusted 0 vote candidate defaults in utils
