### 2025-03-21 052a316
add check for sufficient cast candidates in elections
