+++ src/votekit/elections/election_types/scores/rating.py
@@ -55,12 +55,21 @@ class GeneralRating(Election):
     def _validate_profile(self, profile: PreferenceProfile):
         """
         Ensures that every ballot has a score dictionary and each voter has not gone over their
-        score limit per candidate and total budgrt. Raises a TypeError if no score dictionary,
+        score limit per candidate and total budget. Raises a TypeError if no score dictionary,
         and a value error for budget/score limit violation.
 
         Args:
             profile (PreferenceProfile): Profile to validate.
+
+        Raises:
+            TypeError: no score dictionary in a ballot.
+            TypeError: Ballot must have non-negative scores.
+            ValueError: Ballot violates score limit per candidate.
+            ValueError: Ballot violates score budget.
+            ValueError: Not enough candidates received votes to be elected.
         """
+        if len(profile.candidates_cast) < self.m:
+            raise ValueError("Not enough candidates received votes to be elected.")
 
         for b in profile.ballots:
             if not b.scores:
@@ -105,7 +114,9 @@ class GeneralRating(Election):
             prev_state.remaining, self.m, profile=profile, tiebreak=self.tiebreak
         )
 
-        new_profile = remove_cand([c for s in elected for c in s], profile)
+        new_profile = remove_cand(
+            [c for s in elected for c in s], profile, return_adjusted_count=False
+        )
 
         if store_states:
             if self.score_function:
