+++ src/votekit/elections/election_types/ranking/alaska.py
@@ -141,11 +141,14 @@ class Alaska(RankingElection):
         """
         if prev_state.round_number == 0:
             plurality = Plurality(profile, self.m_1, self.tiebreak)
-
             remaining = plurality.get_elected()
             eliminated = plurality.get_remaining()
             tiebreaks = plurality.election_states[-1].tiebreaks
-            new_profile = remove_cand([c for s in eliminated for c in s], profile)
+            new_profile = remove_cand(
+                [c for s in eliminated for c in s],
+                profile,
+                return_adjusted_count=False,
+            )
             if self.score_function:
                 scores = self.score_function(new_profile)
 
