+++ src/votekit/elections/election_types/ranking/condo_borda.py
@@ -18,6 +18,8 @@ class CondoBorda(RankingElection):
     """
 
     def __init__(self, profile: PreferenceProfile, m: int = 1):
+        if len(profile.candidates_cast) < m:
+            raise ValueError("Not enough candidates received votes to be elected.")
         self.m = m
         super().__init__(profile, score_function=borda_scores)
 
@@ -58,8 +60,10 @@ class CondoBorda(RankingElection):
         else:
             tiebreaks = {}
 
-        new_profile = remove_cand([c for s in elected for c in s], profile)
-
+        new_profile = remove_cand(
+            [c for s in elected for c in s], profile, return_adjusted_count=False
+        )
+        print("max ballot length of new profile", new_profile.max_ballot_length)
         if store_states:
             self.election_states.append(
                 ElectionState(

[…]
