+++ tests/elections/election_types/score/test_limited.py
@@ -134,7 +134,7 @@ def test_errors():
         Limited(profile_no_tied_limited, m=0, k=0)
 
     with pytest.raises(
-        ValueError, match="m must be no more than the number of candidates."
+        ValueError, match="Not enough candidates received votes to be elected."
     ):
         Limited(profile_no_tied_limited, m=5, k=2)
 
@@ -150,7 +150,7 @@ def test_errors():
 
 def test_validate_profile():
     with pytest.raises(TypeError, match="violates score limit"):
-        profile = PreferenceProfile(ballots=[Ballot(scores={"A": 3})])
+        profile = PreferenceProfile(ballots=[Ballot(scores={"A": 3, "B": 2})])
         Limited(profile, m=2, k=2)
 
     with pytest.raises(TypeError, match="violates total score budget"):
@@ -162,5 +162,5 @@ def test_validate_profile():
         Limited(profile, m=1)
 
     with pytest.raises(TypeError, match="All ballots must have score dictionary."):
-        profile = PreferenceProfile(ballots=[Ballot()])
+        profile = PreferenceProfile(ballots=[Ballot(), Ballot(scores={"A": 1, "B": 1})])
         Limited(profile, m=2)
