### 2025-03-24 cd8f111
add adjusted count tests
