### 2025-03-24 ea41302
change voter_set default to empty set
