### 2025-03-25 0bfe569
initial draft of cleanprofile
