+++ tests/cleaning/test_remove_cand.py
@@ -1,7 +1,6 @@
-from votekit.pref_profile import PreferenceProfile
+from votekit.pref_profile import PreferenceProfile, CleanedProfile
 from votekit.ballot import Ballot
 from votekit.cleaning import remove_cand
-from fractions import Fraction
 
 profile_no_ties = PreferenceProfile(
     ballots=[
@@ -20,127 +19,123 @@ profile_with_ties = PreferenceProfile(
 )
 
 
-def test_remove_cand_dif_types():
-    no_a_true = PreferenceProfile(
-        ballots=[
-            Ballot(ranking=[{"B"}], weight=1),
-            Ballot(ranking=[{"B"}, {"C"}], weight=1 / 2),
-            Ballot(ranking=[{"C"}, {"B"}], weight=3),
-        ],
-        candidates=("B", "C"),
+def test_remove_cand():
+    cleaned_profile = remove_cand("A", profile_no_ties)
+
+    assert isinstance(cleaned_profile, CleanedProfile)
+    assert cleaned_profile.parent_profile == profile_no_ties
+    assert cleaned_profile.ballots == (
+        Ballot(ranking=[{"B"}], weight=1),
+        Ballot(ranking=[{"B"}, {"C"}], weight=1 / 2),
+        Ballot(ranking=[{"C"}, {"B"}], weight=3),
     )
+    assert cleaned_profile != profile_no_ties
+    assert cleaned_profile.no_weight_alt_ballot_indices == []
+    assert cleaned_profile.no_ranking_and_no_scores_alt_ballot_indices == []
+    assert cleaned_profile.valid_but_alt_ballot_indices == [0, 1, 2]
+    assert cleaned_profile.unalt_ballot_indices == []
 
-    assert remove_cand("A", profile_no_ties) == no_a_true
-    assert profile_no_ties != no_a_true
 
+def test_remove_mult_cands():
+    cleaned_profile = remove_cand(["A", "B"], profile_no_ties)
 
-def test_remove_cand_no_ties():
-    no_a = remove_cand("A", profile_no_ties)
-    no_a_true = PreferenceProfile(
-        ballots=[
-            Ballot(ranking=[{"B"}], weight=1),
-            Ballot(ranking=[{"B"}, {"C"}], weight=1 / 2),
-            Ballot(ranking=[{"C"}, {"B"}], weight=3),
-        ],
-        candidates=("B", "C"),
+    assert isinstance(cleaned_profile, CleanedProfile)
+    assert cleaned_profile.parent_profile == profile_no_ties
+
+    assert cleaned_profile.group_ballots().ballots == (
+        Ballot(ranking=[{"C"}], weight=7 / 2),
     )
-
-    no_a_b = remove_cand(["A", "B"], profile_no_ties)
-    no_a_b_true = PreferenceProfile(
-        ballots=[
-            Ballot(ranking=[{"C"}], weight=7 / 2),
-        ],
-        candidates=("C",),
-    )
-
-    assert no_a == no_a_true
-    assert no_a_b == no_a_b_true
+    assert cleaned_profile != profile_no_ties
+    assert cleaned_profile.no_weight_alt_ballot_indices == []
+    assert cleaned_profile.no_ranking_and_no_scores_alt_ballot_indices == [0]
+    assert cleaned_profile.valid_but_alt_ballot_indices == [1, 2]
+    assert cleaned_profile.unalt_ballot_indices == []
 
 
-def test_remove_cand_with_ties():
-    no_a = remove_cand("A", profile_with_ties)
-    no_a_true = PreferenceProfile(
-        ballots=[
-            Ballot(ranking=[{"B"}], weight=1),
[…]

[…]
