### 2025-03-25 f6f1b7a
edit ballot equality check
