### 2025-03-25 986bd19
refactor cleaning tests
