### 2025-03-25 3b5bf7c
refactor elections to use CleanedProfile
