+++ src/votekit/elections/election_types/ranking/top_two.py
@@ -69,8 +69,8 @@ class TopTwo(RankingElection):
             remaining = plurality.get_elected()
             eliminated = plurality.get_remaining()
             tiebreaks = plurality.election_states[-1].tiebreaks
-            new_profile = remove_cand(
-                [c for s in eliminated for c in s], profile, return_adjusted_count=False
+            new_profile: PreferenceProfile = remove_cand(
+                [c for s in eliminated for c in s], profile
             )
             if self.score_function:
                 scores = self.score_function(new_profile)
