### 2025-03-28 f0321de
cleaning functions no longer condense ballots
