### 2025-03-28 452e0c2
add condense_profile function
