### 2025-03-28 0266ad8
adding condense step to remove cand in codebase
