+++ tests/graphs/pairwise_comparison/test_pwc_graph.py
@@ -6,7 +6,7 @@ from votekit.ballot import Ballot
 from votekit.graphs.pairwise_comparison_graph import PairwiseComparisonGraph
 from votekit.pref_profile import PreferenceProfile
 from votekit.cvr_loaders import load_csv
-from votekit.cleaning import remove_cand
+from votekit.cleaning import remove_cand, condense_profile
 
 from matplotlib.axes import Axes
 import pytest
@@ -15,9 +15,11 @@ from pathlib import Path
 
 BASE_DIR = Path(__file__).resolve().parents[2]
 CSV_DIR = BASE_DIR / "data/csv/"
-portland_profile = remove_cand(
-    "skipped",
-    load_csv(CSV_DIR / "Portland_D3_Condensed.csv", rank_cols=[1, 2, 3, 4, 5, 6]),
+portland_profile = condense_profile(
+    remove_cand(
+        "skipped",
+        load_csv(CSV_DIR / "Portland_D3_Condensed.csv", rank_cols=[1, 2, 3, 4, 5, 6]),
+    )
 )
 
 
