+++ src/votekit/elections/election_types/ranking/random_dictator.py
@@ -1,7 +1,7 @@
 from .abstract_ranking import RankingElection
 from ....pref_profile import PreferenceProfile
 from ...election_state import ElectionState
-from ....cleaning import remove_cand
+from ....cleaning import remove_cand, condense_profile
 from ....utils import (
     first_place_votes,
     score_dict_to_ranking,
@@ -74,7 +74,7 @@ class RandomDictator(RankingElection):
         winning_cand = random.choices(candidates, weights=weights, k=1)[0]
         elected = (frozenset({winning_cand}),)
 
-        new_profile = remove_cand(winning_cand, profile)
+        new_profile = condense_profile(remove_cand(winning_cand, profile))
 
         if store_states:
             if self.score_function:
