+++ src/votekit/pref_profile/pref_profile.py
@@ -3,11 +3,11 @@ import csv
 from fractions import Fraction
 import pandas as pd
 from pydantic import ConfigDict, field_validator, model_validator
-from typing import Optional
-from .ballot import Ballot
+from ..ballot import Ballot
 from pydantic.dataclasses import dataclass
 from typing_extensions import Self
 from dataclasses import field
+import numpy as np
 
 
 @dataclass(frozen=True, config=ConfigDict(arbitrary_types_allowed=True))
@@ -46,57 +46,142 @@ class PreferenceProfile:
     num_ballots: int = 0
     total_ballot_wt: Fraction = Fraction(0)
 
+    # to and from csv, different type options for scoring, ranking, both
+    # encoding specification
+
+    @model_validator(mode="after")
+    def ballot_list_to_df(self) -> Self:
+        """
+        After init, convert ballot list to df.
+        """
+
+        num_ballots = len(self.ballots)
+
+        ballot_data = {
+            "weight": [np.nan] * num_ballots,
+            "id": [np.nan] * num_ballots,
+            "voter_set": [np.nan] * num_ballots,
+        }
+
+        if self.candidates:
+            ballot_data.update({c: [np.nan] * num_ballots for c in self.candidates})
+
+        if self.max_ballot_length:
+            ballot_data.update(
+                {
+                    f"ranking_{i}": [np.nan] * num_ballots
+                    for i in range(self.max_ballot_length)
+                }
+            )
+
+        for i, b in enumerate(self.ballots):
+            ballot_data["weight"][i] = b.weight
+
+            if b.id:
+                ballot_data["id"][i] = b.id
+            if b.voter_set:
+                ballot_data["voter_set"][i] = b.voter_set
+
+            if b.scores:
+                for c, score in b.scores.items():
+                    if c not in ballot_data:
+                        if self.candidates:
+                            raise ValueError(
+                                f"Candidate {c} found in ballot {b} but not in candidate list {self.candidates}."
+                            )
+                        ballot_data[c] = [np.nan] * num_ballots
+                    ballot_data[c][i] = score
+
+            if b.ranking:
+                for j, cand_set in enumerate(b.ranking):
+                    if f"ranking_{j}" not in ballot_data:
+                        if self.max_ballot_length:
+                            raise ValueError(
+                                f"Max ballot length {self.max_ballot_length} given but ballot {b} has length at least {j}."
+                            )
+                        ballot_data[f"ranking_{j}"] = [np.nan] * num_ballots
+
+                    ballot_data[f"ranking_{j}"][i] = cand_set
+
+        df = pd.DataFrame(ballot_data)
+        temp_col_order = [c for c in df.columns if "ranking" in c] + [
+            "weight",
+            "id",
+            "voter_set",
+        ]
+        col_order = (
+            self.candidates + temp_col_order
+            if self.candidates
+            else sorted([c for c in df.columns if c not in temp_col_order])
+            + temp_col_order
+        )
+        df = df[col_order]
+        df.index.name = "Ballot Index"
+
+        object.__setattr__(self, "df", df)
+        return self
+
     @field_validator("candidates")
     @classmethod
     def cands_must_be_unique(
         cls, candidates: Optional[tuple[str, ...]]
     ) -> Optional[tuple[str, ...]]:
-        if candidates:
-            if not len(set(candidates)) == len(candidates):
-                raise ValueError("All candidates must be unique.")
+        # TODO think about how to do this with DF operations
+        # or how to validate unique candidates given a ballot list or df
+
+        # if candidates:
+        #     if not len(set(candidates)) == len(candidates):
+        #         raise ValueError("All candidates must be unique.")
         return candidates
 
     @model_validator(mode="after")
[…]
