+++ src/votekit/pref_profile/utils.py
@@ -1,0 +1,209 @@
+from votekit.pref_profile import PreferenceProfile
+import pandas as pd
+from votekit.ballot import Ballot
+from functools import partial
+from typing import Optional
+import numpy as np
+
+
+def _convert_ranking_cols_to_ranking(
+    row: pd.Series, ranking_cols: list[str]
+) -> Optional[tuple[frozenset, ...]]:
+    """
+    Convert the ranking cols to a ranking tuple.
+
+    """
+    ranking = []
+
+    for i, col in enumerate(ranking_cols):
+        if pd.isna(row[col]):
+            if not all(pd.isna(row[c]) for c in ranking_cols[i:]):
+                raise ValueError(
+                    (
+                        f"Row {row} has NaN values between valid ranking positions."
+                        " NaN values can only trail on a ranking."
+                    )
+                )
+
+            break
+
+        ranking.append(row[col])
+
+    return tuple(ranking) if ranking else None
+
+
+def _convert_row_to_ballot(
+    row: pd.Series,
+    ranking_cols: list[str],
+    weight_col: str,
+    id_col: str,
+    voter_set_col: str,
+    candidates: list[str],
+) -> Ballot:
+    """
+    Convert a row of a properly formatted df to a Ballot.
+    """
+
+    ranking = _convert_ranking_cols_to_ranking(row, ranking_cols)
+    scores = {c: row[c] if not pd.isna(row[c]) else 0 for c in candidates}
+    id = row[id_col] if not pd.isna(row[id_col]) else None
+    voter_set = row[voter_set_col] if not pd.isna(row[voter_set_col]) else set()
+    weight = row[weight_col]
+
+    return Ballot(
+        ranking=ranking,
+        scores=scores,
+        weight=weight,
+        id=id if id else None,
+        voter_set=voter_set if voter_set else set(),
+    )
+
+
+def profile_to_df(profile: PreferenceProfile) -> pd.DataFrame:
+    """
+    Convert a PrefProfile into a pandas dataframe.
+
+    """
+    ballot_data = {
+        c: [np.nan] * profile.num_ballots for c in sorted(profile.candidates)
+    }
+    ballot_data.update(
+        {
+            f"ranking_{i}": [np.nan] * profile.num_ballots
+            for i in range(profile.max_ballot_length)
+        }
+    )
+    ballot_data.update({"weight": [np.nan] * profile.num_ballots})
+    ballot_data.update({"id": [np.nan] * profile.num_ballots})
+    ballot_data.update({"voter_set": [np.nan] * profile.num_ballots})
+
+    for i, b in enumerate(profile.ballots):
+        ballot_data["weight"][i] = b.weight
+
+        if b.id:
+            ballot_data["id"][i] = b.id
+        if b.voter_set:
+            ballot_data["voter_set"][i] = b.voter_set
+
+        if b.scores:
+            for c, score in b.scores.items():
+                ballot_data[c][i] = score
+
+        if b.ranking:
+            for j, cand_set in enumerate(b.ranking):
+                ballot_data[f"ranking_{j}"][i] = cand_set
+
+    df = pd.DataFrame(ballot_data)
+    df.index.name = "Ballot Index"
+    return df
+
+
+def ballot_list_to_df(
+    ballots: list[Ballot], max_ballot_length: int = 0, candidates: list[str] = []
+) -> pd.DataFrame:
+    """
+    Convert a list of ballots into a pandas dataframe.
+
+    """
+    num_ballots = len(ballots)
+
+    ballot_data = {
+        "weight": [np.nan] * num_ballots,
+        "id": [np.nan] * num_ballots,
+        "voter_set": [np.nan] * num_ballots,
+    }
+
[…]
