+++ .pre-commit-config.yaml
@@ -6,14 +6,14 @@ repos:
     hooks:
       - id: check-merge-conflict
       - id: debug-statements
-  - repo: https://github.com/charliermarsh/ruff-pre-commit
-    rev: "v0.0.257"
-    hooks:
-      - id: ruff
-  - repo: https://github.com/ambv/black
-    rev: 24.8.0
-    hooks:
-      - id: black
+  # - repo: https://github.com/charliermarsh/ruff-pre-commit
+  #   rev: "v0.0.257"
+  #   hooks:
+  #     - id: ruff
+  # - repo: https://github.com/ambv/black
+  #   rev: 24.8.0
+  #   hooks:
+  #     - id: black
   # - repo: https://github.com/pre-commit/mirrors-mypy
   #   rev: "v1.1.1"
   #   hooks:
