### 2025-03-31 bf50909
add df tests
