+++ src/votekit/pref_profile.py
@@ -1,0 +1,558 @@
+from __future__ import annotations
+import csv
+from fractions import Fraction
+import pandas as pd
+from pydantic import ConfigDict, field_validator, model_validator
+from .ballot import Ballot
+from pydantic.dataclasses import dataclass
+from typing_extensions import Self
+from dataclasses import field
+import numpy as np
+from typing import Optional
+
+
+@dataclass(frozen=True, config=ConfigDict(arbitrary_types_allowed=True))
+class PreferenceProfile:
+    """
+    PreferenceProfile class, contains ballots and candidates for a given election.
+    This is a frozen class, so you need to create a new PreferenceProfile any time
+    you want to edit the ballots, candidates, etc.
+
+    Args:
+        ballots (tuple[Ballot], optional): Tuple of ``Ballot`` objects. Defaults to empty tuple.
+        candidates (tuple[str], optional): Tuple of candidate strings. Defaults to empty tuple.
+            If empty, computes this from any candidate listed on a ballot with positive weight.
+        max_ballot_length (int, optional): The length of the longest allowable ballot, i.e., how
+            many candidates are allowed to be ranked in an election. Defaults to longest observed
+            ballot.
+        contains_rankings (bool, optional): Whether or not the profile contains ballots with
+            rankings. If not provided, will set itself to correct value given the ballot list.
+        contains_scores (bool, optional): Whether or not the profile contains ballots with
+            scores. If not provided, will set itself to correct value given the ballot list.
+        contains_rankings_and_scores (bool): Whether or not the profile contains ballots with
+            rankings and scores. If not provided, will set itself to correct value given the ballot
+            list.
+
+    Parameters:
+        ballots (tuple[Ballot]): Tuple of ``Ballot`` objects.
+        candidates (tuple[str]): Tuple of candidate strings.
+        max_ballot_length (int): The length of the longest allowable ballot, i.e., how
+            many candidates are allowed to be ranked in an election.
+        df (pandas.DataFrame): Data frame view of the ballots.
+        candidates_cast (tuple[str]): Tuple of candidates who appear on any ballot with positive
+            weight, either in the ranking or in the score dictionary.
+        total_ballot_wt (Fraction): Sum of ballot weights.
+        num_ballots (int): Length of ballot list.
+        contains_rankings (bool): Whether or not the profile contains ballots with
+            rankings.
+        contains_scores (bool): Whether or not the profile contains ballots with
+            scores.
+        contains_rankings_and_scores (bool): Whether or not the profile contains ballots with
+            rankings and scores.
+
+    """
+
+    ballots: tuple[Ballot, ...] = field(default_factory=tuple)
+    candidates: tuple[str, ...] = field(default_factory=tuple)
+    max_ballot_length: int = 0
+    df: pd.DataFrame = field(default_factory=pd.DataFrame)
+    candidates_cast: tuple[str, ...] = field(default_factory=tuple)
+    num_ballots: int = 0
+    total_ballot_wt: Fraction = Fraction(0)
+    contains_rankings: Optional[bool] = None
+    contains_scores: Optional[bool] = None
+    contains_rankings_and_scores: Optional[bool] = None
+
+    @field_validator("candidates")
+    @classmethod
+    def cands_must_be_unique(
+        cls, candidates: Optional[tuple[str, ...]]
+    ) -> Optional[tuple[str, ...]]:
+        if candidates:
+            if not len(set(candidates)) == len(candidates):
+                raise ValueError("All candidates must be unique.")
+        return candidates
+
+    @model_validator(mode="after")
+    def ballot_list_to_df(self) -> Self:
+        num_ballots = len(self.ballots)
+
+        ballot_data = {
[…]
