+++ src/votekit/pref_profile/pref_profile.py
@@ -1,543 +1,0 @@
-from __future__ import annotations
-import csv
-from fractions import Fraction
-import pandas as pd
-from pydantic import ConfigDict, field_validator, model_validator
-from ..ballot import Ballot
-from pydantic.dataclasses import dataclass
-from typing_extensions import Self
-from dataclasses import field
-import numpy as np
-
-
-@dataclass(frozen=True, config=ConfigDict(arbitrary_types_allowed=True))
-class PreferenceProfile:
-    """
-    PreferenceProfile class, contains ballots and candidates for a given election.
-    This is a frozen class, so you need to create a new PreferenceProfile any time
-    you want to edit the ballots, candidates, etc.
-
-    Args:
-        ballots (tuple[Ballot], optional): Tuple of ``Ballot`` objects. Defaults to empty tuple.
-        candidates (tuple[str], optional): Tuple of candidate strings. Defaults to empty tuple.
-            If empty, computes this from any candidate listed on a ballot with positive weight.
-        max_ballot_length (int, optional): The length of the longest allowable ballot, i.e., how
-            many candidates are allowed to be ranked in an election. Defaults to longest observed
-            ballot.
-
-    Parameters:
-        ballots (tuple[Ballot]): Tuple of ``Ballot`` objects.
-        candidates (tuple[str]): Tuple of candidate strings.
-        max_ballot_length (int): The length of the longest allowable ballot, i.e., how
-            many candidates are allowed to be ranked in an election.
-        df (pandas.DataFrame): Data frame view of the ballots.
-        candidates_cast (tuple[str]): Tuple of candidates who appear on any ballot with positive
-            weight, either in the ranking or in the score dictionary.
-        total_ballot_wt (Fraction): Sum of ballot weights.
-        num_ballots (int): Length of ballot list.
-
-    """
-
-    ballots: tuple[Ballot, ...] = field(default_factory=tuple)
-    candidates: tuple[str, ...] = field(default_factory=tuple)
-    max_ballot_length: int = 0
-    df: pd.DataFrame = field(default_factory=pd.DataFrame)
-    candidates_cast: tuple[str, ...] = field(default_factory=tuple)
-    num_ballots: int = 0
-    total_ballot_wt: Fraction = Fraction(0)
-
-    # to and from csv, different type options for scoring, ranking, both
-    # encoding specification
-
-    @model_validator(mode="after")
-    def ballot_list_to_df(self) -> Self:
-        """
-        After init, convert ballot list to df.
-        """
-
-        num_ballots = len(self.ballots)
-
-        ballot_data = {
-            "weight": [np.nan] * num_ballots,
-            "id": [np.nan] * num_ballots,
-            "voter_set": [np.nan] * num_ballots,
-        }
-
-        if self.candidates:
-            ballot_data.update({c: [np.nan] * num_ballots for c in self.candidates})
-
-        if self.max_ballot_length:
-            ballot_data.update(
-                {
-                    f"ranking_{i}": [np.nan] * num_ballots
-                    for i in range(self.max_ballot_length)
-                }
-            )
-
-        for i, b in enumerate(self.ballots):
-            ballot_data["weight"][i] = b.weight
-
-            if b.id:
-                ballot_data["id"][i] = b.id
-            if b.voter_set:
-                ballot_data["voter_set"][i] = b.voter_set
-
-            if b.scores:
-                for c, score in b.scores.items():
-                    if c not in ballot_data:
-                        if self.candidates:
-                            raise ValueError(
-                                f"Candidate {c} found in ballot {b} but not in candidate list {self.candidates}."
-                            )
-                        ballot_data[c] = [np.nan] * num_ballots
-                    ballot_data[c][i] = score
-
-            if b.ranking:
-                for j, cand_set in enumerate(b.ranking):
[…]

[…]
