### 2025-04-01 756e9d3
add tests for profile df
