### 2025-04-01 70f8170
add __add__ tests
