+++ tests/pref_profile/df/test_pp_df_ranking.py
@@ -12,7 +12,6 @@ ballots_rankings = [
 ]
 
 
-
 def test_pp_df_rankings():
     pp = PreferenceProfile(ballots=ballots_rankings)
     data = {
