### 2025-04-01 671e6de
assorted profile method tests
