### 2025-04-01 009f15d
convert df to ballot list for grouping
