+++ src/votekit/pref_profile.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
-import csv
+
+# import csv
 from fractions import Fraction
 import pandas as pd
 from pydantic import ConfigDict, field_validator, model_validator
@@ -9,6 +10,83 @@ from typing_extensions import Self
 from dataclasses import field
 import numpy as np
 from typing import Optional
+from functools import partial
+import warnings
+
+
+def _convert_ranking_cols_to_ranking(
+    row: pd.Series, ranking_cols: list[str]
+) -> Optional[tuple[frozenset, ...]]:
+    """
+    Convert the ranking cols to a ranking tuple.
+
+    """
+    ranking = []
+    for i, col in enumerate(ranking_cols):
+        if pd.isna(row[col]):
+            if not all(pd.isna(row[c]) for c in ranking_cols[i:]):
+                raise ValueError(
+                    f"Row {row} has NaN values between valid ranking positions. "
+                    "NaN values can only trail on a ranking."
+                )
+
+            break
+
+        ranking.append(row[col])
+
+    return tuple(ranking) if ranking else None
+
+
+def _convert_row_to_ballot(
+    row: pd.Series,
+    ranking_cols: list[str],
+    weight_col: str,
+    id_col: str,
+    voter_set_col: str,
+    candidates: list[str],
+) -> Ballot:
+    """
+    Convert a row of a properly formatted df to a Ballot.
+    """
+    ranking = _convert_ranking_cols_to_ranking(row, ranking_cols)
+    scores = {c: row[c] for c in candidates if c in row and not pd.isna(row[c])}
+    id = row[id_col] if not pd.isna(row[id_col]) else None
+    voter_set = row[voter_set_col]
+    weight = row[weight_col]
+
+    return Ballot(
+        ranking=ranking,
+        scores=scores if scores else None,
+        weight=weight,
+        id=id,
+        voter_set=voter_set,
+    )
+
+
+def _df_to_ballot_tuple(
+    df: pd.DataFrame,
+    candidates: list[str],
+    ranking_cols: list[str] = [],
+    weight_col: str = "weight",
+    id_col: str = "id",
+    voter_set_col: str = "voter_set",
+) -> tuple[Ballot]:
+    """
+    Convert a df into a list of ballots.
+    """
+    return tuple(
+        df.apply(
+            partial(
+                _convert_row_to_ballot,
+                ranking_cols=ranking_cols,
+                weight_col=weight_col,
+                id_col=id_col,
+                candidates=candidates,
+                voter_set_col=voter_set_col,
+            ),
+            axis="columns",
+        )
+    )
 
 
 @dataclass(frozen=True, config=ConfigDict(arbitrary_types_allowed=True))
@@ -74,7 +152,7 @@ class PreferenceProfile:
         ballot_data = {
             "weight": [np.nan] * num_ballots,
             "id": [np.nan] * num_ballots,
-            "voter_set": [np.nan] * num_ballots,
+            "voter_set": [set()] * num_ballots,
         }
 
         if self.candidates:
@@ -101,7 +179,7 @@ class PreferenceProfile:
                 ballot_data["voter_set"][i] = b.voter_set
 
             if b.scores:
-                if self.contains_scores == False:
+                if self.contains_scores is False:
                     raise ValueError(
                         (
                             f"Ballot {b} has scores {b.scores} but contains_scores is "
@@ -117,13 +195,14 @@ class PreferenceProfile:
                     if c not in ballot_data:
                         if self.candidates:
                             raise ValueError(
-                                f"Candidate {c} found in ballot {b} but not in candidate list {self.candidates}."
+                                f"Candidate {c} found in ballot {b} but not in "
+                                f"candidate list {self.candidates}."
[…]
