### 2025-04-03 953463c
add csv method and tests
