### 2025-04-03 c24139c
add pickle methods and tests
