+++ src/votekit/pref_profile.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
-
-# import csv
+import csv
 from fractions import Fraction
 import pandas as pd
 from pydantic import ConfigDict, field_validator, model_validator
@@ -12,7 +11,7 @@ import numpy as np
 from typing import Optional
 from functools import partial
 import warnings
-
+import pickle
 
 def _convert_ranking_cols_to_ranking(
     row: pd.Series, ranking_cols: list[str]
@@ -67,9 +66,9 @@ def _df_to_ballot_tuple(
     df: pd.DataFrame,
     candidates: list[str],
     ranking_cols: list[str] = [],
-    weight_col: str = "weight",
-    id_col: str = "id",
-    voter_set_col: str = "voter_set",
+    weight_col: str = "Weight",
+    id_col: str = "ID",
+    voter_set_col: str = "Voter Set",
 ) -> tuple[Ballot]:
     """
     Convert a df into a list of ballots.
@@ -150,9 +149,9 @@ class PreferenceProfile:
         num_ballots = len(self.ballots)
 
         ballot_data = {
-            "weight": [np.nan] * num_ballots,
-            "id": [np.nan] * num_ballots,
-            "voter_set": [set()] * num_ballots,
+            "Weight": [np.nan] * num_ballots,
+            "ID": [np.nan] * num_ballots,
+            "Voter Set": [set()] * num_ballots,
         }
 
         if self.candidates:
@@ -161,7 +160,7 @@ class PreferenceProfile:
         if self.max_ballot_length:
             ballot_data.update(
                 {
-                    f"ranking_{i+1}": [np.nan] * num_ballots
+                    f"Ranking_{i+1}": [np.nan] * num_ballots
                     for i in range(self.max_ballot_length)
                 }
             )
@@ -171,12 +170,12 @@ class PreferenceProfile:
         contains_scores_indicator = False
 
         for i, b in enumerate(self.ballots):
-            ballot_data["weight"][i] = b.weight
+            ballot_data["Weight"][i] = b.weight
 
             if b.id:
-                ballot_data["id"][i] = b.id
+                ballot_data["ID"][i] = b.id
             if b.voter_set:
-                ballot_data["voter_set"][i] = b.voter_set
+                ballot_data["Voter Set"][i] = b.voter_set
 
             if b.scores:
                 if self.contains_scores is False:
@@ -221,21 +220,21 @@ class PreferenceProfile:
                                 )
                         if b.weight > 0 and c not in candidates_cast:
                             candidates_cast.append(c)
-                    if f"ranking_{j+1}" not in ballot_data:
+                    if f"Ranking_{j+1}" not in ballot_data:
                         if self.max_ballot_length:
                             raise ValueError(
                                 f"Max ballot length {self.max_ballot_length} given but "
                                 "ballot {b} has length at least {j+1}."
                             )
-                        ballot_data[f"ranking_{j+1}"] = [np.nan] * num_ballots
+                        ballot_data[f"Ranking_{j+1}"] = [np.nan] * num_ballots
 
-                    ballot_data[f"ranking_{j+1}"][i] = cand_set
+                    ballot_data[f"Ranking_{j+1}"][i] = cand_set
 
         df = pd.DataFrame(ballot_data)
-        temp_col_order = [c for c in df.columns if "ranking" in c] + [
-            "weight",
-            "id",
-            "voter_set",
+        temp_col_order = [c for c in df.columns if "Ranking" in c] + [
+            "Weight",
+            "ID",
+            "Voter Set",
         ]
 
         if self.candidates and contains_scores_indicator:
@@ -255,9 +254,6 @@ class PreferenceProfile:
         if not self.candidates:
[…]
