+++ tests/pref_profile/test_pref_profile.py
@@ -10,9 +10,6 @@ def test_init():
     empty_profile = PreferenceProfile()
     assert empty_profile.ballots == ()
     assert not empty_profile.candidates
-    assert empty_profile.df.equals(
-        pd.DataFrame({"Ranking": [], "Scores": [], "Weight": [], "Percent": []})
-    )
     assert not empty_profile.candidates_cast
     assert not empty_profile.total_ballot_wt
     assert not empty_profile.num_ballots
@@ -26,7 +23,7 @@ def test_unique_cands_validator():
     PreferenceProfile(candidates=("A", "B"))
 
 
-def test_ballots():
+def test_ballots_frozen():
     p = PreferenceProfile(ballots=[Ballot()])
     b_list = p.ballots
 
@@ -38,7 +35,7 @@ def test_ballots():
         p.ballots = (Ballot(weight=5),)
 
 
-def test_candidates():
+def test_candidates_frozen():
     profile_no_cands = PreferenceProfile(
         ballots=[Ballot(ranking=[{"A"}, {"B"}]), Ballot(ranking=[{"C"}, {"B"}])]
     )
@@ -73,273 +70,118 @@ def test_get_candidates_received_votes():
     assert set(vote_cands) == {"A", "B", "C", "E"}
 
 
-def test_ballot_length_default():
-    profile = PreferenceProfile(
-        ballots=(
-            Ballot(ranking=({"A"}, {"B"}, {"C", "D"})),
-            Ballot(ranking=({"A"}, {"B"}), weight=Fraction(3, 2)),
-            Ballot(ranking=({"C"}, {"B"}), scores={"A": 4}, weight=2),
-            Ballot(scores={"A": 4}),
-        )
-    )
-
-    assert profile.max_ballot_length == 3
 
 
-def test_ballot_length_no_default():
-    profile = PreferenceProfile(
-        ballots=(
-            Ballot(ranking=({"A"}, {"B"}, {"C", "D"})),
-            Ballot(ranking=({"A"}, {"B"}), weight=Fraction(3, 2)),
-            Ballot(ranking=({"C"}, {"B"}), scores={"A": 4}, weight=2),
-            Ballot(scores={"A": 4}),
-        ),
-        max_ballot_length=4,
-    )
 
-    assert profile.max_ballot_length == 4
+# def test_df_head():
+#     profile = PreferenceProfile(
+#         ballots=(
+#             Ballot(ranking=({"A"}, {"B"}, {"C"}), weight=Fraction(1), scores={"A": 4}),
+#             Ballot(ranking=({"B"}, {"C"}, {"E"}), weight=Fraction(2)),
+#             Ballot(ranking=({"B", "C"}, {"E"}), weight=Fraction(1)),
+#         )
+#     )
 
-    with pytest.raises(
-        ValueError, match="Profile contains a ballot with a ranking that is too long."
-    ):
-        profile = PreferenceProfile(
-            ballots=(
-                Ballot(ranking=({"A"}, {"B"}, {"C", "D"})),
-                Ballot(ranking=({"A"}, {"B"}), weight=Fraction(3, 2)),
-                Ballot(ranking=({"C"}, {"B"}), scores={"A": 4}, weight=2),
-                Ballot(scores={"A": 4}),
-            ),
-            max_ballot_length=2,
-        )
+#     true_df = pd.DataFrame(
+#         {
+#             "Ranking": [
+#                 ("A", "B", "C"),
+#                 ("B", "C", "E"),
+#                 (f"{set({'B', 'C'})} (Tie)", "E"),
+#             ],
[…]
