### 2025-04-03 0b2ccc0
add error tests
