+++ src/votekit/pref_profile.py
@@ -99,7 +99,7 @@ class PreferenceProfile:
         ballots (tuple[Ballot], optional): Tuple of ``Ballot`` objects. Defaults to empty tuple.
         candidates (tuple[str], optional): Tuple of candidate strings. Defaults to empty tuple.
             If empty, computes this from any candidate listed on a ballot with positive weight.
-        max_ballot_length (int, optional): The length of the longest allowable ballot, i.e., how
+        max_ranking_length (int, optional): The length of the longest allowable ballot, i.e., how
             many candidates are allowed to be ranked in an election. Defaults to longest observed
             ballot.
         contains_rankings (bool, optional): Whether or not the profile contains ballots with
@@ -110,7 +110,7 @@ class PreferenceProfile:
     Parameters:
         ballots (tuple[Ballot]): Tuple of ``Ballot`` objects.
         candidates (tuple[str]): Tuple of candidate strings.
-        max_ballot_length (int): The length of the longest allowable ballot, i.e., how
+        max_ranking_length (int): The length of the longest allowable ballot, i.e., how
             many candidates are allowed to be ranked in an election.
         df (pandas.DataFrame): Data frame view of the ballots.
         candidates_cast (tuple[str]): Tuple of candidates who appear on any ballot with positive
@@ -122,17 +122,33 @@ class PreferenceProfile:
         contains_scores (bool): Whether or not the profile contains ballots with
             scores.
 
+    Raises:
+        ValueError: contains_rankings is set to False but a ballot contains a ranking.
+        ValueError: contains_rankings is set to True but no ballot contains a ranking.
+        ValueError: contains_scores is set to False but a ballot contains a score.
+        ValueError: contains_scores is set to True but no ballot contains a score.
+        ValueError: max_ranking_length is set but a ballot ranking excedes the length.
+        ValueError: a candidate is found on a ballot that is not listed on a provided 
+            candidate list.
+        ValueError: candidates must be unique.
+     
+    Warns:
+        UserWarning: max_ranking_length is set but contains_rankings is False. 
+            Sets max_ranking_length to 0.
+
     """
 
     ballots: tuple[Ballot, ...] = field(default_factory=tuple)
     candidates: tuple[str, ...] = field(default_factory=tuple)
-    max_ballot_length: int = 0
+    max_ranking_length: int = 0
     df: pd.DataFrame = field(default_factory=pd.DataFrame)
     candidates_cast: tuple[str, ...] = field(default_factory=tuple)
     num_ballots: int = 0
     total_ballot_wt: Fraction = Fraction(0)
     contains_rankings: Optional[bool] = None
     contains_scores: Optional[bool] = None
+
+
 
     @field_validator("candidates")
     @classmethod
@@ -157,11 +173,11 @@ class PreferenceProfile:
         if self.candidates:
             ballot_data.update({c: [np.nan] * num_ballots for c in self.candidates})
 
-        if self.max_ballot_length:
+        if self.max_ranking_length:
             ballot_data.update(
                 {
                     f"Ranking_{i+1}": [np.nan] * num_ballots
-                    for i in range(self.max_ballot_length)
+                    for i in range(self.max_ranking_length)
                 }
             )
 
@@ -221,9 +237,9 @@ class PreferenceProfile:
                         if b.weight > 0 and c not in candidates_cast:
                             candidates_cast.append(c)
                     if f"Ranking_{j+1}" not in ballot_data:
-                        if self.max_ballot_length:
+                        if self.max_ranking_length:
                             raise ValueError(
-                                f"Max ballot length {self.max_ballot_length} given but "
+                                f"Max ballot length {self.max_ranking_length} given but "
                                 "ballot {b} has length at least {j+1}."
                             )
[…]
