+++ src/votekit/utils.py
@@ -294,14 +294,14 @@ def score_profile_from_rankings(
     """
     Score the candidates based on a score vector. For example, the vector (1,0,...) would
     return the first place votes for each candidate. Vectors should be non-increasing and
-    non-negative. Vector should be as long as ``max_ballot_length`` in the profile.
+    non-negative. Vector should be as long as ``max_ranking_length`` in the profile.
     If it is shorter, we add 0s. Unlisted candidates receive 0 points.
 
 
     Args:
         profile (PreferenceProfile): Profile to score.
         score_vector (Sequence[Union[float, Fraction]]): Score vector. Should be
-            non-increasing and non-negative. Vector should be as long as ``max_ballot_length`` in
+            non-increasing and non-negative. Vector should be as long as ``max_ranking_length`` in
             the profile. If it is shorter, we add 0s.
         to_float (bool, optional): If True, compute scores as floats instead of Fractions.
             Defaults to False.
@@ -318,7 +318,7 @@ def score_profile_from_rankings(
     """
     validate_score_vector(score_vector)
 
-    max_length = profile.max_ballot_length
+    max_length = profile.max_ranking_length
     if len(score_vector) < max_length:
         score_vector = list(score_vector) + [0] * (max_length - len(score_vector))
 
@@ -380,7 +380,7 @@ def first_place_votes(
     """
     # equiv to score vector of (1,0,0,...)
     return score_profile_from_rankings(
-        profile, [1] + [0] * (profile.max_ballot_length - 1), to_float, tie_convention
+        profile, [1] + [0] * (profile.max_ranking_length - 1), to_float, tie_convention
     )
 
 
@@ -442,7 +442,7 @@ def borda_scores(
             Dictionary mapping candidates to Borda scores.
     """
     if not borda_max:
-        borda_max = profile.max_ballot_length
+        borda_max = profile.max_ranking_length
 
     score_vector = list(range(borda_max, 0, -1))
 
@@ -750,7 +750,7 @@ def ballot_lengths(
 ) -> Union[dict[int, float], dict[int, Fraction]]:
     """
     Compute the frequency of ballot lengths in the profile.
-    Includes all lengths from 1 to ``max_ballot_length`` as keys.
+    Includes all lengths from 1 to ``max_ranking_length`` as keys.
     Ballots must have rankings.
 
     Args:
@@ -765,7 +765,7 @@ def ballot_lengths(
         TypeError: All ballots must have rankings.
     """
 
-    ballot_lengths = {i: Fraction(0) for i in range(1, profile.max_ballot_length + 1)}
+    ballot_lengths = {i: Fraction(0) for i in range(1, profile.max_ranking_length + 1)}
 
     for ballot in profile.ballots:
         if not ballot.ranking:
