+++ tests/pref_profile/df/test_pp_df_ranking_and_score.py
@@ -60,7 +60,7 @@ def test_pp_df_ranking_and_score_args():
         contains_rankings=True,
         contains_scores=True,
         candidates=["A", "B", "C", "D", "E"],
-        max_ballot_length=3,
+        max_ranking_length=3,
     )
     data = {
         "A": [
