### 2025-04-03 61880a2
change str method
