+++ tests/pref_profile/utils/test_pp_ballot_dict.py
@@ -1,6 +1,7 @@
 from fractions import Fraction
 from votekit.ballot import Ballot
 from votekit.pref_profile import PreferenceProfile
+from votekit.pref_profile.utils import profile_to_ballot_dict
 
 
 def test_to_ballot_dict():
@@ -12,12 +13,12 @@ def test_to_ballot_dict():
             Ballot(scores={"A": 4}),
         )
     )
-    rv = profile.to_ballot_dict(standardize=False)
+    rv = profile_to_ballot_dict(profile, standardize=False)
     assert rv[Ballot(ranking=({"A"}, {"B"}))] == Fraction(5, 2)
     assert rv[Ballot(ranking=({"C"}, {"B"}), scores={"A": 4})] == Fraction(2, 1)
-    assert rv[Ballot(scores={"A": 4})] == Fraction(1)  
+    assert rv[Ballot(scores={"A": 4})] == Fraction(1)
 
-    rv = profile.to_ballot_dict(standardize=True)
+    rv = profile_to_ballot_dict(profile, standardize=True)
     assert rv[Ballot(ranking=({"A"}, {"B"}))] == Fraction(5, 11)
     assert rv[Ballot(ranking=({"C"}, {"B"}), scores={"A": 4})] == Fraction(4, 11)
-    assert rv[Ballot(scores={"A": 4})] == Fraction(2,11)
+    assert rv[Ballot(scores={"A": 4})] == Fraction(2, 11)
