+++ src/votekit/pref_profile/pref_profile.py
@@ -3,7 +3,8 @@ import csv
 from fractions import Fraction
 import pandas as pd
 from pydantic import ConfigDict, field_validator, model_validator
-from .ballot import Ballot
+from ..ballot import Ballot
+from .utils import _df_to_ballot_tuple
 from pydantic.dataclasses import dataclass
 from typing_extensions import Self
 from dataclasses import field
@@ -12,80 +13,6 @@ from typing import Optional
 from functools import partial
 import warnings
 import pickle
-
-def _convert_ranking_cols_to_ranking(
-    row: pd.Series, ranking_cols: list[str]
-) -> Optional[tuple[frozenset, ...]]:
-    """
-    Convert the ranking cols to a ranking tuple.
-
-    """
-    ranking = []
-    for i, col in enumerate(ranking_cols):
-        if pd.isna(row[col]):
-            if not all(pd.isna(row[c]) for c in ranking_cols[i:]):
-                raise ValueError(
-                    f"Row {row} has NaN values between valid ranking positions. "
-                    "NaN values can only trail on a ranking."
-                )
-
-            break
-
-        ranking.append(row[col])
-
-    return tuple(ranking) if ranking else None
-
-
-def _convert_row_to_ballot(
-    row: pd.Series,
-    ranking_cols: list[str],
-    weight_col: str,
-    id_col: str,
-    voter_set_col: str,
-    candidates: list[str],
-) -> Ballot:
-    """
-    Convert a row of a properly formatted df to a Ballot.
-    """
-    ranking = _convert_ranking_cols_to_ranking(row, ranking_cols)
-    scores = {c: row[c] for c in candidates if c in row and not pd.isna(row[c])}
-    id = row[id_col] if not pd.isna(row[id_col]) else None
-    voter_set = row[voter_set_col]
-    weight = row[weight_col]
-
-    return Ballot(
-        ranking=ranking,
-        scores=scores if scores else None,
-        weight=weight,
-        id=id,
-        voter_set=voter_set,
-    )
-
-
-def _df_to_ballot_tuple(
-    df: pd.DataFrame,
-    candidates: list[str],
-    ranking_cols: list[str] = [],
-    weight_col: str = "Weight",
-    id_col: str = "ID",
-    voter_set_col: str = "Voter Set",
-) -> tuple[Ballot]:
-    """
-    Convert a df into a list of ballots.
-    """
-    return tuple(
-        df.apply(
-            partial(
-                _convert_row_to_ballot,
-                ranking_cols=ranking_cols,
-                weight_col=weight_col,
-                id_col=id_col,
-                candidates=candidates,
-                voter_set_col=voter_set_col,
-            ),
-            axis="columns",
-        )
-    )
 
 
 @dataclass(frozen=True, config=ConfigDict(arbitrary_types_allowed=True))
@@ -128,12 +55,12 @@ class PreferenceProfile:
         ValueError: contains_scores is set to False but a ballot contains a score.
         ValueError: contains_scores is set to True but no ballot contains a score.
         ValueError: max_ranking_length is set but a ballot ranking excedes the length.
-        ValueError: a candidate is found on a ballot that is not listed on a provided 
+        ValueError: a candidate is found on a ballot that is not listed on a provided
             candidate list.
         ValueError: candidates must be unique.
-     
+
     Warns:
-        UserWarning: max_ranking_length is set but contains_rankings is False. 
+        UserWarning: max_ranking_length is set but contains_rankings is False.
             Sets max_ranking_length to 0.
 
     """
@@ -147,8 +74,6 @@ class PreferenceProfile:
     total_ballot_wt: Fraction = Fraction(0)
     contains_rankings: Optional[bool] = None
     contains_scores: Optional[bool] = None
-
-
 
     @field_validator("candidates")
     @classmethod
[…]
