+++ src/votekit/pref_profile/__init__.py
@@ -1,0 +1,7 @@
+from .pref_profile import PreferenceProfile  # noqa
+from .utils import (
+    profile_to_ballot_dict, # noqa
+    profile_to_ranking_dict, # noqa
+    profile_to_scores_dict, # noqa
+    convert_row_to_ballot, # noqa
+)

[…]
