### 2025-04-04 a9c30bd
distances refactor with utils
