+++ tests/pref_profile/test_pp_csv.py
@@ -1,50 +1,123 @@
 from votekit.ballot import Ballot
 from votekit.pref_profile import PreferenceProfile
 
-filepath= "tests/data/csv"
+filepath = "tests/data/csv"
+
+
 def test_csv_bijection_rankings():
-    profile_rankings = PreferenceProfile(ballots=(Ballot(ranking = ({"A", "B"}, frozenset(), {"C"}), id="u", voter_set={"Chris", "Peter"}, weight = 3/2),
-                                     Ballot(ranking = ({"A", "B"}, frozenset(), {"C"}), id="v", voter_set={"Moon"}, weight = 1/2),
-                                     Ballot(ranking = ({"A"}, {"B"},), id="yes"),
-                                     Ballot(ranking = ({"A"}, {"B"},), id = "hi"),
-                                     Ballot(ranking = ({"A"}, {"B"},), id = "X29"))*5,
+    profile_rankings = PreferenceProfile(
+        ballots=(
+            Ballot(
+                ranking=({"A", "B"}, frozenset(), {"C"}),
+                id="u",
+                voter_set={"Chris", "Peter"},
+                weight=3 / 2,
+            ),
+            Ballot(
+                ranking=({"A", "B"}, frozenset(), {"C"}),
+                id="v",
+                voter_set={"Moon"},
+                weight=1 / 2,
+            ),
+            Ballot(
+                ranking=(
+                    {"A"},
+                    {"B"},
+                ),
+                id="yes",
+            ),
+            Ballot(
+                ranking=(
+                    {"A"},
+                    {"B"},
+                ),
+                id="hi",
+            ),
+            Ballot(
+                ranking=(
+                    {"A"},
+                    {"B"},
+                ),
+                id="X29",
+            ),
+        )
+        * 5,
+        max_ranking_length=3,
+        candidates=["A", "B", "C", "D", "E"],
+    )
 
-                                    max_ranking_length=3,
-                                    candidates=["A", "B", "C", "D", "E"])
-
-    
     profile_rankings.to_csv(f"{filepath}/test_csv_pp_rankings.csv")
-    read_profile =  PreferenceProfile.from_csv(f"{filepath}/test_csv_pp_rankings.csv")
+    read_profile = PreferenceProfile.from_csv(f"{filepath}/test_csv_pp_rankings.csv")
     assert profile_rankings == read_profile
 
 
 def test_csv_bijection_scores():
-    profile_scores = PreferenceProfile(ballots=(
-                                     Ballot(scores = {"A": 2, "B": 4, "D": 1}, ),
-                                     Ballot(scores = {"A": 2, "B": 4, "D": 1}),
-                                     Ballot(scores = {"A": 2, "B": 4, "C": 1}, id="yes"),
-                                     Ballot(scores = {"A": 2, "B": 4, "C": 1}, id = "hi"),
-                                     Ballot(scores = {"A": 5, "B": 4, "C": 1}, id = "X29"))*5,
-                                    candidates=["A", "B", "C", "D", "E"])
-    
+    profile_scores = PreferenceProfile(
+        ballots=(
+            Ballot(
+                scores={"A": 2, "B": 4, "D": 1},
+            ),
+            Ballot(scores={"A": 2, "B": 4, "D": 1}),
+            Ballot(scores={"A": 2, "B": 4, "C": 1}, id="yes"),
+            Ballot(scores={"A": 2, "B": 4, "C": 1}, id="hi"),
+            Ballot(scores={"A": 5, "B": 4, "C": 1}, id="X29"),
+        )
+        * 5,
+        candidates=["A", "B", "C", "D", "E"],
+    )
+
[…]
