+++ src/votekit/pref_profile/__init__.py
@@ -4,4 +4,6 @@ from .utils import (
     profile_to_ranking_dict,  # noqa
     profile_to_scores_dict,  # noqa
     convert_row_to_ballot,  # noqa
+    profile_df_head,  # noqa
+    profile_df_tail,  # noqa
 )
