+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - renamed `cleaning.deduplicate_profile` to `cleaning.remove_repeated_candidates`.
 - moved `remove_cand` from `utils` to `cleaning`, and removed the older function `remove_noncands`.
 - renamed `PreferenceProfile.condense_ballots()` to `PreferenceProfile.group_ballots()`.
+<<<<<<< HEAD
+=======
+
+
+>>>>>>> 19b0812 (changelog)
 
 ## Fixed
 - `PreferenceProfile.group_ballots()` now also groups the `voter_set` attribute of ballots.
