### 2025-04-07 afc536e
Refactor ballot_list_to_df
