+++ src/votekit/pref_profile/pref_profile.py
@@ -9,7 +9,7 @@ from pydantic.dataclasses import dataclass
 from typing_extensions import Self
 from dataclasses import field
 import numpy as np
-from typing import Optional
+from typing import Optional, Tuple
 from functools import partial
 import warnings
 import pickle
@@ -30,9 +30,13 @@ class PreferenceProfile:
             many candidates are allowed to be ranked in an election. Defaults to longest observed
             ballot.
         contains_rankings (bool, optional): Whether or not the profile contains ballots with
-            rankings. If not provided, will set itself to correct value given the ballot list.
+            rankings. If no boolean is provided, then the appropriate boolean value will be
+            interpreted from the input preference profile (i.e. if some ballot in the profile has
+            a ranking, then this will be set  to `True`).
         contains_scores (bool, optional): Whether or not the profile contains ballots with
-            scores. If not provided, will set itself to correct value given the ballot list.
+            scores. If no boolean is provided, then the appropriate boolean value will be
+            interpreted from the input preference profile (i.e. if some ballot in the profile has
+            a ranking, then this will be set  to `True`).
 
     Parameters:
         ballots (tuple[Ballot]): Tuple of ``Ballot`` objects.
@@ -85,8 +89,105 @@ class PreferenceProfile:
                 raise ValueError("All candidates must be unique.")
         return candidates
 
-    @model_validator(mode="after")
-    def ballot_list_to_df(self) -> Self:
+    def __update_ballot_scores_data(
+        self,
+        ballot_data: dict[str, list],
+        idx: int,
+        ballot: Ballot,
+        candidates_cast: list[str],
+        num_ballots: int,
+    ) -> None:
+        if self.contains_scores is False:
+            raise ValueError(
+                (
+                    f"Ballot {ballot} has scores {ballot.scores} but contains_scores is "
+                    "set to False."
+                )
+            )
+
+        for c, score in ballot.scores.items():
+            if ballot.weight > 0 and c not in candidates_cast:
+                candidates_cast.append(c)
+
+            if c not in ballot_data:
+                if self.candidates:
+                    raise ValueError(
+                        f"Candidate {c} found in ballot {ballot} but not in "
+                        f"candidate list {self.candidates}."
+                    )
+                ballot_data[c] = [np.nan] * num_ballots
+            ballot_data[c][idx] = score
+
+    def __update_ballot_rankings_data(
+        self,
+        ballot_data: dict[str, list],
+        idx: int,
+        ballot: Ballot,
+        candidates_cast: list[str],
+        num_ballots: int,
+    ) -> None:
+        if self.contains_rankings is False:
+            raise ValueError(
+                (
+                    f"Ballot {ballot} has ranking {ballot.ranking} but contains_rankings is"
+                    " set to False."
+                )
+            )
+
+        for j, cand_set in enumerate(ballot.ranking):
+            for c in cand_set:
+                if self.candidates:
+                    if c not in self.candidates:
+                        raise ValueError(
+                            f"Candidate {c} found in ballot {ballot} but not in "
+                            f"candidate list {self.candidates}."
+                        )
+                if ballot.weight > 0 and c not in candidates_cast:
+                    candidates_cast.append(c)
+            if f"Ranking_{j+1}" not in ballot_data:
+                if self.max_ranking_length:
+                    raise ValueError(
+                        f"Max ballot length {self.max_ranking_length} given but "
+                        "ballot {b} has length at least {j+1}."
+                    )
+                ballot_data[f"Ranking_{j+1}"] = [np.nan] * num_ballots
+
+            ballot_data[f"Ranking_{j+1}"][idx] = cand_set
+
+    def __update_ballot_data_attrs(
[…]
