+++ tests/pref_profile/df/test_pp_df_ranking.py
@@ -6,7 +6,7 @@ from fractions import Fraction
 
 ballots_rankings = [
     Ballot(ranking=({"A"}, {"B"}, {"C"}), weight=2),
-    Ballot(ranking=({"A", "B"}, frozenset(), {"D"}), id="X29", voter_set={"Chris"}),
+    Ballot(ranking=({"A", "B"}, frozenset(), {"D"}), voter_set={"Chris"}),
     Ballot(),
     Ballot(weight=0),
 ]
@@ -18,7 +18,6 @@ def test_pp_df_rankings():
         "Ranking_1": [frozenset({"A"}), frozenset({"A", "B"}), np.nan, np.nan],
         "Ranking_2": [frozenset({"B"}), frozenset(), np.nan, np.nan],
         "Ranking_3": [frozenset({"C"}), frozenset({"D"}), np.nan, np.nan],
-        "ID": [np.nan, "X29", np.nan, np.nan],
         "Voter Set": [set(), {"Chris"}, set(), set()],
         "Weight": [Fraction(2), Fraction(1), Fraction(1), Fraction(0)],
     }
@@ -41,7 +40,6 @@ def test_pp_df_rankings_args():
         "Ranking_2": [frozenset({"B"}), frozenset(), np.nan, np.nan],
         "Ranking_3": [frozenset({"C"}), frozenset({"D"}), np.nan, np.nan],
         "Ranking_4": [np.nan, np.nan, np.nan, np.nan],
-        "ID": [np.nan, "X29", np.nan, np.nan],
         "Voter Set": [set(), {"Chris"}, set(), set()],
         "Weight": [Fraction(2), Fraction(1), Fraction(1), Fraction(0)],
     }
