### 2025-04-08 423ff11
add ProfileError
