### 2025-04-11 9a544d9
refactor to from csv method
