+++ src/votekit/pref_profile/pref_profile.py
@@ -10,7 +10,6 @@ from typing_extensions import Self
 from dataclasses import field
 import numpy as np
 from typing import Optional, Tuple
-from functools import partial
 import warnings
 import pickle
 from .profile_error import ProfileError
@@ -442,77 +441,332 @@ class PreferenceProfile:
 
     __repr__ = __str__
 
-    def to_csv(self, fpath: str):
+    def __csv_header(
+        self, candidate_mapping: dict[str, str], include_voter_set: bool
+    ) -> list[list]:
         """
-        Saves PreferenceProfile to CSV.
+        Construct the header rows for the PrefProfile a custom CSV format.
+
+        Args:
+            include_voter_set (bool): Whether or not to include the voter set of each
+                ballot.
+        """
+        header = [
+            ["VoteKit PreferenceProfile"],
+            ["Candidates"],
+            [f"({c}:{cand_label})" for c, cand_label in candidate_mapping.items()],
+        ]
+        header += [["Max Ranking Length"], [str(self.max_ranking_length)]]
+        header += [["Includes Voter Set"], [str(include_voter_set)]]
+        header += [["="] * 10]
+
+        return header
+
+    def __csv_score_list(self, ballot: Ballot) -> list:
+        """
+        Create the list of score data for a ballot in the profile.
+
+        Args:
+            ballot (Ballot): Ballot.
+
+        """
+        if ballot.scores:
+            return [float(ballot.scores.get(c, 0)) for c in self.candidates]
+
+        return [""] * len(self.candidates)
+
+    def __csv_ranking_list(
+        self, ballot: Ballot, candidate_mapping: dict[str, str]
+    ) -> list:
+        """
+        Create the list of ranking data for a ballot in the profile.
+
+        Args:
+            ballot (Ballot): Ballot.
+            candidate_mapping (dict[str, int]): Mapping candidate names to integers.
+
+        """
+        if ballot.ranking:
+            ranking_list = [
+                set([candidate_mapping[c] for c in cand_set]) if cand_set else "{}"
+                for cand_set in ballot.ranking
+            ]
+            if len(ranking_list) != self.max_ranking_length:
+                ranking_list += [""] * (self.max_ranking_length - len(ranking_list))
+
+            return ranking_list
+
+        return [""] * self.max_ranking_length
+
+    def __csv_ballot_row(
+        self, ballot: Ballot, include_voter_set: bool, candidate_mapping: dict[str, str]
+    ) -> list[list]:
+        """
+        Create the row for a ballot in the profile.
+
+        Args:
+            ballot (Ballot): Ballot.
+            include_voter_set (bool): Whether or not to include the voter set of each
+                ballot.
+            candidate_mapping (dict[str, int]): Mapping candidate names to integers.
+
+        """
+        row = []
+        if self.contains_scores:
+            row += self.__csv_score_list(ballot)
+        row += ["&"]
+
+        if self.contains_rankings:
+            row += self.__csv_ranking_list(ballot, candidate_mapping)
+        row += ["&"]
+
+        n, d = ballot.weight.as_integer_ratio()
+        row += [f"{n}/{d}"]
+
+        if include_voter_set:
+            row += ["&"] + [v for v in sorted(ballot.voter_set)]
+
+        return row
+
+    def __csv_data_column_names(
+        self, include_voter_set: bool, candidate_mapping: dict[str, str]
+    ) -> list:
+        """
+        Create the data column header.
+
+        Args:
+            include_voter_set (bool): Whether or not to include the voter set of each
+                ballot.
+        """
+        data_col_names = []
+        if self.contains_scores:
+            data_col_names += [
+                f"{cand_label}" for cand_label in candidate_mapping.values()
+            ]
+        data_col_names += ["&"]
[…]
