+++ src/votekit/pref_profile/utils.py
@@ -12,12 +12,15 @@ from functools import partial
 from .profile_error import ProfileError
 
 
-def _convert_ranking_cols_to_ranking(row: pd.Series) -> Optional[tuple[frozenset, ...]]:
+def _convert_ranking_cols_to_ranking(
+    row: pd.Series, max_ranking_length: int
+) -> Optional[tuple[frozenset, ...]]:
     """
     Convert the ranking cols to a ranking tuple in profile.df.
 
     Args:
         row (pd.Series): Row of a profile.df.
+        max_ranking_length (int, optional): The maximum length of a ranking.
 
     Returns:
         Optional[tuple[frozenset, ...]]: Ranking of ballot.
@@ -27,10 +30,14 @@ def _convert_ranking_cols_to_ranking(row: pd.Series) -> Optional[tuple[frozenset
 
     """
     ranking = []
-    ranking_cols = [c for c in row.index if "Ranking_" in c]
-    for i, col in enumerate(ranking_cols):
-        if pd.isna(row[col]):
-            if not all(pd.isna(row[c]) for c in ranking_cols[i:]):
+    ranking_cols_idxs = [f"Ranking_{i+1}" for i in range(max_ranking_length)]
+
+    if any(idx not in row.index for idx in ranking_cols_idxs):
+        raise ValueError(f"Row has improper ranking columns: {row.index}.")
+
+    for i, col_idx in enumerate(ranking_cols_idxs):
+        if pd.isna(row[col_idx]):
+            if not all(pd.isna(row[idx]) for idx in ranking_cols_idxs[i:]):
                 raise ValueError(
                     f"Row {row} has NaN values between valid ranking positions. "
                     "NaN values can only trail on a ranking."
@@ -38,14 +45,13 @@ def _convert_ranking_cols_to_ranking(row: pd.Series) -> Optional[tuple[frozenset
 
             break
 
-        ranking.append(row[col])
+        ranking.append(row[col_idx])
 
     return tuple(ranking) if ranking else None
 
 
 def convert_row_to_ballot(
-    row: pd.Series,
-    candidates: tuple[str, ...],
+    row: pd.Series, candidates: tuple[str, ...], max_ranking_length: int = 0
 ) -> Ballot:
     """
     Convert a row of a properly formatted profile.df to a Ballot.
@@ -53,11 +59,15 @@ def convert_row_to_ballot(
     Args:
         row (pd.Series): Row of a profile.df.
         candidates (tuple[str,...]): The name of the candidates.
+        max_ranking_length (int, optional): The maximum length of a ranking. Defaults to 0, which
+            is used for ballots with no ranking.
 
     Returns:
         Ballot: Ballot corresponding to the row of the df.
     """
-    ranking = _convert_ranking_cols_to_ranking(row)
+    ranking = None
+    if max_ranking_length > 0:
+        ranking = _convert_ranking_cols_to_ranking(row, max_ranking_length)
     scores = {c: row[c] for c in candidates if c in row and not pd.isna(row[c])}
     voter_set = row["Voter Set"]
     weight = row["Weight"]
@@ -71,8 +81,7 @@ def convert_row_to_ballot(
 
 
 def _df_to_ballot_tuple(
-    df: pd.DataFrame,
-    candidates: tuple[str, ...],
+    df: pd.DataFrame, candidates: tuple[str, ...], max_ranking_length: int = 0
 ) -> tuple[Ballot, ...]:
     """
     Convert a properly formatted profile.df into a list of ballots.
@@ -80,6 +89,8 @@ def _df_to_ballot_tuple(
     Args:
         df (pd.DataFrame): A profile.df.
         candidates (tuple[str,...]): The candidates.
+        max_ranking_length (int, optional): The maximum length of a ranking. Defaults to 0, which
[…]
