+++ paper/paper.md
@@ -141,8 +141,7 @@ parameters designed to be tailored by the user to the specific locality. Availab
 
 -   **Approval-based (set).** Approval voting, Bloc plurality.
 
-See generally [@electoralhandbook; @STV; @Borda; @TopTwo; @SequentialRCV] for
-references. This list does not include every method that has attracted theoretical investigation; rather, it is oriented to methods used or considered for political representation, such as the final-four system in Alaska or the sequential RCV in Utah local elections. In addition, `VoteKit` is flexible enough to allow users to write custom voting rules.
+For references for the listed election methods, see: [@electoralhandbook; @STV; @Borda; @TopTwo; @SequentialRCV]. This list does not include every method that has attracted theoretical investigation; rather, it is oriented to methods used or considered for political representation, such as the final-four system in Alaska or the sequential RCV in Utah local elections. In addition, `VoteKit` is flexible enough to allow users to write custom voting rules.
 
 Reform advocates also need to describe voting mechanisms and their
 likely outcomes effectively to members of their communities. The end-to-end pipeline provided by `VoteKit` allows advocates to toggle different system settings and compare expected outcomes. For example, 
