+++ paper/paper.bib
@@ -234,7 +234,7 @@ WA School District},
   author = {Gerdus Benad\`e and Christopher Donnay and Moon Duchin and Thomas Weighill},
   title  = "Proportionality for ranked voting, in theory and practice",
   year   = "2024",
-  journal = "Preprint.",
+  journal = "Preprint",
     howpublished={\url{https://mggg.org/PRVTP}}
 }
 
@@ -244,7 +244,8 @@ WA School District},
   author = {Moon Duchin and David Shmoys and Kristopher Tapp},
   title  = "Learning blocs and slates from ranked-choice ballots",
   year   = "2024",
-  journal = "Preprint"
+  journal = "Preprint",
+   howpublished={\url{https://mggg.org/clustering}}
 }
 
 @misc{FairVote,
