### 2025-04-15 acfda94
fix article type for clustering
