+++ paper/paper.bib
@@ -240,12 +240,12 @@ WA School District},
 
 
 
-@article{duchin_tapp_24,
+@misc{duchin_tapp_24,
   author = {Moon Duchin and David Shmoys and Kristopher Tapp},
   title  = "Learning blocs and slates from ranked-choice ballots",
   year   = "2024",
   journal = "Preprint",
-   howpublished={\url{https://mggg.org/clustering}}
+  howpublished={\url{https://mggg.org/clustering}}
 }
 
 @misc{FairVote,
