### 2025-04-15 #7477 @cdonnay
@britta-wstnr we're ready! I made the final edits and added the link to the preprint. Everything should be updated in the JOSS branch now.

Here is the version number of VoteKit with the additions from this review process: 3.1.0
Here is the DOI of the Zenodo archive of VoteKit 3.1.0: 10.5281/zenodo.14975305
I reviewed the orcids of the authors and made sure the licenses match.
