### 2025-04-15 66b3c6c
add test group by equality
