### 2025-04-15 55050af
refactor from_csv
