+++ src/votekit/pref_profile/pref_profile.py
@@ -15,6 +15,135 @@ import pickle
 from .profile_error import ProfileError
 
 
+def _parse_profile_data_from_csv(
+    csv_data: list[list[str]],
+) -> Tuple[dict[str, str], int, bool, bool, bool, list[int]]:
+    """
+    Parse the profile data from a PreferenceProfile csv.
+
+    Args:
+        csv_data (list[list[str]]): Data from csv.
+
+    Returns:
+        Tuple[dict[str, str], int, bool, bool, bool, list[int]]:
+            inv_candidate_mapping, max_ranking_length, contains_rankings, contains_scores,
+            includes_voter_set, break_indices
+    """
+    candidate_row = [c_tuple.strip("()").split(":") for c_tuple in csv_data[2]]
+    inv_candidate_mapping = {prefix: cand for cand, prefix in candidate_row}
+
+    max_ranking_length = int(csv_data[4][0])
+    contains_rankings = max_ranking_length > 0
+
+    includes_voter_set = csv_data[6][0] == "True"
+
+    ballot_data_column_names = csv_data[8]
+    contains_scores = (
+        list(inv_candidate_mapping.values())[0] in ballot_data_column_names
+    )
+    break_indices = [
+        i for i, col_name in enumerate(ballot_data_column_names) if col_name == "&"
+    ]
+
+    return (
+        inv_candidate_mapping,
+        max_ranking_length,
+        contains_rankings,
+        contains_scores,
+        includes_voter_set,
+        break_indices,
+    )
+
+
+def _parse_ballot_from_csv(
+    ballot_row: list[str],
+    contains_rankings: bool,
+    contains_scores: bool,
+    includes_voter_set: bool,
+    break_indices: list[int],
+    inv_candidate_mapping: dict[str, str],
+) -> Ballot:
+    """
+    Parse a ballot from a PreferenceProfile csv row.
+
+    Args:
+        ballot_row (list[str]): Row from the csv file containing ballot data.
+        contains_rankings (bool): Whether or not the csv contains rankings.
+        contains_scores (bool): Whether or not the csv contains scores.
+        includes_voter_set (bool): Whether or not the csv contains voter sets.
+        break_indices (list[int]): Where the columns of the csv change from one data type to
+            another.
+        inv_candidate_mapping (dict[str, str]): The iverted candidate mapping of prefix
+            to the cand.
+
+    Returns:
+        Ballot: Ballot formatted from row of csv.
+    """
+    candidates = list(inv_candidate_mapping.values())
+    ranking_start = break_indices[0] + 1
+    ranking_end = break_indices[1]
+
+    scores = None
+    formatted_ranking = None
+    voter_set = set()
+
+    num, denom = ballot_row[break_indices[1] + 1].split("/")
+    weight = Fraction(int(num), int(denom))
+
+    if contains_scores:
+        scores = {
+            c: Fraction(float(ballot_row[i]))
+            for i, c in enumerate(candidates)
+            if ballot_row[i]
+        }
+
+    if contains_rankings:
+        ranking = ballot_row[ranking_start:ranking_end]
+        temp_ranking = [
+            cand_set.strip("{}").split(", ") for cand_set in ranking if cand_set != ""
+        ]
+        formatted_ranking = tuple(
+            [
+                (
+                    frozenset(inv_candidate_mapping[c.strip("'")] for c in cand_set)
+                    if cand_set != [""]
+                    else frozenset()
+                )
+                for cand_set in temp_ranking
+            ]
+        )
+
+    if includes_voter_set:
+        voter_set = set(ballot_row[break_indices[-1] + 1 :])
+
+    return Ballot(
[…]
