### 2025-04-15 e12acdb
add doc strings to profile
