+++ src/votekit/pref_profile/pref_profile.py
@@ -214,6 +214,15 @@ class PreferenceProfile:
     def cands_must_be_unique(
         cls, candidates: Optional[tuple[str, ...]]
     ) -> Optional[tuple[str, ...]]:
+        """
+        Checks that candidate names are unique.
+
+        Args:
+            candidates (tuple[str], optional): Candidate names.
+
+        Returns:
+            Optional[tuple[str]]: Candidate names.
+        """
         if candidates:
             if not len(set(candidates)) == len(candidates):
                 raise ProfileError("All candidates must be unique.")
@@ -223,6 +232,11 @@ class PreferenceProfile:
     def cands_not_ranking_columns(
         self,
     ) -> Self:
+        """
+        Ensures that candidate names do not match ranking columns.
+        Also added protection for from_csv method.
+
+        """
         for cand in self.candidates:
             if any(f"Ranking_{i}" == cand for i in range(len(self.candidates))):
                 raise ProfileError(
@@ -241,6 +255,16 @@ class PreferenceProfile:
         candidates_cast: list[str],
         num_ballots: int,
     ) -> None:
+        """
+        Update the score data from a ballot.
+
+        Args:
+            ballot_data (dict[str, list]): Dictionary storing ballot data.
+            idx (int): Index of ballot.
+            ballot (Ballot): Ballot.
+            candidates_cast (list[str]): List of candidates who have received votes.
+            num_ballots (int): Total number of ballots.
+        """
         if ballot.scores:
             if self.contains_scores is False:
                 raise ProfileError(
@@ -271,6 +295,16 @@ class PreferenceProfile:
         candidates_cast: list[str],
         num_ballots: int,
     ) -> None:
+        """
+        Update the ranking data from a ballot.
+
+        Args:
+            ballot_data (dict[str, list]): Dictionary storing ballot data.
+            idx (int): Index of ballot.
+            ballot (Ballot): Ballot.
+            candidates_cast (list[str]): List of candidates who have received votes.
+            num_ballots (int): Total number of ballots.
+        """
 
         if ballot.ranking:
             if self.contains_rankings is False:
@@ -309,6 +343,16 @@ class PreferenceProfile:
         candidates_cast: list[str],
         num_ballots: int,
     ) -> None:
+        """
+        Update all ballot data from a ballot.
+
+        Args:
+            ballot_data (dict[str, list]): Dictionary storing ballot data.
+            idx (int): Index of ballot.
+            ballot (Ballot): Ballot.
+            candidates_cast (list[str]): List of candidates who have received votes.
+            num_ballots (int): Total number of ballots.
+        """
         ballot_data["Weight"][idx] = ballot.weight
 
         if ballot.voter_set:
@@ -333,6 +377,13 @@ class PreferenceProfile:
             )
 
     def __init_ballot_data(self) -> Tuple[int, dict[str, list]]:
+        """
+        Create the ballot data objects.
+
+        Returns:
+            Tuple[int, dict[str, list]]: num_ballots, ballot_data
+
+        """
         num_ballots = len(self.ballots)
 
         ballot_data: dict[str, list] = {
@@ -356,7 +407,18 @@ class PreferenceProfile:
         self,
         ballot_data: dict[str, list],
         contains_scores_indicator: bool,
-    ):
+    ) -> pd.DataFrame:
+        """
+        Create a pandas dataframe from the ballot data.
+
+        Args:
+            ballot_data (dict[str, list]): Dictionary storing ballot data.
+            contains_scores_indicator (bool): Whether or not the profile contains ballots
+                with scores.
+
+        Returns:
+            pd.DataFrame: Dataframe of profile.
+        """
         df = pd.DataFrame(ballot_data)
[…]
