### 2025-04-15 7ff4180
add checks for file paths in to_ methods
