### 2025-04-17 c1a7db8
add robust from_csv validators
