+++ src/votekit/pref_profile/pref_profile.py
@@ -38,9 +38,7 @@ def _parse_profile_data_from_csv(
     includes_voter_set = csv_data[6][0] == "True"
 
     ballot_data_column_names = csv_data[8]
-    contains_scores = (
-        list(inv_candidate_mapping.values())[0] in ballot_data_column_names
-    )
+    contains_scores = list(inv_candidate_mapping.keys())[0] in ballot_data_column_names
     break_indices = [
         i for i, col_name in enumerate(ballot_data_column_names) if col_name == "&"
     ]
@@ -121,27 +119,448 @@ def _parse_ballot_from_csv(
     )
 
 
+def _validate_csv_header_values(header_data: list[list[str]]):
+    """
+    Validate that the values of the header rows are correct.
+
+    Args:
+        header_data (list[list[str]]): The rows of the csv file.
+
+    Raises:
+        ValueError: If csv header rows are improperly formatted for VoteKit.
+    """
+    boiler_plate = (
+        "This usually indicates that you are loading a csv that was not made with "
+        "PreferenceProfile.to_csv()."
+    )
+
+    if any(char not in c_tuple for c_tuple in header_data[2] for char in "(:)"):
+        raise ValueError(
+            (
+                "csv file is improperly formatted. Row 2 should contain tuples mapping candidates "
+                "to their unique prefixes. For example, (Chris:Ch), (Colleen: Co)."
+                + boiler_plate
+            )
+        )
+
+    if len(header_data[4]) != 1:
+        raise ValueError(
+            (
+                "csv file is improperly formatted. Row 4 should be a single non-negative integer. "
+                + boiler_plate
+            )
+        )
+    else:
+        try:
+            max_ranking_length = int(header_data[4][0])
+
+            if max_ranking_length < 0:
+                raise ValueError(
+                    (
+                        "csv file is improperly formatted. Row 4 should be a single"
+                        " non-negative integer. " + boiler_plate
+                    )
+                )
+        except ValueError:
+            raise ValueError(
+                (
+                    "csv file is improperly formatted. Row 4 should be a single"
+                    " non-negative integer. " + boiler_plate
+                )
+            )
+
+    if len(header_data[6]) != 1 or header_data[6][0] not in ["True", "False"]:
+        raise ValueError(
+            (
+                "csv file is improperly formatted. Row 6 should be 'True' or 'False'. This "
+                + boiler_plate
+            )
+        )
+
+
+def _validate_csv_header_rows(header_data: list[list[str]]):
+    """
+    Validate that the names of the header rows are correct.
+
+    Args:
+        header_data (list[list[str]]): The rows of the csv file.
+
+    Raises:
+        ValueError: If csv headers rows are improperly formatted for VoteKit.
+    """
+    boiler_plate = (
+        "This usually indicates that you are loading a csv that was not made with "
+        "PreferenceProfile.to_csv()."
+    )
+
+    if header_data[-1] != ["="] * 10:
+        raise ValueError(
+            (
+                "csv file is improperly formatted. Row 7 should be '=,=,=,=,=,=,=,=,=,='. This "
+                + boiler_plate
+            )
+        )
+
+    if header_data[0] != ["VoteKit PreferenceProfile"]:
+        raise ValueError(
+            (
+                "csv file is improperly formatted. Row 0 should be 'VoteKit PreferenceProfile'."
+                + boiler_plate
+            )
+        )
+
+    if header_data[1] != ["Candidates"]:
+        raise ValueError(
+            (
+                "csv file is improperly formatted. Row 1 should be 'Candidates'. This "
+                + boiler_plate
+            )
+        )
+
+    if header_data[3] != ["Max Ranking Length"]:
[…]
