+++ tests/pref_profile/test_pp_csv.py
@@ -53,21 +53,21 @@ def test_csv_bijection_scores():
     profile_scores = PreferenceProfile(
         ballots=(
             Ballot(
-                scores={"A": 2, "B": 4, "D": 1},
+                scores={"Alex": 2, "Allen": 4, "D": 1},
             ),
-            Ballot(scores={"A": 2, "B": 4, "D": 1}),
+            Ballot(scores={"Alex": 2, "Allen": 4, "D": 1}),
             Ballot(
-                scores={"A": 2, "B": 4, "C": 1},
-            ),
-            Ballot(
-                scores={"A": 2, "B": 4, "C": 1},
+                scores={"Alex": 2, "Allen": 4, "C": 1},
             ),
             Ballot(
-                scores={"A": 5, "B": 4, "C": 1},
+                scores={"Alex": 2, "Allen": 4, "C": 1},
+            ),
+            Ballot(
+                scores={"Alex": 5, "Allen": 4, "C": 1},
             ),
         )
         * 5,
-        candidates=["A", "B", "C", "D", "E"],
+        candidates=["Alex", "Allen", "C", "D", "E"],
     )
 
     profile_scores.to_csv(f"{filepath}/test_csv_pp_scores.csv", include_voter_set=True)
@@ -122,6 +122,287 @@ def test_csv_bijection_mixed():
     assert profile_mixed == read_profile
 
 
-def test_csv_error():
+def test_csv_bijection_mixed_no_voter_set():
+    profile_mixed = PreferenceProfile(
+        ballots=(
+            Ballot(
+                ranking=({"A", "B"}, frozenset(), {"C"}),
+                voter_set={"Chris", "Peter"},
+                weight=3 / 2,
+            ),
+            Ballot(
+                ranking=({"A", "B"}, frozenset(), {"C"}),
+                voter_set={"Moon"},
+                weight=1 / 2,
+            ),
+            Ballot(scores={"A": 2, "B": 4, "D": 1}),
+            Ballot(scores={"A": 2, "B": 4, "D": 1}),
+            Ballot(
+                ranking=(
+                    {"A"},
+                    {"B"},
+                ),
+                scores={"A": 2, "B": 4, "C": 1},
+            ),
+            Ballot(
+                ranking=(
+                    {"A"},
+                    {"B"},
+                ),
+                scores={"A": 2, "B": 4, "C": 1},
+            ),
+            Ballot(
+                ranking=(
+                    {"A"},
+                    {"B"},
+                ),
+                scores={"A": 5, "B": 4, "C": 1},
+            ),
+        )
+        * 2,
+        max_ranking_length=3,
+        candidates=["A", "B", "C", "D", "E"],
+    )
+
+    profile_mixed.to_csv(
+        f"{filepath}/test_csv_pp_mixed_no_voter_set.csv", include_voter_set=False
+    )
+    read_profile = PreferenceProfile.from_csv(
+        f"{filepath}/test_csv_pp_mixed_no_voter_set.csv"
+    )
+    assert profile_mixed.df.drop(columns=["Voter Set"]).equals(
+        read_profile.df.drop(columns=["Voter Set"])
+    )
+    assert (read_profile.df["Voter Set"] == set()).all()
+
+
+def test_csv_filepath_error():
     with pytest.raises(ValueError, match="File path must be provided."):
         PreferenceProfile().to_csv("")
+
+
+def test_csv_misformatted_header_rows_error():
+    with pytest.raises(
+        ValueError,
+        match=(
+            "csv file is improperly formatted. Row 0 should be "
+            "'VoteKit PreferenceProfile'."
+        ),
+    ):
+        PreferenceProfile.from_csv(f"{filepath}/test_csv_pp_misformat_header_0.csv")
[…]
