### 2025-04-20 acd9bf0
add todo notes
