### 2025-04-20 07209af
add draft of new profile.df usage
