### 2025-04-20 80f0fd3
strip whitespace ballot candidates
