### 2025-04-20 58db52a
add strip whitespace for votersets in csv
