### 2025-04-20 3a85d57
add more descriptive error messages
