+++ src/votekit/pref_profile/pref_profile.py
@@ -138,42 +138,42 @@ def _validate_csv_header_values(header_data: list[list[str]]):
         raise ValueError(
             (
                 "csv file is improperly formatted. Row 2 should contain tuples mapping candidates "
-                "to their unique prefixes. For example, (Chris:Ch), (Colleen: Co)."
-                + boiler_plate
+                "to their unique prefixes. For example, (Chris:Ch), (Colleen: Co). "
+                f"Not {header_data[2]}. " + boiler_plate
             )
         )
 
     if len(header_data[4]) != 1:
         raise ValueError(
             (
-                "csv file is improperly formatted. Row 4 should be a single non-negative integer. "
-                + boiler_plate
+                "csv file is improperly formatted. Row 4 should be a single non-negative integer, "
+                f"not {header_data[4]}. " + boiler_plate
             )
         )
-    else:
-        try:
-            max_ranking_length = int(header_data[4][0])
 
-            if max_ranking_length < 0:
-                raise ValueError(
-                    (
-                        "csv file is improperly formatted. Row 4 should be a single"
-                        " non-negative integer. " + boiler_plate
-                    )
-                )
-        except ValueError:
+    try:
+        max_ranking_length = int(header_data[4][0])
+
+        if max_ranking_length < 0:
             raise ValueError(
                 (
                     "csv file is improperly formatted. Row 4 should be a single"
-                    " non-negative integer. " + boiler_plate
+                    f" non-negative integer, not {max_ranking_length}. " + boiler_plate
                 )
             )
+    except ValueError:
+        raise ValueError(
+            (
+                "csv file is improperly formatted. Row 4 should be a single"
+                f" non-negative integer, not {header_data[4][0]}. " + boiler_plate
+            )
+        )
 
     if len(header_data[6]) != 1 or header_data[6][0] not in ["True", "False"]:
         raise ValueError(
             (
-                "csv file is improperly formatted. Row 6 should be 'True' or 'False'. This "
-                + boiler_plate
+                "csv file is improperly formatted. Row 6 should be 'True' or 'False',"
+                f" not {header_data[6]}. " + boiler_plate
             )
         )
 
@@ -196,40 +196,40 @@ def _validate_csv_header_rows(header_data: list[list[str]]):
     if header_data[-1] != ["="] * 10:
         raise ValueError(
             (
-                "csv file is improperly formatted. Row 7 should be '=,=,=,=,=,=,=,=,=,='. This "
-                + boiler_plate
+                "csv file is improperly formatted. Row 7 should be '=,=,=,=,=,=,=,=,=,=', "
+                f"not {header_data[-1]}. " + boiler_plate
             )
         )
 
     if header_data[0] != ["VoteKit PreferenceProfile"]:
         raise ValueError(
             (
-                "csv file is improperly formatted. Row 0 should be 'VoteKit PreferenceProfile'."
-                + boiler_plate
+                "csv file is improperly formatted. Row 0 should be 'VoteKit PreferenceProfile',"
+                f"not {header_data[0]}. " + boiler_plate
             )
         )
 
     if header_data[1] != ["Candidates"]:
         raise ValueError(
             (
-                "csv file is improperly formatted. Row 1 should be 'Candidates'. This "
-                + boiler_plate
+                "csv file is improperly formatted. Row 1 should be 'Candidates', "
+                f"not {header_data[1]}. " + boiler_plate
             )
         )
 
     if header_data[3] != ["Max Ranking Length"]:
         raise ValueError(
             (
-                "csv file is improperly formatted. Row 3 should be 'Max Ranking Length'. This "
-                + boiler_plate
+                "csv file is improperly formatted. Row 3 should be 'Max Ranking Length', "
[…]
