+++ src/votekit/pref_profile/pref_profile.py
@@ -354,26 +354,30 @@ def _validate_csv_ballot_score(
             )
         )
 
-    elif contains_scores:
-        if len(ballot_row[:break_idx]) != len(candidates):
+    elif not contains_scores and break_idx == 0:
+        return
+
+    if len(ballot_row[:break_idx]) != len(candidates):
+        raise ValueError(
+            (
+                f"csv file is improperly formatted. Ballot in row {row_index} is missing "
+                f"some scores: {ballot_row[:break_idx]} ." + boiler_plate
+            )
+        )
+
+    for score in ballot_row[:break_idx]:
+        if not score:
+            continue
+
+        try:
+            float(score)
+        except ValueError:
             raise ValueError(
                 (
-                    f"csv file is improperly formatted. Ballot in row {row_index} is missing "
-                    f"some scores: {ballot_row[:break_idx]} ." + boiler_plate
+                    f"csv file is improperly formatted. Ballot in row {row_index} has "
+                    f"non-float score value: {score}. " + boiler_plate
                 )
             )
-
-        for score in ballot_row[:break_idx]:
-            if score:
-                try:
-                    float(score)
-                except ValueError:
-                    raise ValueError(
-                        (
-                            f"csv file is improperly formatted. Ballot in row {row_index} has "
-                            f"non-float score value: {score}. " + boiler_plate
-                        )
-                    )
 
 
 def _validate_csv_ballot_ranking(
@@ -409,28 +413,33 @@ def _validate_csv_ballot_ranking(
                 + boiler_plate
             )
         )
-    elif max_ranking_length > 0:
-        for ranking_set in ballot_row[break_idxs[0] + 1 : break_idxs[1]]:
-            if ranking_set:
-                if ranking_set[0] != "{" or ranking_set[-1] != "}":
-                    raise ValueError(
-                        (
-                            f"csv file is improperly formatted. Ballot in row {row_index} has "
-                            f"improperly formatted ranking set: {ranking_set}. "
-                            + boiler_plate
-                        )
-                    )
+    elif max_ranking_length == 0:
+        return
 
-            if ranking_set not in ["", "{}"]:
-                cand_set = ranking_set.strip("{}").split(", ")
-                for c in cand_set:
-                    if c.strip("'") not in candidate_prefixes:
-                        raise ValueError(
-                            (
-                                f"csv file is improperly formatted. Ballot in row {row_index} has "
-                                f"undefined candidate prefix: {c}. " + boiler_plate
-                            )
-                        )
+    for ranking_set in ballot_row[break_idxs[0] + 1 : break_idxs[1]]:
+        if not ranking_set:
+            continue
+
+        if ranking_set[0] != "{" or ranking_set[-1] != "}":
+            raise ValueError(
+                (
+                    f"csv file is improperly formatted. Ballot in row {row_index} has "
+                    f"improperly formatted ranking set: {ranking_set}. " + boiler_plate
+                )
+            )
+
+        if ranking_set in ["", "{}"]:
+            continue
+
+        cand_set = ranking_set.strip("{}").split(", ")
+        for c in cand_set:
+            if c.strip("'") not in candidate_prefixes:
+                raise ValueError(
+                    (
+                        f"csv file is improperly formatted. Ballot in row {row_index} has "
+                        f"undefined candidate prefix: {c}. " + boiler_plate
+                    )
+                )
 
 
 def _validate_csv_ballot_weight(
@@ -460,31 +469,31 @@ def _validate_csv_ballot_weight(
                 " entry that is too long or short. " + boiler_plate
             )
         )
+
+    if "/" not in ballot_row[break_idxs[1] + 1]:
+        raise ValueError(
+            (
+                f"csv file is improperly formatted. Ballot in row {row_index} has a "
[…]
