### 2025-04-23 #7477 @britta-wstnr
Great, @cdonnay ! 🎉 I have downloaded the paper and will progress through my [post-review checklist](https://github.com/openjournals/joss-reviews/issues/7477#issuecomment-2668200870) now!
