### 2025-04-28 ae80cd2
first draft of cleaning with new CleanedProfile
