### 2025-04-28 38592df
fix parent_profile reinstatiation, add tests to catch
