+++ tests/cleaning/test_remove_repeated_cands.py
@@ -21,7 +21,7 @@ def test_remove_repeated_candidates():
     assert cleaned_profile != profile
     assert cleaned_profile.no_weight_altr_ballot_indices == set()
     assert cleaned_profile.no_ranking_and_no_scores_altr_ballot_indices == set()
-    assert cleaned_profile.valid_but_altr_ballot_indices == {0, 1}
+    assert cleaned_profile.nonempty_but_altr_ballot_indices == {0, 1}
     assert cleaned_profile.unaltr_ballot_indices == set()
 
 
@@ -48,7 +48,7 @@ def test_remove_repeated_candidates_ties():
     assert cleaned_profile != profile
     assert cleaned_profile.no_weight_altr_ballot_indices == set()
     assert cleaned_profile.no_ranking_and_no_scores_altr_ballot_indices == set()
-    assert cleaned_profile.valid_but_altr_ballot_indices == {0, 1}
+    assert cleaned_profile.nonempty_but_altr_ballot_indices == {0, 1}
     assert cleaned_profile.unaltr_ballot_indices == set()
 
 

[…]
