### 2025-04-29 4aa40ae
add repr to cleanedprofile
