### 2025-04-29 1518f69
change equivalence condition in condense profile
