+++ tests/elections/election_types/ranking/test_random_dictator.py
@@ -34,7 +34,7 @@ def test_random_dictator_simple():
     test_profile = PreferenceProfile(ballots=ballots, candidates=candidates)
 
     winner_counts = {c: 0 for c in candidates}
-    trials = 5000
+    trials = 500
 
     # Parallel execution
     n_jobs = -1  # Use all available cores
@@ -80,7 +80,7 @@ def test_random_dictator_4_candidates_without_ties():
     tot_fpv = sum(fpv.values())
     fpv = {c: v / tot_fpv for c, v in fpv.items()}
 
-    trials = 5000
+    trials = 500
 
     # Parallel execution
     n_jobs = -1  # Use all available cores
@@ -113,7 +113,7 @@ def test_random_dictator_4_candidates_with_ties():
 
     test_profile = PreferenceProfile(ballots=ballots, candidates=candidates)
 
-    trials = 5000
+    trials = 500
 
     fpv = first_place_votes(test_profile, tie_convention="average")
     tot_fpv = sum(fpv.values())
@@ -140,7 +140,7 @@ def test_random_dictator_4_candidates_large_sample(all_possible_ranked_ballots):
 
     ballots = all_possible_ranked_ballots(candidates)
 
-    trials = 5000
+    trials = 500
 
     for i, ballot in enumerate(ballots):
         if "A" in ballot.ranking[0]:
