### 2025-05-02 42a6228
strong type checking for boolean conditionals
