### 2025-05-02 0b06a79
change condensed ranking type check
