+++ tests/cleaning/test_remove_cand.py
@@ -31,10 +31,10 @@ def test_remove_cand():
         Ballot(ranking=[{"C"}, {"B"}, frozenset()], weight=3),
     )
     assert cleaned_profile != profile_no_ties
-    assert cleaned_profile.no_weight_altr_ballot_indices == set()
-    assert cleaned_profile.no_ranking_and_no_scores_altr_ballot_indices == set()
-    assert cleaned_profile.nonempty_but_altr_ballot_indices == {0, 1, 2}
-    assert cleaned_profile.unaltr_ballot_indices == set()
+    assert cleaned_profile.no_wt_altr_idxs == set()
+    assert cleaned_profile.no_rank_no_score_altr_idxs == set()
+    assert cleaned_profile.nonempty_altr_idxs == {0, 1, 2}
+    assert cleaned_profile.unaltr_idxs == set()
 
 
 def test_remove_mult_cands():
@@ -51,10 +51,10 @@ def test_remove_mult_cands():
         ]
     )
     assert cleaned_profile != profile_no_ties
-    assert cleaned_profile.no_weight_altr_ballot_indices == set()
-    assert cleaned_profile.no_ranking_and_no_scores_altr_ballot_indices == set()
-    assert cleaned_profile.nonempty_but_altr_ballot_indices == {0, 1, 2}
-    assert cleaned_profile.unaltr_ballot_indices == set()
+    assert cleaned_profile.no_wt_altr_idxs == set()
+    assert cleaned_profile.no_rank_no_score_altr_idxs == set()
+    assert cleaned_profile.nonempty_altr_idxs == {0, 1, 2}
+    assert cleaned_profile.unaltr_idxs == set()
 
 
 def test_remove_cand_with_ties():
@@ -71,10 +71,10 @@ def test_remove_cand_with_ties():
         ]
     )
     assert cleaned_profile != profile_with_ties
-    assert cleaned_profile.no_weight_altr_ballot_indices == set()
-    assert cleaned_profile.no_ranking_and_no_scores_altr_ballot_indices == set()
-    assert cleaned_profile.nonempty_but_altr_ballot_indices == {0, 1, 2}
-    assert cleaned_profile.unaltr_ballot_indices == set()
+    assert cleaned_profile.no_wt_altr_idxs == set()
+    assert cleaned_profile.no_rank_no_score_altr_idxs == set()
+    assert cleaned_profile.nonempty_altr_idxs == {0, 1, 2}
+    assert cleaned_profile.unaltr_idxs == set()
 
 
 def test_remove_cands_scores():
@@ -123,7 +123,7 @@ def test_remove_cands_scores():
         ),
     )
     assert cleaned_profile != profile
-    assert cleaned_profile.no_weight_altr_ballot_indices == set()
-    assert cleaned_profile.no_ranking_and_no_scores_altr_ballot_indices == set()
-    assert cleaned_profile.nonempty_but_altr_ballot_indices == {0, 1, 2, 3, 4}
-    assert cleaned_profile.unaltr_ballot_indices == set()
+    assert cleaned_profile.no_wt_altr_idxs == set()
+    assert cleaned_profile.no_rank_no_score_altr_idxs == set()
+    assert cleaned_profile.nonempty_altr_idxs == {0, 1, 2, 3, 4}
+    assert cleaned_profile.unaltr_idxs == set()
