### 2025-05-05 20230ff
use remove_and_condense in elections
