+++ src/votekit/elections/election_types/ranking/condo_borda.py
@@ -1,7 +1,7 @@
 from .abstract_ranking import RankingElection
 from ....pref_profile import PreferenceProfile
 from ...election_state import ElectionState
-from ....cleaning import remove_cand, condense_profile
+from ....cleaning import remove_and_condense
 from ....utils import elect_cands_from_set_ranking, borda_scores
 from ....graphs import PairwiseComparisonGraph
 
@@ -60,9 +60,8 @@ class CondoBorda(RankingElection):
         else:
             tiebreaks = {}
 
-        new_profile = condense_profile(
-            remove_cand([c for s in elected for c in s], profile)
-        )
+        new_profile = remove_and_condense([c for s in elected for c in s], profile)
+
         if store_states:
             self.election_states.append(
                 ElectionState(
