+++ src/votekit/elections/election_types/ranking/plurality_veto.py
@@ -7,7 +7,7 @@ from ....utils import (
     score_dict_to_ranking,
     tiebreak_set,
 )
-from ....cleaning import remove_cand, condense_profile
+from ....cleaning import remove_and_condense
 from fractions import Fraction
 import numpy as np
 from typing import Optional, Literal
@@ -225,15 +225,11 @@ class PluralityVeto(RankingElection):
             for c in eliminated_cands:
                 self.eliminated_dict[c] = True
 
-            new_profile = condense_profile(
-                remove_cand(
-                    eliminated_cands,
-                    profile,
-                    remove_zero_weight_ballots=False,
-                    remove_empty_ballots=False,
-                ),
-                remove_empty_ballots=False,
+            new_profile = remove_and_condense(
+                eliminated_cands,
+                profile,
                 remove_zero_weight_ballots=False,
+                remove_empty_ballots=False,
             )
 
             self.ballot_list = new_profile.ballots
