+++ src/votekit/elections/election_types/ranking/borda.py
@@ -1,7 +1,7 @@
 from .abstract_ranking import RankingElection
 from ....pref_profile import PreferenceProfile
 from ...election_state import ElectionState
-from ....cleaning import remove_cand, condense_profile
+from ....cleaning import remove_and_condense
 from ....utils import (
     elect_cands_from_set_ranking,
     validate_score_vector,
@@ -91,9 +91,7 @@ class Borda(RankingElection):
             prev_state.remaining, self.m, profile=profile, tiebreak=self.tiebreak
         )
 
-        new_profile = condense_profile(
-            remove_cand([c for s in elected for c in s], profile)
-        )
+        new_profile = remove_and_condense([c for s in elected for c in s], profile)
 
         if store_states:
             if self.score_function:  # mypy
