+++ src/votekit/elections/election_types/ranking/dominating_sets.py
@@ -1,7 +1,7 @@
 from .abstract_ranking import RankingElection
 from ....pref_profile import PreferenceProfile
 from ...election_state import ElectionState
-from ....cleaning import remove_cand, condense_profile
+from ....cleaning import remove_and_condense
 from ....graphs import PairwiseComparisonGraph
 
 
@@ -48,7 +48,7 @@ class DominatingSets(RankingElection):
 
         pwc_graph = PairwiseComparisonGraph(profile)
         dominating_tiers = pwc_graph.get_dominating_tiers()
-        new_profile = condense_profile(remove_cand(list(dominating_tiers[0]), profile))
+        new_profile = remove_and_condense(list(dominating_tiers[0]), profile)
 
         if store_states:
             elected = (frozenset(dominating_tiers[0]),)
