+++ src/votekit/elections/election_types/scores/rating.py
@@ -5,7 +5,7 @@ from ....utils import (
     score_profile_from_ballot_scores,
     elect_cands_from_set_ranking,
 )
-from ....cleaning import remove_cand, condense_profile
+from ....cleaning import remove_and_condense
 from typing import Optional, Union
 from fractions import Fraction
 
@@ -114,9 +114,7 @@ class GeneralRating(Election):
             prev_state.remaining, self.m, profile=profile, tiebreak=self.tiebreak
         )
 
-        new_profile = condense_profile(
-            remove_cand([c for s in elected for c in s], profile)
-        )
+        new_profile = remove_and_condense([c for s in elected for c in s], profile)
 
         if store_states:
             if self.score_function:
