+++ src/votekit/elections/election_types/ranking/top_two.py
@@ -2,7 +2,7 @@ from .abstract_ranking import RankingElection
 from ....pref_profile import PreferenceProfile
 from ...election_state import ElectionState
 from ....utils import first_place_votes
-from ....cleaning import remove_cand, condense_profile
+from ....cleaning import remove_and_condense
 from ..ranking import Plurality
 from typing import Optional, Literal
 from functools import partial
@@ -69,9 +69,10 @@ class TopTwo(RankingElection):
             remaining = plurality.get_elected()
             eliminated = plurality.get_remaining()
             tiebreaks = plurality.election_states[-1].tiebreaks
-            new_profile: PreferenceProfile = condense_profile(
-                remove_cand([c for s in eliminated for c in s], profile)
+            new_profile: PreferenceProfile = remove_and_condense(
+                [c for s in eliminated for c in s], profile
             )
+
             if self.score_function:
                 scores = self.score_function(new_profile)
 
