### 2025-05-05 021a817
add remove_and_condense
