+++ src/votekit/cleaning.py
@@ -36,10 +36,10 @@ def clean_profile(
     """
     new_ballots_and_idxs = [(Ballot(), -1)] * len(profile.ballots)
 
-    no_weight_altr_ballot_indices = set()
-    no_ranking_and_no_scores_altr_ballot_indices = set()
-    nonempty_but_altr_ballot_indices = set()
-    unaltr_ballot_indices = set()
+    no_wt_altr_idxs = set()
+    no_rank_no_score_altr_idxs = set()
+    nonempty_altr_idxs = set()
+    unaltr_idxs = set()
 
     for integer_idx, (i, b_row) in enumerate(profile.df.iterrows()):
         assert isinstance(i, int)
@@ -51,17 +51,17 @@ def clean_profile(
         new_b = clean_ballot_func(b)
 
         if new_b == b:
-            unaltr_ballot_indices.add(i)
+            unaltr_idxs.add(i)
 
         else:
             if (new_b.ranking or new_b.scores) and new_b.weight > 0:
-                nonempty_but_altr_ballot_indices.add(i)
+                nonempty_altr_idxs.add(i)
 
             if new_b.weight == 0:
-                no_weight_altr_ballot_indices.add(i)
+                no_wt_altr_idxs.add(i)
 
             if not (new_b.ranking or new_b.scores):
-                no_ranking_and_no_scores_altr_ballot_indices.add(i)
+                no_rank_no_score_altr_idxs.add(i)
 
         new_ballots_and_idxs[integer_idx] = (new_b, i)
     if remove_empty_ballots:
@@ -82,10 +82,10 @@ def clean_profile(
         max_ranking_length=(
             profile.max_ranking_length if retain_original_max_ranking_length else 0
         ),
-        no_weight_altr_ballot_indices=no_weight_altr_ballot_indices,
-        no_ranking_and_no_scores_altr_ballot_indices=no_ranking_and_no_scores_altr_ballot_indices,
-        nonempty_but_altr_ballot_indices=nonempty_but_altr_ballot_indices,
-        unaltr_ballot_indices=unaltr_ballot_indices,
+        no_wt_altr_idxs=no_wt_altr_idxs,
+        no_rank_no_score_altr_idxs=no_rank_no_score_altr_idxs,
+        nonempty_altr_idxs=nonempty_altr_idxs,
+        unaltr_idxs=unaltr_idxs,
         parent_profile=profile,
         df_index_column=list(new_idxs),
     )
@@ -147,6 +147,8 @@ def remove_repeated_candidates(
     Given a profile, if a candidate appears multiple times on a ballot, keep the first instance and
     remove any further instances. Does not condense any empty rankings as as result.
     Only works on ranking ballots, not score ballots.
+
+    Wrapper for clean_profile.
 
     Args:
         profile (PreferenceProfile): Profile to remove repeated candidates from.
@@ -232,6 +234,9 @@ def remove_cand(
     Given a profile, remove the given candidate(s) from the ballots. Does not condense the
     resulting ballots. Removes candidates from score dictionary as well.
 
+    Wrapper for clean_profile that does some extra processing to ensure the candidate list
+    is handled correctly.
+
     Args:
         removed (Union[str, list]): Candidate or list of candidates to be removed.
         profile (PreferenceProfile): Profile to remove repeated candidates from.
@@ -273,10 +278,10 @@ def remove_cand(
         max_ranking_length=cleaned_profile.max_ranking_length,
         parent_profile=cleaned_profile.parent_profile,
         df_index_column=cleaned_profile.df_index_column,
-        no_weight_altr_ballot_indices=cleaned_profile.no_weight_altr_ballot_indices,
[…]
