### 2025-05-06 ec9e3b1
add test to remove and condense
