### 2025-05-06 70600c3
remove group ballots from transfers
