### 2025-05-06 717923a
remove group ballots from utils
