### 2025-05-08 #7477 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
