### 2025-05-12 ffcacaa
add ballot cache property, from df init
