+++ src/votekit/pref_profile/pref_profile.py
@@ -2,14 +2,13 @@ from __future__ import annotations
 import csv
 from fractions import Fraction
 import pandas as pd
-from pydantic import ConfigDict, field_validator, model_validator
+from pydantic import ConfigDict, field_validator
 from ..ballot import Ballot
-from .utils import _df_to_ballot_tuple
+from .utils import _df_to_ballot_tuple, convert_row_to_ballot
 from pydantic.dataclasses import dataclass
-from typing_extensions import Self
-from dataclasses import field
+from dataclasses import field, InitVar
 import numpy as np
-from typing import Optional, Tuple
+from typing import Optional, Tuple, TypeVar, Callable, Any
 import warnings
 import pickle
 from .profile_error import ProfileError
@@ -598,7 +597,10 @@ def _validate_csv_format(csv_data: list[list[str]]):
     _validate_csv_ballot_rows(csv_data)
 
 
-@dataclass(frozen=True, config=ConfigDict(arbitrary_types_allowed=True))
+@dataclass(
+    frozen=True,
+    config=ConfigDict(arbitrary_types_allowed=True),
+)
 class PreferenceProfile:
     """
     PreferenceProfile class, contains ballots and candidates for a given election.
@@ -637,6 +639,7 @@ class PreferenceProfile:
             scores.
 
     Raises:
+        ProfileError: a data frame and ballot list are passed to the init method.
         ProfileError: contains_rankings is set to False but a ballot contains a ranking.
         ProfileError: contains_rankings is set to True but no ballot contains a ranking.
         ProfileError: contains_scores is set to False but a ballot contains a score.
@@ -653,7 +656,7 @@ class PreferenceProfile:
 
     """
 
-    ballots: tuple[Ballot, ...] = field(default_factory=tuple)
+    ballots: InitVar[tuple[Ballot, ...]] = tuple()
     candidates: tuple[str, ...] = field(default_factory=tuple)
     max_ranking_length: int = 0
     df: pd.DataFrame = field(default_factory=pd.DataFrame)
@@ -662,6 +665,29 @@ class PreferenceProfile:
     total_ballot_wt: Fraction = Fraction(0)
     contains_rankings: Optional[bool] = None
     contains_scores: Optional[bool] = None
+
+    def __post_init__(self, ballots: tuple[Ballot, ...]):
+        """
+        Use the post init to handle the creation of a ballot df.
+
+        Args:
+            ballots (tuple[Ballot,...]): Tuple of ballots.
+        """
+        self._cands_not_ranking_columns()
+        if not self.df.equals(pd.DataFrame()) and ballots != tuple():
+            raise ProfileError(
+                "Cannot pass a dataframe and a ballot list to profile init method. Must pick one."
+            )
+
+        elif self.df.equals(pd.DataFrame()) and ballots != tuple():
+            self._ballot_list_to_df(ballots)
+
+        elif not self.df.equals(pd.DataFrame()) and ballots == tuple():
+            self._set_class_attrs_from_init_df()
+
+        self._find_max_ranking_length()
+        self._find_total_ballot_wt()
+        self._find_num_ballots()
 
     @field_validator("candidates")
     @classmethod
@@ -685,17 +711,16 @@ class PreferenceProfile:
 
         return tuple([c.strip() for c in candidates])
 
-    @model_validator(mode="after")
-    def cands_not_ranking_columns(
+    def _cands_not_ranking_columns(
         self,
-    ) -> Self:
+    ):
         """
         Ensures that candidate names do not match ranking columns.
         Also added protection for from_csv method.
 
         """
         if self.candidates == tuple():
-            return self
+            return
 
         for cand in self.candidates:
[…]
