### 2025-05-13 cfc7889
add tests for from_df init
